@@ -1,6 +1,6 @@
 {
   "name": "udonarium",
-  "version": "1.8.5",
+  "version": "1.9.3",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -12,54 +12,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.4",
-    "@angular/common": "^7.2.4",
-    "@angular/compiler": "^7.2.4",
-    "@angular/core": "^7.2.4",
-    "@angular/forms": "^7.2.4",
-    "@angular/http": "^7.2.4",
-    "@angular/platform-browser": "^7.2.4",
-    "@angular/platform-browser-dynamic": "^7.2.4",
-    "@angular/router": "^7.2.4",
-    "core-js": "^2.5.4",
+    "@angular/animations": "^8.1.1",
+    "@angular/common": "^8.1.1",
+    "@angular/compiler": "^8.1.1",
+    "@angular/core": "^8.1.1",
+    "@angular/forms": "^8.1.1",
+    "@angular/platform-browser": "^8.1.1",
+    "@angular/platform-browser-dynamic": "^8.1.1",
+    "@angular/router": "^8.1.1",
+    "core-js": "^2.6.9",
     "crypto-js": "^3.1.9-1",
-    "file-saver": "^2.0.1",
-    "js-yaml": "^3.13.0",
+    "file-saver": "^2.0.2",
+    "js-yaml": "^3.13.1",
     "jszip": "~3.1.5",
     "lzbase62": "^1.4.9",
     "msgpack-lite": "^0.1.26",
-    "rxjs": "^6.4.0",
-    "tslib": "^1.9.0",
+    "pako": "^1.0.10",
+    "rxjs": "^6.5.2",
+    "tslib": "^1.10.0",
     "vkbeautify": "^0.99.1",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.1",
-    "@angular/cli": "^7.3.1",
-    "@angular/compiler-cli": "^7.2.11",
-    "@angular/language-service": "^7.2.4",
+    "@angular-devkit/build-angular": "~0.801.1",
+    "@angular/cli": "^8.1.1",
+    "@angular/compiler-cli": "^8.1.1",
+    "@angular/language-service": "^8.1.1",
     "@types/crypto-js": "^3.1.33",
     "@types/file-saver": "^1.3.0",
-    "@types/jasmine": "~2.8.8",
+    "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/jquery": "^2.0.41",
     "@types/jqueryui": "^1.11.32",
-    "@types/js-yaml": "^3.9.0",
-    "@types/jszip": "^0.0.32",
+    "@types/js-yaml": "^3.12.1",
+    "@types/jszip": "^3.1.6",
     "@types/msgpack-lite": "^0.1.5",
     "@types/node": "~8.9.4",
+    "@types/pako": "^1.0.1",
     "@types/vkbeautify": "^0.99.1",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.4"
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.5"
   }
 }
