@@ -1,4 +1,4 @@
-import { animate, keyframes, state, style, transition, trigger } from '@angular/animations';
+import { animate, keyframes, style, transition, trigger } from '@angular/animations';
 import {
   AfterViewInit,
   Component,
@@ -23,7 +23,6 @@ import { PointerDeviceService } from 'service/pointer-device.service';
   ],
   animations: [
     trigger('flyInOut', [
-      state('in', style({ transform: 'scale(1, 1)' })),
       transition('void => *', [
         animate('100ms ease-out', keyframes([
           style({ transform: 'scale(0.8, 0.8)', opacity: '0', offset: 0 }),
@@ -33,22 +32,13 @@ import { PointerDeviceService } from 'service/pointer-device.service';
       transition('* => void', [
         animate(100, style({ transform: 'scale(0, 0)' }))
       ])
-    ]),
-    trigger('bgInOut', [
-      state('in', style({ 'background-color': 'rgba(30, 30, 30, 0.3)' })),
-      transition('void => *', [
-        style({ 'background-color': 'rgba(30, 30, 30, 0.0)' }), animate(200)
-      ]),
-      transition('* => void', [
-        animate(200, style({ 'background-color': 'rgba(30, 30, 30, 0.0)' }))
-      ])
-    ]),
+    ])
   ]
 })
 export class UIPanelComponent implements OnInit, OnDestroy, AfterViewInit {
-  @ViewChild('draggablePanel') draggablePanel: ElementRef;
-  @ViewChild('scrollablePanel') scrollablePanel: ElementRef;
-  @ViewChild('content', { read: ViewContainerRef }) content: ViewContainerRef;
+  @ViewChild('draggablePanel', { static: true }) draggablePanel: ElementRef;
+  @ViewChild('scrollablePanel', { static: true }) scrollablePanel: ElementRef;
+  @ViewChild('content', { read: ViewContainerRef, static: true }) content: ViewContainerRef;
 
   @Input() set title(title: string) { this.panelService.title = title; }
   @Input() set left(left: number) { this.panelService.left = left; }
