@@ -2,10 +2,10 @@ import { Card } from './card';
 import { ImageFile } from './core/file-storage/image-file';
 import { SyncObject, SyncVar } from './core/synchronize-object/decorator';
 import { ObjectNode } from './core/synchronize-object/object-node';
-import { ObjectStore } from './core/synchronize-object/object-store';
 import { DataElement } from './data-element';
 import { PeerCursor } from './peer-cursor';
 import { TabletopObject } from './tabletop-object';
+import { moveToTopmost } from './tabletop-object-util';
 
 @SyncObject('card-stack')
 export class CardStack extends TabletopObject {
@@ -27,7 +27,7 @@ export class CardStack extends TabletopObject {
     }
     return null;
   }
-  get cards(): Card[] { return this.cardRoot ? <Card[]>this.cardRoot.children.concat() : []; }
+  get cards(): Card[] { return this.cardRoot ? <Card[]>this.cardRoot.children : []; }
   get topCard(): Card { return this.isEmpty ? null : this.cards[0]; }
   get isEmpty(): boolean { return this.cards.length < 1 }
   get imageFile(): ImageFile { return this.topCard ? this.topCard.imageFile : null; }
@@ -38,6 +38,7 @@ export class CardStack extends TabletopObject {
     for (let card of this.cards) {
       card.index = Math.random() * length;
       card.rotate = Math.floor(Math.random() * 2) * 180;
+      this.setSamePositionFor(card);
     }
     return this.cards;
   }
@@ -45,14 +46,10 @@ export class CardStack extends TabletopObject {
   drawCard(): Card {
     let card = this.topCard ? <Card>this.cardRoot.removeChild(this.topCard) : null;
     if (card) {
-      card.location.name = this.location.name;
-      card.location.x = this.location.x;
-      card.location.y = this.location.y;
-      card.posZ = this.posZ;
       card.rotate += this.rotate;
       if (360 < card.rotate) card.rotate -= 360;
+      this.setSamePositionFor(card);
       card.toTopmost();
-      card.update();
     }
     return card;
   }
@@ -61,39 +58,45 @@ export class CardStack extends TabletopObject {
     let cards = this.cards;
     for (let card of cards) {
       this.cardRoot.removeChild(card);
-      card.location.name = this.location.name;
-      card.location.x = this.location.x;
-      card.location.y = this.location.y;
-      card.posZ = this.posZ;
       card.rotate += this.rotate;
+      this.setSamePositionFor(card);
       if (360 < card.rotate) card.rotate -= 360;
     }
     return cards;
   }
 
   faceUp() {
-    if (this.topCard) this.topCard.faceUp();
+    if (this.topCard) {
+      this.topCard.faceUp();
+      this.setSamePositionFor(this.topCard);
+    }
   }
 
   faceDown() {
-    if (this.topCard) this.topCard.faceDown();
+    if (this.topCard) {
+      this.topCard.faceDown();
+      this.setSamePositionFor(this.topCard);
+    }
   }
 
   faceUpAll() {
     for (let card of this.cards) {
       card.faceUp();
+      this.setSamePositionFor(card);
     }
   }
 
   faceDownAll() {
     for (let card of this.cards) {
       card.faceDown();
+      this.setSamePositionFor(card);
     }
   }
 
   uprightAll() {
     for (let card of this.cards) {
       card.rotate = 0;
+      this.setSamePositionFor(card);
     }
   }
 
@@ -104,48 +107,44 @@ export class CardStack extends TabletopObject {
   }
 
   putOnTop(card: Card): Card {
-    if (!this.cardRoot) return;
+    if (!this.cardRoot) return null;
     if (!this.topCard) return this.putOnBottom(card);
     card.owner = '';
     card.zindex = 0;
     let delta = Math.abs(card.rotate - this.rotate);
     if (180 < delta) delta = 360 - delta;
     card.rotate = delta <= 90 ? 0 : 180;
-    card.location.name = this.identifier;
-    card.update();
+    this.setSamePositionFor(card);
     return <Card>this.cardRoot.insertBefore(card, this.topCard);
   }
 
   putOnBottom(card: Card): Card {
-    if (!this.cardRoot) return;
+    if (!this.cardRoot) return null;
     card.owner = '';
     card.zindex = 0;
     let delta = Math.abs(card.rotate - this.rotate);
     if (180 < delta) delta = 360 - delta;
     card.rotate = delta <= 90 ? 0 : 180;
-    card.location.name = this.identifier;
-    card.update();
+    this.setSamePositionFor(card);
     return <Card>this.cardRoot.appendChild(card);
   }
 
   toTopmost() {
-    let object: any[] = ObjectStore.instance.getObjects('card-stack');
-    object = object.concat(ObjectStore.instance.getObjects('card'));
-    object.sort((a, b) => {
-      if (a.zindex < b.zindex) return -1;
-      if (a.zindex > b.zindex) return 1;
-      return 0;
-    });
-    let last = object[object.length - 1];
-    if (last === this) return;
-    let max = last.zindex;
-    if (this.zindex <= max) this.zindex = max + 1;
+    moveToTopmost(this, ['card']);
+  }
 
-    if (object.length * 16 < max) {
-      for (let i = 0; i < object.length; i++) {
-        object[i].zindex = i;
-      }
-    }
+  // override
+  setLocation(location: string) {
+    super.setLocation(location);
+    let cards = this.cards;
+    for (let card of cards) card.setLocation(location);
+  }
+
+  private setSamePositionFor(card: Card) {
+    card.location.name = this.location.name;
+    card.location.x = this.location.x;
+    card.location.y = this.location.y;
+    card.posZ = this.posZ;
   }
 
   static create(name: string, identifier?: string): CardStack {
