@@ -1,10 +1,10 @@
 import { ImageFile } from './core/file-storage/image-file';
 import { SyncObject, SyncVar } from './core/synchronize-object/decorator';
-import { ObjectStore } from './core/synchronize-object/object-store';
 import { Network } from './core/system';
 import { DataElement } from './data-element';
 import { PeerCursor } from './peer-cursor';
 import { TabletopObject } from './tabletop-object';
+import { moveToTopmost } from './tabletop-object-util';
 
 export enum CardState {
   FRONT,
@@ -47,23 +47,7 @@ export class Card extends TabletopObject {
   }
 
   toTopmost() {
-    let object: any[] = ObjectStore.instance.getObjects('card-stack');
-    object = object.concat(ObjectStore.instance.getObjects('card'));
-    object.sort((a, b) => {
-      if (a.zindex < b.zindex) return -1;
-      if (a.zindex > b.zindex) return 1;
-      return 0;
-    });
-    let last = object[object.length - 1];
-    if (last === this) return;
-    let max = last.zindex;
-    if (this.zindex <= max) this.zindex = max + 1;
-
-    if (object.length * 16 < max) {
-      for (let i = 0; i < object.length; i++) {
-        object[i].zindex = i;
-      }
-    }
+    moveToTopmost(this, ['card-stack']);
   }
 
   static create(name: string, fornt: string, back: string, size: number = 2, identifier?: string): Card {
