@@ -21,11 +21,12 @@ interface ReceivedChank {
   id: string;
   chanks: Uint8Array[];
   length: number;
+  byteLength: number;
 };
 
 export class SkyWayDataConnection extends EventEmitter {
   private chunkSize = 15.5 * 1024;
-  private receivedChankMap: Map<string, ReceivedChank> = new Map();
+  private receivedMap: Map<string, ReceivedChank> = new Map();
 
   get open(): boolean { return this.conn.open; }
   get remoteId(): string { return this.conn.remoteId; }
@@ -57,9 +58,11 @@ export class SkyWayDataConnection extends EventEmitter {
 
     let id = UUID.generateUuid();
 
+    let sliceData: Uint8Array = null;
+    let chank: DataChank = null;
     for (let sliceIndex = 0; sliceIndex < total; sliceIndex++) {
-      let sliceData = encodedData.slice(sliceIndex * this.chunkSize, (sliceIndex + 1) * this.chunkSize);
-      let chank: DataChank = { id: id, data: sliceData, index: sliceIndex, total: total };
+      sliceData = encodedData.slice(sliceIndex * this.chunkSize, (sliceIndex + 1) * this.chunkSize);
+      chank = { id: id, data: sliceData, index: sliceIndex, total: total };
       this.conn.send(MessagePack.encode(chank));
     }
   }
@@ -72,26 +75,25 @@ export class SkyWayDataConnection extends EventEmitter {
       return;
     }
 
-    let receivedChank = this.receivedChankMap.get(chank.id);
-    if (receivedChank == null) {
-      receivedChank = { id: chank.id, chanks: new Array(chank.total), length: 0 };
-      this.receivedChankMap.set(chank.id, receivedChank);
+    let received = this.receivedMap.get(chank.id);
+    if (received == null) {
+      received = { id: chank.id, chanks: new Array(chank.total), length: 0, byteLength: 0 };
+      this.receivedMap.set(chank.id, received);
     }
 
-    if (receivedChank.chanks[chank.index] != null) return;
+    if (received.chanks[chank.index] != null) return;
 
-    receivedChank.length++;
-    receivedChank.chanks[chank.index] = chank.data;
+    received.length++;
+    received.byteLength += chank.data.byteLength;
+    received.chanks[chank.index] = chank.data;
 
-    if (receivedChank.length < chank.total) return;
-    this.receivedChankMap.delete(chank.id);
+    if (received.length < chank.total) return;
+    this.receivedMap.delete(chank.id);
 
-    let sumLength = 0;
-    for (let c of receivedChank.chanks) sumLength += c.byteLength;
-    let uint8Array = new Uint8Array(sumLength);
+    let uint8Array = new Uint8Array(received.byteLength);
 
     let pos = 0;
-    for (let c of receivedChank.chanks) {
+    for (let c of received.chanks) {
       uint8Array.set(c, pos);
       pos += c.byteLength;
     }
