@@ -48,7 +48,7 @@
     Opal.defn(self, '$getHelpMessage', TMP_DarkDaysDrive_getHelpMessage_4 = function $$getHelpMessage() {
       var self = this;
 
-      return "" + "・判定\n" + "スペシャル／ファンブル／成功／失敗を判定\n" + "・各種表\n" + "RTT\tランダム特技決定表\n" + "ABRT アビリティ決定表\n" + "DT ダメージ表\n" + "FT 失敗表\n" + "GJT 大成功表\n" + "ITT 移動トラブル表\n" + "NTT 任務トラブル表\n" + "STT 襲撃トラブル表\n" + "HTT 変身トラブル表\n" + "DET ドライブイベント表\n" + "BET ブレイクイベント表\n" + "CT キャンプ表\n" + "KZT 関係属性表\n" + "・D66ダイスあり\n"
+      return "" + "・判定\n" + "スペシャル／ファンブル／成功／失敗を判定\n" + "・各種表\n" + "RTT\tランダム特技決定表\n" + "ABRT アビリティ決定表\n" + "DT ダメージ表\n" + "FT 失敗表\n" + "GJT 大成功表\n" + "ITT 移動トラブル表\n" + "NTT 任務トラブル表\n" + "STT 襲撃トラブル表\n" + "HTT 変身トラブル表\n" + "DET ドライブイベント表\n" + "BET ブレイクイベント表\n" + "CT キャンプ表\n" + "KZT 関係属性表\n" + "IA イケメンアクション決定表\n" + " IAA 遠距離\n" + " IAB 移動\n" + " IAC 近距離\n" + " IAD 善人\n" + " IAE 悪人\n" + " IAF 幼い\n" + " IAG バカ\n" + " IAH 渋い\n" + " IAI 賢い\n" + " IAJ 超自然\n" + "・D66ダイスあり\n"
     }, TMP_DarkDaysDrive_getHelpMessage_4.$$arity = 0);
     
     Opal.defn(self, '$check_2D6', TMP_DarkDaysDrive_check_2D6_5 = function $$check_2D6(total_n, dice_n, signOfInequality, diff, dice_cnt, dice_max, n1, n_max) {
@@ -133,7 +133,7 @@ if (item == null) item = nil;
           return item
         }}, TMP_9.$$s = self, TMP_9.$$arity = 1, TMP_9))
     }, TMP_DarkDaysDrive_getD66Table_10.$$arity = 1);
-    (Opal.class_variable_set($DarkDaysDrive, '@@tables', $hash2(["ABRT", "DT", "FT", "GJT", "ITT", "NTT", "STT", "HTT", "DET", "BET", "CT", "KZT"], {"ABRT": $hash2(["name", "type", "table"], {"name": "アビリティ決定表", "type": "D66S", "table": ["11:インストラクター(P155)", "12:運送業(P155)", "13:運転手(P155)", "14:カフェ店員(P155)", "15:趣味人(P155)", "16:ショップ店員(P155)", "22:正社員(P156)", "23:大工(P156)", "24:探偵(P156)", "25:バイヤー(P156)", "26:俳優(P156)", "33:派遣社員(P156)", "34:犯罪者(P157)", "35:バンドマン(P157)", "36:バーテンダー(P157)", "44:ヒモ(P157)", "45:ホスト(P157)", "46:ホテルマン(P157)", "55:無職(P158)", "56:用心棒(P158)", "66:料理人(P158)"]}), "DT": $hash2(["name", "type", "table"], {"name": "ダメージ表", "type": "1D6", "table": ["疲れ", "痛み", "焦り", "不調", "ショック", "ケガ"]}), "FT": $hash2(["name", "type", "table"], {"name": "失敗表", "type": "1D6", "table": ["任意のアイテムを一つ失う", "１ダメージを受ける", "【所持金ランク】が１減少する（最低０）", "２ダメージを受ける", "【所持金ランク】が２減少する（最低０）", "標的レベルが１増加する"]}), "GJT": $hash2(["name", "type", "table"], {"name": "大成功表", "type": "1D6", "table": ["主人からお褒めの言葉をいただく", "ダメージを１回復する", "ダメージを１回復する", "関係のチェックを一つ消す", "ダメージを２回復する", "【所持金ランク】が１増加する"]}), "ITT": $hash2(["name", "type", "table"], {"name": "移動トラブル表", "type": "1D6", "table": ["検問（P220)", "急な腹痛（P220)", "黒煙（P221)", "蚊（P221)", "落とし物（P222)", "空腹（P222)"]}), "NTT": $hash2(["name", "type", "table"], {"name": "任務トラブル表", "type": "1D6", "table": ["通報（P223)", "プレッシャー（P223)", "マナー違反（P224)", "志願者（P224)", "仲間割れ（P225)", "狩人の噂（P225)"]}), "STT": $hash2(["name", "type", "table"], {"name": "襲撃トラブル表", "type": "1D6", "table": ["孤独な追跡者（P226)", "地元の若者たち（P226)", "V-FILES（P227)", "チンピラの群れ（P227)", "孤独な狩人（P228)", "狩人の群れ（P228)"]}), "HTT": $hash2(["name", "type", "table"], {"name": "変身トラブル表", "type": "D66N", "table": ["11:あれを食べたい(P214)", "12:あれを着たい(P214)", "13:あれを見たい(P215)", "14:あれを狩りたい(P215)", "15:あれを踊りたい(P216)", "16:あれに入りたい(P216)", "21:強奪(P217)", "22:暴行(P217)", "23:虐殺(P218)", "24:誘拐(P218)", "25:無精(P219)", "26:失踪(P219)", "31:あれを食べたい(P214)", "32:あれを着たい(P214)", "33:あれを見たい(P215)", "34:あれを狩りたい(P215)", "35:あれを踊りたい(P216)", "36:あれに入りたい(P216)", "41:強奪(P217)", "42:暴行(P217)", "43:虐殺(P218)", "44:誘拐(P218)", "45:無精(P219)", "46:失踪(P219)", "51:あれを食べたい(P214)", "52:あれを着たい(P214)", "53:あれを見たい(P215)", "54:あれを狩りたい(P215)", "55:あれを踊りたい(P216)", "56:あれに入りたい(P216)", "61:強奪(P217)", "62:暴行(P217)", "63:虐殺(P218)", "64:誘拐(P218)", "65:無精(P219)", "66:失踪(P219)"]}), "DET": $hash2(["name", "type", "table"], {"name": "ドライブイベント表", "type": "1D6", "table": ["身の上話をする。目標が背景分野で選択している特技がドライブ判定の指定特技になる。", "スキル自慢をする。目標が仕事分野で選択している特技がドライブ判定の指定特技になる。", "むかし行った場所の話をする。目標が捜索分野で選択している特技がドライブ判定の指定特技になる。", "趣味の話をする。目標が趣味分野で選択している特技がドライブ判定の指定特技になる。", "テーマがない雑談をする。目標が雰囲気分野で選択している特技がドライブ判定の指定特技になる。", "物騒な話をする。目標が戦闘法分野で選択している特技がドライブ判定の指定特技になる。"]}), "BET": $hash2(["name", "type", "table"], {"name": "ブレイクイベント表", "type": "1D6", "table": ["イケメンの車は風光明美な場所に到着する。197ページの「観光地」を参照。", "イケメンの車は明るい光に照らされた小さな店に到着する。197ページの「コンビニ」を参照。", "イケメンの車は巨大かつ何でも売っている店に到着する。198ページの「ホームセンター」を参照。", "イケメンの車はドライバーたちの憩いの地に到着する。198ページの「サービスエリア」を参照。", "イケメンの車は大きなサービスエリアのような場所に到着する。199ページの「道の駅」を参照。", "イケメンの車は闇の底に隠された秘密の場所に到着する。199ページの「友愛会支部」を参照。"]}), "CT": $hash2(["name", "type", "table"], {"name": "キャンプ表", "type": "1D6", "table": ["無料仮眠所・いい感じの空き地：定員無制限／居住性-2／価格0／発見率2", "カプセルホテル：定員1／居住性-1／価格3／発見率2", "ラブホテル：定員2／居住性0／価格4／発見率1", "ビジネスホテル：定員2／居住性0／価格4／発見率1", "観光ホテル：定員4／居住性1／価格5／発見率1", "高級ホテル：定員4／居住性2／価格6／発見率0"]}), "KZT": $hash2(["name", "type", "table"], {"name": "関係属性表", "type": "1D6", "table": ["軽蔑", "反感", "混乱", "興味", "共感", "憧れ"]})})));
+    (Opal.class_variable_set($DarkDaysDrive, '@@tables', $hash2(["ABRT", "DT", "FT", "GJT", "ITT", "NTT", "STT", "HTT", "DET", "BET", "CT", "KZT", "IA", "IAA", "IAB", "IAC", "IAD", "IAE", "IAF", "IAG", "IAH", "IAI", "IAJ"], {"ABRT": $hash2(["name", "type", "table"], {"name": "アビリティ決定表", "type": "D66S", "table": ["11:インストラクター(P155)", "12:運送業(P155)", "13:運転手(P155)", "14:カフェ店員(P155)", "15:趣味人(P155)", "16:ショップ店員(P155)", "22:正社員(P156)", "23:大工(P156)", "24:探偵(P156)", "25:バイヤー(P156)", "26:俳優(P156)", "33:派遣社員(P156)", "34:犯罪者(P157)", "35:バンドマン(P157)", "36:バーテンダー(P157)", "44:ヒモ(P157)", "45:ホスト(P157)", "46:ホテルマン(P157)", "55:無職(P158)", "56:用心棒(P158)", "66:料理人(P158)"]}), "DT": $hash2(["name", "type", "table"], {"name": "ダメージ表", "type": "1D6", "table": ["疲れ", "痛み", "焦り", "不調", "ショック", "ケガ"]}), "FT": $hash2(["name", "type", "table"], {"name": "失敗表", "type": "1D6", "table": ["任意のアイテムを一つ失う", "１ダメージを受ける", "【所持金ランク】が１減少する（最低０）", "２ダメージを受ける", "【所持金ランク】が２減少する（最低０）", "標的レベルが１増加する"]}), "GJT": $hash2(["name", "type", "table"], {"name": "大成功表", "type": "1D6", "table": ["主人からお褒めの言葉をいただく", "ダメージを１回復する", "ダメージを１回復する", "関係のチェックを一つ消す", "ダメージを２回復する", "【所持金ランク】が１増加する"]}), "ITT": $hash2(["name", "type", "table"], {"name": "移動トラブル表", "type": "1D6", "table": ["検問（P220)", "急な腹痛（P220)", "黒煙（P221)", "蚊（P221)", "落とし物（P222)", "空腹（P222)"]}), "NTT": $hash2(["name", "type", "table"], {"name": "任務トラブル表", "type": "1D6", "table": ["通報（P223)", "プレッシャー（P223)", "マナー違反（P224)", "志願者（P224)", "仲間割れ（P225)", "狩人の噂（P225)"]}), "STT": $hash2(["name", "type", "table"], {"name": "襲撃トラブル表", "type": "1D6", "table": ["孤独な追跡者（P226)", "地元の若者たち（P226)", "V-FILES（P227)", "チンピラの群れ（P227)", "孤独な狩人（P228)", "狩人の群れ（P228)"]}), "HTT": $hash2(["name", "type", "table"], {"name": "変身トラブル表", "type": "D66N", "table": ["11:あれを食べたい(P214)", "12:あれを着たい(P214)", "13:あれを見たい(P215)", "14:あれを狩りたい(P215)", "15:あれを踊りたい(P216)", "16:あれに入りたい(P216)", "21:強奪(P217)", "22:暴行(P217)", "23:虐殺(P218)", "24:誘拐(P218)", "25:無精(P219)", "26:失踪(P219)", "31:あれを食べたい(P214)", "32:あれを着たい(P214)", "33:あれを見たい(P215)", "34:あれを狩りたい(P215)", "35:あれを踊りたい(P216)", "36:あれに入りたい(P216)", "41:強奪(P217)", "42:暴行(P217)", "43:虐殺(P218)", "44:誘拐(P218)", "45:無精(P219)", "46:失踪(P219)", "51:あれを食べたい(P214)", "52:あれを着たい(P214)", "53:あれを見たい(P215)", "54:あれを狩りたい(P215)", "55:あれを踊りたい(P216)", "56:あれに入りたい(P216)", "61:強奪(P217)", "62:暴行(P217)", "63:虐殺(P218)", "64:誘拐(P218)", "65:無精(P219)", "66:失踪(P219)"]}), "DET": $hash2(["name", "type", "table"], {"name": "ドライブイベント表", "type": "1D6", "table": ["身の上話をする。目標が背景分野で選択している特技がドライブ判定の指定特技になる。", "スキル自慢をする。目標が仕事分野で選択している特技がドライブ判定の指定特技になる。", "むかし行った場所の話をする。目標が捜索分野で選択している特技がドライブ判定の指定特技になる。", "趣味の話をする。目標が趣味分野で選択している特技がドライブ判定の指定特技になる。", "テーマがない雑談をする。目標が雰囲気分野で選択している特技がドライブ判定の指定特技になる。", "物騒な話をする。目標が戦闘法分野で選択している特技がドライブ判定の指定特技になる。"]}), "BET": $hash2(["name", "type", "table"], {"name": "ブレイクイベント表", "type": "1D6", "table": ["イケメンの車は風光明美な場所に到着する。197ページの「観光地」を参照。", "イケメンの車は明るい光に照らされた小さな店に到着する。197ページの「コンビニ」を参照。", "イケメンの車は巨大かつ何でも売っている店に到着する。198ページの「ホームセンター」を参照。", "イケメンの車はドライバーたちの憩いの地に到着する。198ページの「サービスエリア」を参照。", "イケメンの車は大きなサービスエリアのような場所に到着する。199ページの「道の駅」を参照。", "イケメンの車は闇の底に隠された秘密の場所に到着する。199ページの「友愛会支部」を参照。"]}), "CT": $hash2(["name", "type", "table"], {"name": "キャンプ表", "type": "1D6", "table": ["無料仮眠所・いい感じの空き地：定員無制限／居住性-2／価格0／発見率2", "カプセルホテル：定員1／居住性-1／価格3／発見率2", "ラブホテル：定員2／居住性0／価格4／発見率1", "ビジネスホテル：定員2／居住性0／価格4／発見率1", "観光ホテル：定員4／居住性1／価格5／発見率1", "高級ホテル：定員4／居住性2／価格6／発見率0"]}), "KZT": $hash2(["name", "type", "table"], {"name": "関係属性表", "type": "1D6", "table": ["軽蔑", "反感", "混乱", "興味", "共感", "憧れ"]}), "IA": $hash2(["name", "type", "table"], {"name": "イケメンアクション決定表", "type": "D66S", "table": ["11:遠距離", "12:遠距離", "13:移動", "14:移動", "15:近距離", "16:近距離", "22:善人", "23:善人", "24:悪人", "25:悪人", "26:幼い", "33:幼い", "34:バカ", "35:バカ", "36:渋い", "44:渋い", "45:賢い", "46:賢い", "55:超自然", "56:超自然", "66:振り直しor自由選択"]}), "IAA": $hash2(["name", "type", "table"], {"name": "イケメンアクション（遠距離）表", "type": "1D6", "table": ["目を合わせて微笑む（かっこよさ：4）", "場所を譲る（かっこよさ：4）", "髪をかきあげる（かっこよさ：5）", "複雑なポーズで座る（かっこよさ：5）", "物憂げな表情で振り返る（かっこよさ：6）", "ものを上に持つ（かっこよさ：6）"]}), "IAB": $hash2(["name", "type", "table"], {"name": "イケメンアクション（移動）表", "type": "1D6", "table": ["車道側を歩く（かっこよさ：4）", "乗り物から降りる（かっこよさ：4）", "真剣な表情で近づく（かっこよさ：4）", "馬に乗る（かっこよさ：6）", "ダメージを受けつつ移動（かっこよさ：6）", "瞬間移動（かっこよさ：6）"]}), "IAC": $hash2(["name", "type", "table"], {"name": "イケメンアクション（近距離）表", "type": "1D6", "table": ["黙って見つめる（かっこよさ：3）", "ちょっとしたプレゼント（かっこよさ：3）", "顎クイ（かっこよさ：5）", "壁ドン（かっこよさ：5）", "お姫様抱っこ（かっこよさ：7）", "床ドン（かっこよさ：7）"]}), "IAD": $hash2(["name", "type", "table"], {"name": "イケメンアクション（善人）表", "type": "1D6", "table": ["手を引いて逃げる（かっこよさ：4）", "毛布を掛ける（かっこよさ：4）", "守る（かっこよさ：5）", "笑って去る（かっこよさ：5）", "全てを捧げる（かっこよさ：6）", "悪堕ち（かっこよさ：6）"]}), "IAE": $hash2(["name", "type", "table"], {"name": "イケメンアクション（悪人）表", "type": "1D6", "table": ["攻撃する（かっこよさ：4）", "暗く笑う（かっこよさ：4）", "奪う（かっこよさ：4）", "目論見を口にする（かっこよさ：6）", "罠にかける（かっこよさ：6）", "改心する（かっこよさ：6）"]}), "IAF": $hash2(["name", "type", "table"], {"name": "イケメンアクション（幼い）表", "type": "1D6", "table": ["甘える（かっこよさ：3）", "疲れる（かっこよさ：3）", "無邪気な発言（かっこよさ：5）", "おねだり（かっこよさ：5）", "上目遣い（かっこよさ：7）", "抱きつく（かっこよさ：7）"]}), "IAG": $hash2(["name", "type", "table"], {"name": "イケメンアクション（バカ）表", "type": "1D6", "table": ["苦悩する（かっこよさ：4）", "屈託のない笑顔（かっこよさ：4）", "転ぶ（かっこよさ：4）", "叫ぶ（かっこよさ：6）", "間違える（かっこよさ：6）", "怖がる（かっこよさ：6）"]}), "IAH": $hash2(["name", "type", "table"], {"name": "イケメンアクション（渋い）表", "type": "1D6", "table": ["説教（かっこよさ：4）", "気づかせる（かっこよさ：4）", "見守る（かっこよさ：5）", "残心（かっこよさ：5）", "称える（かっこよさ：6）", "いい位置を取る（かっこよさ：6）"]}), "IAI": $hash2(["name", "type", "table"], {"name": "イケメンアクション（賢い）表", "type": "1D6", "table": ["難しい本を読む（かっこよさ：3）", "アドバイスをする（かっこよさ：3）", "眼鏡を持ち上げる（かっこよさ：5）", "状況を解説する（かっこよさ：5）", "計算できなくなる（かっこよさ：7）", "大丈夫だと言う（かっこよさ：7）"]}), "IAJ": $hash2(["name", "type", "table"], {"name": "イケメンアクション（超自然）表", "type": "1D6", "table": ["水に濡れる（かっこよさ：4）", "風を纏う（かっこよさ：4）", "地割れ（かっこよさ：5）", "火を放つ（かっこよさ：5）", "闇を生み出す（かっこよさ：6）", "光る（かっこよさ：6）"]})})));
     return self.$setPrefixes($rb_plus(["RTT"], (($a = $DarkDaysDrive.$$cvars['@@tables']) == null ? nil : $a).$keys()));
   })($nesting[0], Opal.const_get_relative($nesting, 'DiceBot'), $nesting)
 })(Opal);
