@@ -32,6 +32,17 @@ DET ドライブイベント表
 BET ブレイクイベント表
 CT キャンプ表
 KZT 関係属性表
+IA イケメンアクション決定表
+ IAA 遠距離
+ IAB 移動
+ IAC 近距離
+ IAD 善人
+ IAE 悪人
+ IAF 幼い
+ IAG バカ
+ IAH 渋い
+ IAI 賢い
+ IAJ 超自然
 ・D66ダイスあり
 INFO_MESSAGE_TEXT
   end
@@ -89,8 +100,7 @@ INFO_MESSAGE_TEXT
 
     return output
   end
-  
-  
+
   def getTableDiceCommandResult(command)
 
     info = @@tables[command]
@@ -325,6 +335,156 @@ V-FILES（P227)
 共感
 憧れ
 },},
+
+     'IA' => {
+      :name => "イケメンアクション決定表",
+      :type => 'D66S',
+      :table => %w{
+11:遠距離
+12:遠距離
+13:移動
+14:移動
+15:近距離
+16:近距離
+22:善人
+23:善人
+24:悪人
+25:悪人
+26:幼い
+33:幼い
+34:バカ
+35:バカ
+36:渋い
+44:渋い
+45:賢い
+46:賢い
+55:超自然
+56:超自然
+66:振り直しor自由選択
+},},
+
+    'IAA' => {
+      :name => "イケメンアクション（遠距離）表",
+      :type => '1D6',
+      :table => %w{
+目を合わせて微笑む（かっこよさ：4）
+場所を譲る（かっこよさ：4）
+髪をかきあげる（かっこよさ：5）
+複雑なポーズで座る（かっこよさ：5）
+物憂げな表情で振り返る（かっこよさ：6）
+ものを上に持つ（かっこよさ：6）
+},},
+
+    'IAB' => {
+      :name => "イケメンアクション（移動）表",
+      :type => '1D6',
+      :table => %w{
+車道側を歩く（かっこよさ：4）
+乗り物から降りる（かっこよさ：4）
+真剣な表情で近づく（かっこよさ：4）
+馬に乗る（かっこよさ：6）
+ダメージを受けつつ移動（かっこよさ：6）
+瞬間移動（かっこよさ：6）
+},},
+
+'IAC' => {
+      :name => "イケメンアクション（近距離）表",
+      :type => '1D6',
+      :table => %w{
+黙って見つめる（かっこよさ：3）
+ちょっとしたプレゼント（かっこよさ：3）
+顎クイ（かっこよさ：5）
+壁ドン（かっこよさ：5）
+お姫様抱っこ（かっこよさ：7）
+床ドン（かっこよさ：7）
+},},
+
+'IAD' => {
+      :name => "イケメンアクション（善人）表",
+      :type => '1D6',
+      :table => %w{
+手を引いて逃げる（かっこよさ：4）
+毛布を掛ける（かっこよさ：4）
+守る（かっこよさ：5）
+笑って去る（かっこよさ：5）
+全てを捧げる（かっこよさ：6）
+悪堕ち（かっこよさ：6）
+},},
+
+'IAE' => {
+      :name => "イケメンアクション（悪人）表",
+      :type => '1D6',
+      :table => %w{
+攻撃する（かっこよさ：4）
+暗く笑う（かっこよさ：4）
+奪う（かっこよさ：4）
+目論見を口にする（かっこよさ：6）
+罠にかける（かっこよさ：6）
+改心する（かっこよさ：6）
+},},
+
+'IAF' => {
+      :name => "イケメンアクション（幼い）表",
+      :type => '1D6',
+      :table => %w{
+甘える（かっこよさ：3）
+疲れる（かっこよさ：3）
+無邪気な発言（かっこよさ：5）
+おねだり（かっこよさ：5）
+上目遣い（かっこよさ：7）
+抱きつく（かっこよさ：7）
+},},
+
+'IAG' => {
+      :name => "イケメンアクション（バカ）表",
+      :type => '1D6',
+      :table => %w{
+苦悩する（かっこよさ：4）
+屈託のない笑顔（かっこよさ：4）
+転ぶ（かっこよさ：4）
+叫ぶ（かっこよさ：6）
+間違える（かっこよさ：6）
+怖がる（かっこよさ：6）
+},},
+
+'IAH' => {
+      :name => "イケメンアクション（渋い）表",
+      :type => '1D6',
+      :table => %w{
+説教（かっこよさ：4）
+気づかせる（かっこよさ：4）
+見守る（かっこよさ：5）
+残心（かっこよさ：5）
+称える（かっこよさ：6）
+いい位置を取る（かっこよさ：6）
+},},
+
+'IAI' => {
+      :name => "イケメンアクション（賢い）表",
+      :type => '1D6',
+      :table => %w{
+難しい本を読む（かっこよさ：3）
+アドバイスをする（かっこよさ：3）
+眼鏡を持ち上げる（かっこよさ：5）
+状況を解説する（かっこよさ：5）
+計算できなくなる（かっこよさ：7）
+大丈夫だと言う（かっこよさ：7）
+},},
+
+'IAJ' => {
+      :name => "イケメンアクション（超自然）表",
+      :type => '1D6',
+      :table => %w{
+水に濡れる（かっこよさ：4）
+風を纏う（かっこよさ：4）
+地割れ（かっこよさ：5）
+火を放つ（かっこよさ：5）
+闇を生み出す（かっこよさ：6）
+光る（かっこよさ：6）
+},}
+
+
+
   }
 
   setPrefixes(['RTT'] + @@tables.keys)
