@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 class LogHorizon < DiceBot
-  setPrefixes(['\d+LH.*', 'PC.*', 'EC.*', 'GC.*', 'CC.*', 'CTR.*', 'MTR.*', 'ITR.*', 'HTR.*','GTR.*', 'PTAG', 'KOYU', 'MGR.', 'HLOC', 'PCNM', 'IAT.*', 'TIAS', 'ABDC', 'MII.*', 'ESCT.*', 'CSCT.*', 'ESTL.*'])
+  setPrefixes(['\d+LH.*', 'PC.*', 'EC.*', 'GC.*', 'CC.*', 'CTR.*', 'MTR.*', 'ITR.*', 'OTR.*', 'HTR.*','GTR.*', 'PTAG', 'KOYU', 'MGR.', 'HLOC', 'PCNM', 'IAT.*', 'TIAS', 'ABDC', 'MII.*', 'ESCT.*', 'CSCT.*', 'ESTL.*'])
 
   def initialize
     super
@@ -31,15 +31,19 @@ class LogHorizon < DiceBot
 　$z：＄を付けるとダイス目を z 固定。表の特定の値参照用に。省略可能。
 　例） PCT1　ECT2+1　GCT3-1　CCT3$5
 ・財宝表(tTRSx±y$)
-　CTRS 金銭／MTRS 魔法素材／ITRS 換金アイテム／※HTRS ヒロイン／GTRS ゴブリン財宝表
+　CTRS 金銭／MTRS 魔法素材／ITRS 換金アイテム／OTRS そのほか／※HTRS ヒロイン／GTRS ゴブリン財宝表
 　x：CRを指定。省略時はダイス値 0 固定で修正値の表参照。《ゴールドフィンガー》使用時など。
 　±y：修正値。＋と－の計算に対応。省略可能。
 　$：＄を付けると財宝表のダイス目を7固定（1回分のプライズ用）。省略可能。
 　例） CTRS1　MTRS2+1　ITRS3-1　ITRS+27　CTRS3$
 ・パーソナリティタグ表(PTAG)
 ・交友表(KOYU)
-・イースタル探索表(ESTLx) xはCRを指定
-・プレフィックスドマジックアイテム効果表(MGRx) xはMGを指定。
+・イースタル探索表(ESTLx±y$z)
+　x：CRを指定。省略時はダイス値 0 固定で修正値の表参照。
+　±y：修正値。＋と－の計算に対応。省略可能。
+　$z：＄を付けるとダイス目を z 固定。特定CRの表参照用に。省略可能。
+　例） ESTL1　ESTL+15　ESTL2+1$5　ESTL2-1$5
+・プレフィックスドマジックアイテム効果表(MGRx) xはMGを指定。(LHZB1用)
 †楽器種別表(MIIx) xは楽器の種類(1～6を指定)、省略可能
 　1 打楽器１／2 鍵盤楽器／3 弦楽器１／4 弦楽器２／5 管楽器１／6 管楽器２
 ☆特殊消耗表(tSCTx±y$z)　消耗表と同様、ただしCRは省略可能。
@@ -73,26 +77,26 @@ MESSAGETEXT
     modifyText = ($2 || '')
     difficultyText = $4
 
-	#修正値の計算
-	modify = getValue( modifyText, 0 )
+    #修正値の計算
+    modify = getValue( modifyText, 0 )
 
-	#目標値の計算
+    #目標値の計算
     difficulty = getValue( difficultyText, nil )
 
-	#ダイスロール
-	dice, dice_str = roll(diceCount, 6)
+    #ダイスロール
+    dice, dice_str = roll(diceCount, 6)
     diceList = dice_str.split(/,/).collect{|i|i.to_i}.sort
 
-	total = dice + modify
+    total = dice + modify
 
-	#出力用ダイスコマンドを生成
-	command =  "#{diceCount}LH#{modifyText}"
-	command += ">=#{difficulty}" unless(difficulty.nil?)
+    #出力用ダイスコマンドを生成
+    command =  "#{diceCount}LH#{modifyText}"
+    command += ">=#{difficulty}" unless(difficulty.nil?)
 
-	#出力文の生成
-	result = "(#{command}) ＞ #{dice}[#{dice_str}]#{modifyText} ＞ #{total}"
+    #出力文の生成
+    result = "(#{command}) ＞ #{dice}[#{dice_str}]#{modifyText} ＞ #{total}"
 
-	#クリティカル・ファンブルチェック
+    #クリティカル・ファンブルチェック
     if( isCritical(diceList) )
       result += " ＞ クリティカル！"
     elsif( isFamble(diceList, diceCount) )
@@ -123,7 +127,7 @@ MESSAGETEXT
   end
 
   def isCritical(diceList)
-	(diceList.select{|i| i == 6 }.size >= 2)
+    (diceList.select{|i| i == 6 }.size >= 2)
   end
 
   def isFamble(diceList, diceCount)
@@ -202,6 +206,46 @@ MESSAGETEXT
                [6, '[疲労:35]を受ける'],
                [7, '[疲労:40]を受ける'],
               ],
+              [
+               [0, '消耗なし'],
+               [1, '[疲労:25]を受ける'],
+               [2, '[疲労:35]を受ける'],
+               [3, '[疲労:45]を受ける'],
+               [4, '[疲労:55]を受ける'],
+               [5, '[疲労:65]を受ける'],
+               [6, '[疲労:65]を受け、【因果力】を1点失う'],
+               [7, '[疲労:65]を受け、【因果力】を2点失う'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '[疲労:40]を受ける'],
+               [2, '[疲労:60]を受ける'],
+               [3, '[疲労:80]を受ける'],
+               [4, '[疲労:80]を受け、【因果力】を1点失う'],
+               [5, '[疲労:80]を受け、【因果力】を2点失う'],
+               [6, '[疲労:90]を受け、【因果力】を2点失う'],
+               [7, '[疲労:90]を受け、【因果力】を3点失う'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '[疲労:60]を受ける'],
+               [2, '[疲労:85]を受ける'],
+               [3, '[疲労:110]を受ける'],
+               [4, '[疲労:100]を受け、【因果力】を1点失う'],
+               [5, '[疲労:120]を受け、【因果力】を1点失う'],
+               [6, '[疲労:100]を受け、【因果力】を2点失う'],
+               [7, '[疲労:100]を受け、【因果力】を3点失う'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '[疲労:80]を受ける'],
+               [2, '[疲労:120]を受ける'],
+               [3, '[疲労:120]を受け、【因果力】を1点失う'],
+               [4, '[疲労:120]を受け、【因果力】を1点失う'],
+               [5, '[疲労:120]を受け、【因果力】を2点失う'],
+               [6, '[疲労:120]を受け、【因果力】を2点失う'],
+               [7, '[疲労:120]を受け、【因果力】を3点失う'],
+              ],
              ]
 
     return tableName, tables
@@ -228,6 +272,46 @@ MESSAGETEXT
                [5, '【因果力】を2点失う'],
                [6, '【因果力】を2点失う'],
                [7, '【因果力】を3点失う'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '【因果力】を1点失う'],
+               [2, '【因果力】を1点失う'],
+               [3, '【因果力】を2点失う'],
+               [4, '【因果力】を2点失う'],
+               [5, '【因果力】を2点失う'],
+               [6, '【因果力】を3点失う'],
+               [7, '【因果力】を4点失う'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '【因果力】を1点失う'],
+               [2, '【因果力】を2点失う'],
+               [3, '【因果力】を2点失う'],
+               [4, '【因果力】を2点失う'],
+               [5, '【因果力】を3点失う'],
+               [6, '【因果力】を1点失い、特技喪失する'],
+               [7, '【因果力】を2点失い、特技喪失する'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '【因果力】を1点失う'],
+               [2, '【因果力】を2点失う'],
+               [3, '【因果力】を2点失う'],
+               [4, '【因果力】を3点失う'],
+               [5, '【因果力】を1点失い、特技喪失する'],
+               [6, '【因果力】を2点失い、特技喪失する'],
+               [7, '【因果力】を3点失い、特技喪失する'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '【因果力】を1点失う'],
+               [2, '【因果力】を2点失う'],
+               [3, '【因果力】を3点失う'],
+               [4, '【因果力】を1点失い、特技喪失する'],
+               [5, '【因果力】を4点失う'],
+               [6, '【因果力】を2点失い、特技喪失する'],
+               [7, '【因果力】を3点失い、特技喪失する'],
               ],
              ]
 
@@ -256,6 +340,46 @@ MESSAGETEXT
                [6, '[消耗品]アイテムを3個失う'],
                [7, '[消耗品]アイテムを4個失う'],
               ],
+              [
+               [0, '消耗なし'],
+               [1, '[消耗品]アイテムを1個失う'],
+               [2, '[消耗品]アイテムを2個失う'],
+               [3, '[消耗品]アイテムを2個失う'],
+               [4, '[消耗品]アイテムを3個失う'],
+               [5, '[消耗品]アイテムを3個失う'],
+               [6, '[消耗品]アイテムを3個失い、[圧迫:1]を受ける'],
+               [7, '[消耗品]アイテムを3個失い、[圧迫:2]を受ける'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '[消耗品]アイテムを2個失う'],
+               [2, '[消耗品]アイテムを2個失う'],
+               [3, '[消耗品]アイテムを2個失い、[圧迫:1]を受ける'],
+               [4, '[消耗品]アイテムを3個失う'],
+               [5, '[消耗品]アイテムを3個失い、[圧迫:1]を受ける'],
+               [6, '[消耗品]アイテムを4個失う'],
+               [7, '[消耗品]アイテムを4個失い、[圧迫:1]を受ける'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '[消耗品]アイテムを2個失う'],
+               [2, '[消耗品]アイテムを2個失い、[圧迫:1]を受ける'],
+               [3, '[消耗品]アイテムを3個失う'],
+               [4, '[消耗品]アイテムを3個失い、[圧迫:1]を受ける'],
+               [5, '[消耗品]アイテムを4個失う'],
+               [6, '[消耗品]アイテムを4個失い、[圧迫:1]を受ける'],
+               [7, '[消耗品]アイテムを4個失い、[圧迫:2]を受ける'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '[消耗品]アイテムを3個失う'],
+               [2, '[消耗品]アイテムを3個失う'],
+               [3, '[消耗品]アイテムを3個失い、[圧迫:1]を受ける'],
+               [4, '[消耗品]アイテムを4個失う'],
+               [5, '[消耗品]アイテムを4個失い、[圧迫:1]を受ける'],
+               [6, '[消耗品]アイテムを4個失い、[圧迫:2]を受ける'],
+               [7, '[消耗品]アイテムを4個失い、[圧迫:3]を受ける'],
+              ],
              ]
 
     return tableName, tables
@@ -282,6 +406,46 @@ MESSAGETEXT
                [5, '所持金を75G失う'],
                [6, '所持金を90G失う'],
                [7, '所持金を100G失う'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '所持金を50G失う'],
+               [2, '所持金を80G失う'],
+               [3, '所持金を110G失う'],
+               [4, '所持金を140G失う'],
+               [5, '所持金を160G失う'],
+               [6, '所持金を190G失う'],
+               [7, '所持金を220G失う'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '所持金を100G失う'],
+               [2, '所持金を150G失う'],
+               [3, '所持金を200G失う'],
+               [4, '所持金を250G失う'],
+               [5, '所持金を300G失う'],
+               [6, '所持金を350G失う'],
+               [7, '所持金を400G失う'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '所持金を160G失う'],
+               [2, '所持金を240G失う'],
+               [3, '所持金を320G失う'],
+               [4, '所持金を400G失う'],
+               [5, '所持金を480G失う'],
+               [6, '所持金を560G失う'],
+               [7, '所持金を640G失う'],
+              ],
+              [
+               [0, '消耗なし'],
+               [1, '所持金を210G失う'],
+               [2, '所持金を320G失う'],
+               [3, '所持金を430G失う'],
+               [4, '所持金を540G失う'],
+               [5, '所持金を650G失う'],
+               [6, '所持金を760G失う'],
+               [7, '所持金を870G失う'],
               ],
              ]
 
@@ -322,8 +486,8 @@ MESSAGETEXT
     #index = (rank - 1) / 5
     index = ((rank - 1) / 5).floor # TKfix Rubyでは常に整数が返るが、JSだと実数になる可能性がある
 
-	index = [0, index].max
-	index = [index, (tables.size - 1)].min
+    index = [0, index].max
+    index = [index, (tables.size - 1)].min
 
     return tables[index]
   end
@@ -353,7 +517,7 @@ MESSAGETEXT
   #財宝表
   def getTresureDiceCommandResult(command)
 
-    return nil unless(/(C|M|I|H|G)TRS(\d*)([\+\-\d]*)(\$)?/ === command)
+    return nil unless(/(C|M|I|O|H|G)TRS(\d*)([\+\-\d]*)(\$)?/ === command)
 
     #TKfix メソッドをまたぐと$xの中身がnilになっている
     reg1 = $1
@@ -361,12 +525,12 @@ MESSAGETEXT
     reg3 = $3
     reg4 = $4
 
-  	type = reg1 #$1
-  	rank = reg2.to_i #$2.to_i
-    is_choice = (reg2.empty? || (not reg4.nil?)) #($2.empty? || (not $4.nil?))
-  	modifyText = reg3 # $3
-  	modify = getValue(modifyText, 0)
-    is_prize = (reg4 == "$") #($4 == "$")
+    type = reg1
+    rank = reg2.to_i
+    is_choice = (reg2.empty? || (not reg4.nil?))
+    modifyText = reg3
+    modify = getValue(modifyText, 0)
+    is_prize = (reg4 == "$")
     dice_value = nil
     dice_value = '7' if is_prize
     is_rank_enable = ( (not is_choice) || is_prize)
@@ -379,6 +543,8 @@ MESSAGETEXT
         getMagicTresureResultTable
       when "I"
         getItemTresureResultTable
+      when "O"
+        getOtherTresureResultTable
       when "H"
         getHeroineTresureResultTable
       when "G"
@@ -389,25 +555,27 @@ MESSAGETEXT
 
     return nil if table.nil?
 
-  	number, dice_str, =
+    number, dice_str, =
       if is_choice
         [dice_value.to_i, dice_value]
       else
         roll(2, 6)
       end
 
-  	number += (rank * (is_rank_enable ? 5 : 0)) + modify
+    number += (rank * (is_rank_enable ? 5 : 0)) + modify
 
     return command if is_choice and (number < getTableMinimum(table))
 
-    number = [getAdjustNumberMin(number, table), 87].min
-
-    result =
-      if type == "H"
-        getHiroineTresureResultString(table, number)
-      else
-        getOtherTresureResultString(table, number)
-      end
+    if type == "H"
+      number = [getAdjustNumberMin(number, table), 87].min
+      result = getHiroineTresureResultString(table, number)
+    elsif type == "G"
+      number = [getAdjustNumberMin(number, table), 87].min
+      result = getOtherTresureResultString(table, number)
+    else
+      number = [getAdjustNumberMin(number, table), 187].min
+      result = getOtherTresureResultStringEx(table, number)
+    end
 
     return "#{tableName}(#{number}#{dice_str ? '[' + dice_str + ']' : ''})：#{result}"
   end
@@ -441,65 +609,181 @@ MESSAGETEXT
     return result
   end
 
+  def getOtherTresureResultStringEx(table, number)
+    result =
+      case number
+      when 163 .. 172
+        get_table_by_number(number - 10, table) + '&200G'
+      when 173 .. 182
+        get_table_by_number(number - 20, table) + '&400G'
+      when 183 .. 187
+        get_table_by_number(number - 30, table) + '&600G'
+      else
+        get_table_by_number(number, table)
+      end
+
+    return result
+  end
+
   def getCashTresureResultTable
     tableName = "金銭財宝表"
     table = [
-             [ 7, '39G'],
-             [ 8, '40G'],
-             [ 9, '42G'],
-             [10, '43G'],
-             [11, '45G'],
-             [12, '46G'],
-             [13, '48G'],
-             [14, '50G'],
-             [15, '52G'],
-             [16, '54G'],
-             [17, '57G'],
-             [18, '59G'],
-             [19, '61G'],
-             [20, '64G'],
-             [21, '67G'],
-             [22, '70G'],
-             [23, '72G'],
-             [24, '75G'],
-             [25, '79G'],
-             [26, '82G'],
-             [27, '85G'],
-             [28, '89G'],
-             [29, '92G'],
-             [30, '96G'],
-             [31, '100G'],
-             [32, '104G'],
-             [33, '108G'],
-             [34, '112G'],
-             [35, '117G'],
-             [36, '121G'],
-             [37, '126G'],
-             [38, '130G'],
-             [39, '135G'],
-             [40, '140G'],
-             [41, '145G'],
-             [42, '150G'],
-             [43, '155G'],
-             [44, '161G'],
-             [45, '166G'],
-             [46, '172G'],
-             [47, '178G'],
-             [48, '183G'],
-             [49, '189G'],
-             [50, '195G'],
-             [51, '201G'],
-             [52, '208G'],
-             [53, '214G'],
-             [54, '221G'],
-             [55, '227G'],
-             [56, '234G'],
-             [57, '241G'],
-             [58, '248G'],
-             [59, '255G'],
-             [60, '262G'],
-             [61, '269G'],
-             [62, '277G'],
+             [  7, '35G'],
+             [  8, '40G'],
+             [  9, '40G'],
+             [ 10, '40G'],
+             [ 11, '45G'],
+             [ 12, '45G'],
+             [ 13, '45G'],
+             [ 14, '50G'],
+             [ 15, '50G'],
+             [ 16, '50G'],
+             [ 17, '55G'],
+             [ 18, '55G'],
+             [ 19, '60G'],
+             [ 20, '60G'],
+             [ 21, '65G'],
+             [ 22, '70G'],
+             [ 23, '70G'],
+             [ 24, '75G'],
+             [ 25, '75G'],
+             [ 26, '80G'],
+             [ 27, '85G'],
+             [ 28, '85G'],
+             [ 29, '90G'],
+             [ 30, '95G'],
+             [ 31, '100G'],
+             [ 32, '100G'],
+             [ 33, '105G'],
+             [ 34, '110G'],
+             [ 35, '115G'],
+             [ 36, '120G'],
+             [ 37, '125G'],
+             [ 38, '130G'],
+             [ 39, '135G'],
+             [ 40, '140G'],
+             [ 41, '145G'],
+             [ 42, '150G'],
+             [ 43, '155G'],
+             [ 44, '160G'],
+             [ 45, '165G'],
+             [ 46, '170G'],
+             [ 47, '175G'],
+             [ 48, '180G'],
+             [ 49, '185G'],
+             [ 50, '195G'],
+             [ 51, '200G'],
+             [ 52, '205G'],
+             [ 53, '210G'],
+             [ 54, '220G'],
+             [ 55, '225G'],
+             [ 56, '230G'],
+             [ 57, '240G'],
+             [ 58, '245G'],
+             [ 59, '255G'],
+             [ 60, '260G'],
+             [ 61, '265G'],
+             [ 62, '275G'],
+             [ 63, '280G'],
+             [ 64, '290G'],
+             [ 65, '300G'],
+             [ 66, '300G'],
+             [ 67, '310G'],
+             [ 68, '320G'],
+             [ 69, '330G'],
+             [ 70, '340G'],
+             [ 71, '340G'],
+             [ 72, '350G'],
+             [ 73, '360G'],
+             [ 74, '370G'],
+             [ 75, '380G'],
+             [ 76, '390G'],
+             [ 77, '400G'],
+             [ 78, '410G'],
+             [ 79, '420G'],
+             [ 80, '430G'],
+             [ 81, '440G'],
+             [ 82, '450G'],
+             [ 83, '460G'],
+             [ 84, '460G'],
+             [ 85, '480G'],
+             [ 86, '490G'],
+             [ 87, '500G'],
+             [ 88, '510G'],
+             [ 89, '520G'],
+             [ 90, '530G'],
+             [ 91, '540G'],
+             [ 92, '550G'],
+             [ 93, '560G'],
+             [ 94, '570G'],
+             [ 95, '580G'],
+             [ 96, '590G'],
+             [ 97, '610G'],
+             [ 98, '620G'],
+             [ 99, '630G'],
+             [100, '640G'],
+             [101, '650G'],
+             [102, '660G'],
+             [103, '680G'],
+             [104, '690G'],
+             [105, '700G'],
+             [106, '710G'],
+             [107, '730G'],
+             [108, '740G'],
+             [109, '750G'],
+             [110, '760G'],
+             [111, '780G'],
+             [112, '790G'],
+             [113, '800G'],
+             [114, '820G'],
+             [115, '830G'],
+             [116, '840G'],
+             [117, '860G'],
+             [118, '870G'],
+             [119, '890G'],
+             [120, '900G'],
+             [121, '910G'],
+             [122, '930G'],
+             [123, '940G'],
+             [124, '960G'],
+             [125, '970G'],
+             [126, '990G'],
+             [127, '1000G'],
+             [128, '1020G'],
+             [129, '1030G'],
+             [130, '1050G'],
+             [131, '1060G'],
+             [132, '1080G'],
+             [133, '1090G'],
+             [134, '1110G'],
+             [135, '1130G'],
+             [136, '1140G'],
+             [137, '1160G'],
+             [138, '1170G'],
+             [139, '1190G'],
+             [140, '1210G'],
+             [141, '1220G'],
+             [142, '1240G'],
+             [143, '1260G'],
+             [144, '1270G'],
+             [145, '1290G'],
+             [146, '1310G'],
+             [147, '1330G'],
+             [148, '1340G'],
+             [149, '1360G'],
+             [150, '1380G'],
+             [151, '1400G'],
+             [152, '1410G'],
+             [153, '1430G'],
+             [154, '1450G'],
+             [155, '1470G'],
+             [156, '1490G'],
+             [157, '1500G'],
+             [158, '1520G'],
+             [159, '1540G'],
+             [160, '1560G'],
+             [161, '1580G'],
+             [162, '1600G'],
             ]
 
     return tableName, table
@@ -508,62 +792,162 @@ MESSAGETEXT
   def getMagicTresureResultTable
     tableName = "魔法素材財宝表"
     table = [
-             [ 7, '魔触媒2[魔触媒2](20G)x2'],
-             [ 8, '魔触媒2[魔触媒2](20G)x2'],
-             [ 9, '魔触媒1[魔触媒1](15G)x3'],
-             [10, '魔触媒1[魔触媒1](15G)x3'],
-             [11, '魔触媒1[魔触媒1](15G)x3'],
-             [12, '腐銀の小片[コア素材](30G)&鉱石のサンプル[換金](15G)'],
-             [13, '新緑の若芽[コア素材](30G)&小さな花の種[換金](20G)'],
-             [14, '命の葉[コア素材](40G)&強靭なツタ[換金](10G)'],
-             [15, '鋭い牙[コア素材](40G)&使い込まれたナイフ[換金](10G)'],
-             [16, '魔触媒3[魔触媒3](25G)x2'],
-             [17, '古びた髑髏[コア素材](40G)&黒い詩集[換金](15G)'],
-             [18, '黒睡蓮の花弁[コア素材](40G)&水キセル[換金](20G)'],
-             [19, '純白の羽根[コア素材](40G)&小さなゴーグル[換金](20G)'],
-             [20, '真紅の爪[コア素材](50G)&小さな鏡[換金](15G)'],
-             [21, '赤熱の小爪[コア素材](50G)&細工物の暖炉[換金](20G)'],
-             [22, '自戒の茨[コア素材](50G)&焦げた竜革[換金](20G)'],
-             [23, '流星のかけら[コア素材](50G)&質素な指輪[換金](10G)x2'],
-             [24, '聖なる繊維[コア素材](50G)&傷ついたメダル[換金](25G)'],
-             [25, '折れたシャフト[コア素材](60G)&ガラクタの山[換金](5G)x4'],
-             [26, '巨人の髭[コア素材](50G)&小さな櫛[換金](20G)'],
-             [27, '精密な歯車[コア素材](60G)&工具鋼の破片[換金](25G)'],
-             [28, 'とがった爪[コア素材](60G)&刺繍の飾り帯[換金](30G)'],
-             [29, '触媒のフラスコ[コア素材](60G)&小さな肖像画[換金](30G)'],
-             [30, '液化魔晶[コア素材](60G)&精密な人形[換金](35G)'],
-             [31, '魔触媒5[魔触媒5](40G)&魔触媒4[魔触媒4](30G)x2'],
-             [32, 'やわらかい石[コア素材](80G)&小型の蒸留器[換金](20G)'],
-             [33, '謎めいた毛皮[コア素材](80G)&奇妙な頭骨[換金](30G)'],
-             [34, '星屑の銀糸[コア素材](80G)&錫引きの星見盤[換金](30G)'],
-             [35, '魔触媒5[魔触媒5](40G)x3'],
-             [36, '小型錬成陣[コア素材](100G)&奇妙なオブジェ[換金](20G)'],
-             [37, '魔触媒6[魔触媒6](50G)&魔触媒5[魔触媒5](40G)x2'],
-             [38, '呪紋の種[コア素材](100G)&砕けた宝石[換金](30G)'],
-             [39, '銀の円環[コア素材](100G)&純白のコサージュ[換金](30G)'],
-             [40, '常眠りの種子[コア素材](100G)&庭師の飾り紐[換金](40G)'],
-             [41, '飴色の粘液[コア素材](1210G)&極彩色の粒[換金](5G)x5'],
-             [42, '魔触媒6[魔触媒6](50G)x3'],
-             [43, '銀色の甲殻[コア素材](120G)&乏しいつぼみ[換金](10G)x4'],
-             [44, '魔触媒7[魔触媒7](60G)&魔触媒6[魔触媒6](50G)x2'],
-             [45, '大きな魔力結晶[コア素材](120G)&妖精のラクガキ[換金](40G)'],
-             [46, '拳士の魄片[コア素材](140G)&白紙の巻物[換金](40G)'],
-             [47, '魔触媒7[魔触媒7](60G)x3'],
-             [48, '蒼い鉱石[コア素材](140G)&天然ガラスの塊[換金](25G)x2'],
-             [49, '巨大な風切り羽[コア素材](140G)&分厚い卵殻[換金](50G)'],
-             [50, '魔触媒8[魔触媒8](70G)x2&魔触媒7[魔触媒7](60G)'],
-             [51, '絶えない火種[コア素材](180G)&漆黒のスス[換金](20G)'],
-             [52, '魔触媒8[魔触媒8](70G)x3'],
-             [53, '魔導錠[コア素材](180G)&つくりかけの錠前[換金](10G)x3'],
-             [54, '七色の透明捻子[コア素材](180G)&開かない細工箱[換金](40G)'],
-             [55, '偏属性魔法結晶[コア素材](220G)&レポートの束[換金](10G)'],
-             [56, '魔触媒9[魔触媒9](90G)&魔触媒8[魔触媒8](70G)x2'],
-             [57, '魔触媒9[魔触媒9](90G)&魔触媒8[魔触媒8](70G)x2'],
-             [58, '派手な羽根飾り[コア素材](220G)&空想動物図鑑[換金](30G)'],
-             [59, '砂のバラ[コア素材](220G)&棘のような水晶[換金](10G)x5'],
-             [60, '魔触媒9[魔触媒9](90G)x3'],
-             [61, '頑丈な胃袋[コア素材](220G)&奇妙な標本[換金](50G)'],
-             [62, '魔触媒10[魔触媒10](110G)&魔触媒9[魔触媒9](90G)x2'],
+             [  7, '魔触媒2[魔触媒2](20G)&魔触媒1[魔触媒1](15G)'],
+             [  8, '新緑の若芽[コア素材](30G)'],
+             [  9, '魔触媒2[魔触媒2](20G)&魔触媒1[魔触媒1](15G)'],
+             [ 10, '怪しい丸薬[コア素材](30G)'],
+             [ 11, '魔触媒2[魔触媒2](20G)&魔触媒1[魔触媒1](15G)'],
+             [ 12, '滑らかな被膜[コア素材](30G)'],
+             [ 13, '魔触媒2[魔触媒2](20G)&魔触媒1[魔触媒1](15G)'],
+             [ 14, '呪法シリンダー[コア素材](30G)'],
+             [ 15, '魔触媒3[魔触媒3](25G)&魔触媒2[魔触媒2](20G)'],
+             [ 16, '反魔水銀[コア素材](40G)'],
+             [ 17, '魔触媒3[魔触媒3](25G)&魔触媒2[魔触媒2](20G)'],
+             [ 18, '鋭い牙[コア素材](40G)'],
+             [ 19, '魔触媒3[魔触媒3](25G)&魔触媒2[魔触媒2](20G)'],
+             [ 20, '自戒の茨[コア素材](50G)'],
+             [ 21, '魔触媒4[魔触媒4](30G)&魔触媒3[魔触媒3](25G)'],
+             [ 22, '流星のかけら[コア素材](50G)'],
+             [ 23, '魔触媒4[魔触媒4](30G)&魔触媒3[魔触媒3](25G)'],
+             [ 24, '聖なる繊維[コア素材](50G)'],
+             [ 25, '魔触媒5[魔触媒5](40G)&魔触媒4[魔触媒4](30G)'],
+             [ 26, '巨人の髭[コア素材](60G)'],
+             [ 27, '魔触媒5[魔触媒5](40G)&魔触媒4[魔触媒4](30G)'],
+             [ 28, '精密な歯車[コア素材](60G)'],
+             [ 29, '魔触媒5[魔触媒5](40G)&魔触媒4[魔触媒4](30G)'],
+             [ 30, '謎めいた毛皮[コア素材](80G)'],
+             [ 31, '魔触媒6[魔触媒6](50G)&魔触媒5[魔触媒5](40G)'],
+             [ 32, '雄々しい角[コア素材](80G)'],
+             [ 33, '魔触媒6[魔触媒6](50G)&魔触媒5[魔触媒5](40G)'],
+             [ 34, 'つややかな繭[コア素材](80G)'],
+             [ 35, '魔触媒7[魔触媒7](60G)&魔触媒6[魔触媒6](50G)'],
+             [ 36, '小型錬成陣[コア素材](100G)'],
+             [ 37, '魔触媒7[魔触媒7](60G)&魔触媒6[魔触媒6](50G)'],
+             [ 38, '銀の円環[コア素材](100G)'],
+             [ 39, '魔触媒7[魔触媒7](60G)&魔触媒6[魔触媒6](50G)'],
+             [ 40, '仙桃果[コア素材](120G)'],
+             [ 41, '魔触媒8[魔触媒8](70G)&魔触媒7[魔触媒7](60G)'],
+             [ 42, '動力ケーブル[コア素材](120G)'],
+             [ 43, '魔触媒8[魔触媒8](70G)&魔触媒7[魔触媒7](60G)'],
+             [ 44, '流星歯車[コア素材](120G)'],
+             [ 45, '魔触媒9[魔触媒9](90G)&魔触媒8[魔触媒8](70G)'],
+             [ 46, '茨のトゲ[コア素材](140G)'],
+             [ 47, '魔触媒9[魔触媒9](90G)&魔触媒8[魔触媒8](70G)'],
+             [ 48, '砕けた剛剣[コア素材](140G)'],
+             [ 49, '魔触媒9[魔触媒9](90G)&魔触媒8[魔触媒8](70G)'],
+             [ 50, '絶えない火種[コア素材](180G)'],
+             [ 51, '魔触媒10[魔触媒10](110G)&魔触媒9[魔触媒9](90G)'],
+             [ 52, '魔導錠[コア素材](180G)'],
+             [ 53, '魔触媒10[魔触媒10](110G)&魔触媒9[魔触媒9](90G)'],
+             [ 54, '青ざめた盾鱗[コア素材](180G)'],
+             [ 55, '魔触媒10[魔触媒10](110G)x2'],
+             [ 56, '破軍の戦帯[コア素材](220G)'],
+             [ 57, '魔触媒11[魔触媒11](120G)&魔触媒10[魔触媒10](110G)'],
+             [ 58, '派手な羽根飾り[コア素材](220G)'],
+             [ 59, '魔触媒11[魔触媒11](120G)&魔触媒10[魔触媒10](110G)'],
+             [ 60, '神隠しの古枝[コア素材](240G)'],
+             [ 61, '魔触媒11[魔触媒11](120G)x2'],
+             [ 62, '呪紋水晶[コア素材](240G)'],
+             [ 63, '魔触媒12[魔触媒12](150G)&魔触媒11[魔触媒11](120G)'],
+             [ 64, '謎めいたフラスコ[コア素材](240G)'],
+             [ 65, '魔触媒12[魔触媒12](150G)&魔触媒11[魔触媒11](120G)'],
+             [ 66, '未熟な竜玉[コア素材](300G)'],
+             [ 67, '魔触媒12[魔触媒12](150G)x2'],
+             [ 68, '陰陽の水銀[コア素材](300G)'],
+             [ 69, '魔触媒13[魔触媒13](170G)&魔触媒12[魔触媒12](150G)'],
+             [ 70, '金剛骨[コア素材](340G)'],
+             [ 71, '魔触媒13[魔触媒13](170G)x2'],
+             [ 72, 'アルヴの研磨剤[コア素材](340G)'],
+             [ 73, '魔触媒14[魔触媒14](190G)&魔触媒13[魔触媒13](170G)'],
+             [ 74, '不死王の心臓[コア素材](340G)'],
+             [ 75, '魔触媒14[魔触媒14](190G)x2'],
+             [ 76, '緑小鬼の大軍旗[コア素材](380G)'],
+             [ 77, '魔触媒14[魔触媒14](190G)x2'],
+             [ 78, '深層アダマン鉱石[コア素材](380G)'],
+             [ 79, '魔触媒15[魔触媒15](220G)&魔触媒14[魔触媒14](190G)'],
+             [ 80, '漆黒の眼球[コア素材](440G)'],
+             [ 81, '魔触媒15[魔触媒15](220G)&魔触媒14[魔触媒14](190G)'],
+             [ 82, '雪桃の実[コア素材](440G)'],
+             [ 83, '魔触媒15[魔触媒15](220G)x2'],
+             [ 84, '四つ葉のアンク[コア素材](440G)'],
+             [ 85, '魔触媒16[魔触媒16](250G)&魔触媒15[魔触媒15](220G)'],
+             [ 86, '金色のたてがみ[コア素材](500G)'],
+             [ 87, '魔触媒16[魔触媒16](250G)&魔触媒15[魔触媒15](220G)'],
+             [ 88, '銀の車輪[コア素材](500G)'],
+             [ 89, '魔触媒16[魔触媒16](250G)x2'],
+             [ 90, '封緑樹の堅枝[コア素材](560G)'],
+             [ 91, '魔触媒17[魔触媒17](280G)&魔触媒16[魔触媒16](250G)'],
+             [ 92, '千年王樹の種子[コア素材](560G)'],
+             [ 93, '魔触媒17[魔触媒17](280G)x2'],
+             [ 94, '巨鳥の鎖骨[コア素材](560G)'],
+             [ 95, '魔触媒17[魔触媒17](280G)x2'],
+             [ 96, '水棲緑鬼の王卵[コア素材](620G)'],
+             [ 97, '魔触媒18[魔触媒18](310G)&魔触媒17[魔触媒17](280G)'],
+             [ 98, '太陽の飾り鋲[コア素材](620G)'],
+             [ 99, '魔触媒18[魔触媒18](310G)x2'],
+             [100, '火精霊の心核[コア素材](680G)'],
+             [101, '魔触媒19[魔触媒19](340G)&魔触媒18[魔触媒18](310G)'],
+             [102, '獄炎狼の房毛[コア素材](680G)'],
+             [103, '魔触媒19[魔触媒19](340G)x2'],
+             [104, '島クジラの巨骨[コア素材](680G)'],
+             [105, '魔触媒19[魔触媒19](340G)x2'],
+             [106, '原初の炎[コア素材](740G)'],
+             [107, '魔触媒20[魔触媒20](370G)&魔触媒19[魔触媒19](340G)'],
+             [108, '魔石草の種[コア素材](740G)'],
+             [109, '魔触媒20[魔触媒20](370G)x2'],
+             [110, '稠密立方魔晶[コア素材](820G)'],
+             [111, '魔触媒21[魔触媒21](410G)&魔触媒20[魔触媒20](370G)'],
+             [112, '清掃用女中人形九号[コア素材](820G)'],
+             [113, '魔触媒21[魔触媒21](410G)&魔触媒20[魔触媒20](370G)'],
+             [114, '飛びイルカの心臓[コア素材](820G)'],
+             [115, '魔触媒21[魔触媒21](410G)x2'],
+             [116, '奈落皇蛇の額鱗[コア素材](900G)'],
+             [117, '魔触媒22[魔触媒22](450G)&魔触媒21[魔触媒21](410G)'],
+             [118, 'ブリザードダイヤ[コア素材](900G)'],
+             [119, '魔触媒22[魔触媒22](450G)&魔触媒21[魔触媒21](410G)'],
+             [120, '火炎向日葵の花びら[コア素材](980G)'],
+             [121, '魔触媒22[魔触媒22](450G)x2'],
+             [122, '紅蓮蝙蝠の鮮血[コア素材](980G)'],
+             [123, '魔触媒23[魔触媒23](490G)&魔触媒22[魔触媒22](450G)'],
+             [124, '落日祈祷書[コア素材](980G)'],
+             [125, '魔触媒24[魔触媒24](530G)&黒檀の馬具[換金](440G)'],
+             [126, '辰砂の組紐[コア素材](1060G)'],
+             [127, '魔触媒23[魔触媒23](490G)x2'],
+             [128, '黴臭い皮の剣帯[コア素材](1060G)'],
+             [129, '魔触媒24[魔触媒24](530G)&魔触媒23[魔触媒23](490G)'],
+             [130, '玻璃蜥蜴の抜け殻[コア素材](1160G)'],
+             [131, '魔触媒24[魔触媒24](530G)x2'],
+             [132, '火炎巨人の尺骨[コア素材](1160G)'],
+             [133, '魔触媒24[魔触媒24](530G)x2'],
+             [134, '高熱練炭[コア素材](1160G)'],
+             [135, '魔触媒26[魔触媒26](620G)&青銅の装置[換金](510G)'],
+             [136, 'すり切れた羽衣[コア素材](1240G)'],
+             [137, '魔触媒25[魔触媒25](580G)x2'],
+             [138, '亡霊のため息[コア素材](1240G)'],
+             [139, '魔触媒25[魔触媒25](580G)x2'],
+             [140, '宵闇色のヒゲ[コア素材](1340G)'],
+             [141, '魔触媒27[魔触媒27](670G)&月長石の腕輪[換金](550G)'],
+             [142, 'コアトルの毒腺[コア素材](1340G)'],
+             [143, '魔触媒26[魔触媒26](620G)x2'],
+             [144, '玄亀の甲羅[コア素材](1340G)'],
+             [145, '魔触媒28[魔触媒28](720G)&宝玉のネックレス[換金](570G)'],
+             [146, '山大鹿の角[コア素材](1440G)'],
+             [147, '魔触媒28[魔触媒28](720G)&消えない燭台[換金](600G)'],
+             [148, '煉獄クモの燃糸[コア素材](1540G)'],
+             [149, '魔触媒27[魔触媒27](670G)x2'],
+             [150, '碧氷晶の単眼球[コア素材](1540G)'],
+             [151, '魔触媒29[魔触媒29](770G)&琥珀のイヤリング[換金](630G)'],
+             [152, '衝撃竜の牙[コア素材](1540G)'],
+             [153, '魔触媒29[魔触媒29](770G)&茨を模したティアラ[換金](650G)'],
+             [154, '鉄鼠の自在尾[コア素材](1540G)'],
+             [155, '魔触媒30[魔触媒30](820G)&石榴石の腕輪[換金](650G)'],
+             [156, '人工生命の素[コア素材](1640G)'],
+             [157, '魔触媒30[魔触媒30](820G)&輝くメガネ[換金](680G)'],
+             [158, '翡翠甲虫の刃翅[コア素材](1640G)'],
+             [159, '魔触媒29[魔触媒29](770G)x2'],
+             [160, '無限渦流の煤[コア素材](1640G)'],
+             [161, '魔触媒30[魔触媒30](820G)&スライム入り試験管[換金](750G)'],
+             [162, '陽喰い鳥の嘴[コア素材](1640G)'],
             ]
 
     return tableName, table
@@ -572,62 +956,326 @@ MESSAGETEXT
   def getItemTresureResultTable
     tableName = "換金アイテム財宝表"
     table = [
-             [ 7, '陶器の絵付きマグカップ[換金](40G)'],
-             [ 8, '木製の騎士像[換金](40G)'],
-             [ 9, '小さな風景画[換金](50G)'],
-             [10, '奇妙な抽象画[換金](50G)'],
-             [11, '夜会の仮面[換金](50G)'],
-             [12, '錫の食器セット[換金](50G)'],
-             [13, '古い詩集[換金](50G)'],
-             [14, '鮮やかな刺繍のハンカチ[換金](50G)'],
-             [15, '陶器の大皿[換金](60G)'],
-             [16, '絵巻物[換金](60G)'],
-             [17, '陶器の水盤[換金](60G)'],
-             [18, '細工物のイヤリング[換金](60G)'],
-             [19, '淑女の肖像画[換金](70G)'],
-             [20, '小さな宝石箱[換金](70G)'],
-             [21, '真鍮の燭台[換金](70G)'],
-             [22, '貴族の古い日記[換金](70G)'],
-             [23, '騎士の肖像画[換金](80G)'],
-             [24, '塗り下駄[換金](80G)'],
-             [25, '精緻なゲーム盤[換金](80G)'],
-             [26, '精緻な静物画[換金](90G)'],
-             [27, '樫の椅子[換金](90G)'],
-             [28, '古いビスクドール[換金](90G)'],
-             [29, 'きらびやかな仮面[換金](100G)'],
-             [30, 'ウサギのぬいぐるみ[換金](100G)'],
-             [31, '真鍮の子鬼像[換金](100G)'],
-             [32, '地方の歴史書[換金](110G)'],
-             [33, '夜会の手袋[換金](110G)'],
-             [34, 'ハイヒール[換金](120G)'],
-             [35, 'おとぎ話の本[換金](120G)'],
-             [36, '少女の肖像画[換金](130G)'],
-             [37, '小さなコケシ[換金](130G)'],
-             [38, '藤の椅子[換金](130G)'],
-             [39, 'イヌとネコのパペット[換金](140G)'],
-             [40, '掛け軸[換金](G140)'],
-             [41, 'クマのぬいぐるみ[換金](150G)'],
-             [42, 'ネコのかぶりもの[換金](150G)'],
-             [43, '学術書[換金](160G)'],
-             [44, 'タカの剥製[換金](170G)'],
-             [45, '大理石の賢者像[換金](170G)'],
-             [46, '刺繍をあしらったクッション[換金](180G)'],
-             [47, 'イヌのかぶりもの[換金](180G)'],
-             [48, '樫の大テーブル[換金](190G)'],
-             [49, '彫金の指輪[換金](190G)'],
-             [50, '花鳥画の掛け軸[換金](200G)'],
-             [51, '最上質の毛布[換金](210G)'],
-             [52, '上質の白粉[換金](210G)'],
-             [53, '磁器の茶器[換金](220G)'],
-             [54, '簡素なティアラ[換金](230G)'],
-             [55, '彫金のイヤリング[換金](230G)'],
-             [56, '豪華なネックレス[換金](240G)'],
-             [57, '上質の香水[換金](250G)'],
-             [58, '山水画の掛け軸[換金](250G)'],
-             [59, '手の込んだドレス/礼服[換金](260G)'],
-             [60, '陶器の絵皿[換金](270G)'],
-             [61, 'キツネのぬいぐるみ[換金](270G)'],
-             [62, '古い歴史書[換金](280G)'],
+             [  7, '破れた書籍[換金](30G)'],
+             [  8, '古ぼけたぬいぐるみ[換金](40G)'],
+             [  9, 'あやしい書籍[換金](40G)'],
+             [ 10, '分厚い書籍[換金](40G)'],
+             [ 11, '子供向けの書籍[換金](40G)'],
+             [ 12, '犬のぬいぐるみ[換金](40G)'],
+             [ 13, '使い込まれたティーセット[換金](40G)'],
+             [ 14, 'クマのぬいぐるみ[換金](50G)'],
+             [ 15, '白磁のティーセット[換金](50G)'],
+             [ 16, '藍色の染料[換金](50G)'],
+             [ 17, '毛糸の帽子[換金](50G)'],
+             [ 18, '恐竜のぬいぐるみ[換金](50G)'],
+             [ 19, '豪華な日記[換金](60G)'],
+             [ 20, '乳白色の角[換金](60G)'],
+             [ 21, '羽根付き帽子[換金](60G)'],
+             [ 22, '魔法の書籍[換金](70G)'],
+             [ 23, '木製の食器[換金](70G)'],
+             [ 24, '猛獣の毛皮[換金](70G)'],
+             [ 25, '魔法使いの帽子[換金](70G)'],
+             [ 26, '貴族の絵本[換金](80G)'],
+             [ 27, '陶製の食器[換金](80G)'],
+             [ 28, '磨かれたティーセット[換金](80G)'],
+             [ 29, 'しなやかな毛皮[換金](90G)'],
+             [ 30, '巨大な角[換金](90G)'],
+             [ 31, '貴重な書籍[換金](100G)'],
+             [ 32, '銀の食器[換金](100G)'],
+             [ 33, '魔獣の毛皮[換金](100G)'],
+             [ 34, '司祭の帽子[換金](110G)'],
+             [ 35, '野獣の牙[換金](110G)'],
+             [ 36, '巨人のスプーン[換金](120G)'],
+             [ 37, '枝分かれした角[換金](120G)'],
+             [ 38, '猫耳帽子[換金](130G)'],
+             [ 39, '木綿のドレス[換金](130G)'],
+             [ 40, '水晶の角[換金](140G)'],
+             [ 41, '琥珀の円盤[換金](140G)'],
+             [ 42, '幻獣の牙[換金](150G)'],
+             [ 43, '古代のドレス[換金](150G)'],
+             [ 44, '王朝風のティーセット[換金](160G)'],
+             [ 45, '傷ひとつない毛皮[換金](160G)'],
+             [ 46, '絹のドレス[換金](170G)'],
+             [ 47, '魔法のぬいぐるみ[換金](170G)'],
+             [ 48, '神獣の毛皮[換金](180G)'],
+             [ 49, '水晶の食器[換金](180G)'],
+             [ 50, '石の呪具[換金](190G)'],
+             [ 51, '禍々しい仮面[換金](200G)'],
+             [ 52, 'ねず耳カチューシャ[換金](200G)'],
+             [ 53, '竜の牙[換金](210G)'],
+             [ 54, 'スライム？の剥製[換金](220G)'],
+             [ 55, '銀のネジ[換金](220G)'],
+             [ 56, '大猫の剥製[換金](230G)'],
+             [ 57, '舞踏会の仮面[換金](240G)'],
+             [ 58, '重厚な牙[換金](240G)'],
+             [ 59, '宝石細工の仮面[換金](250G)'],
+             [ 60, 'サテンのドレス[換金](260G)'],
+             [ 61, '光る角[換金](260G)'],
+             [ 62, '魔法のドレス[換金](270G)'],
+             [ 63, '道化師の仮面[換金](280G)'],
+             [ 64, 'アルヴのティーセット[換金](290G)'],
+             [ 65, '燃えない毛皮[換金](300G)'],
+             [ 66, '雪狼の剥製[換金](300G)'],
+             [ 67, '籐製の椅子[換金](310G)'],
+             [ 68, '赤い石板[換金](320G)'],
+             [ 69, '黒獅子の剥製[換金](330G)'],
+             [ 70, '三角の椅子[換金](340G)'],
+             [ 71, '子供用紳士服[換金](340G)'],
+             [ 72, 'アルヴの食器[換金](350G)'],
+             [ 73, 'ほつれた紳士服[換金](360G)'],
+             [ 74, '上等な紳士服[換金](370G)'],
+             [ 75, '寄木細工の文箱[換金](380G)'],
+             [ 76, '白磁の燭台[換金](390G)'],
+             [ 77, '黄金の牙[換金](400G)'],
+             [ 78, '白木の文箱[換金](410G)'],
+             [ 79, '青銅の燭台[換金](420G)'],
+             [ 80, '輝く燭台[換金](430G)'],
+             [ 81, '王様の椅子[換金](440G)'],
+             [ 82, '古風な紳士服[換金](450G)'],
+             [ 83, '思い出のドレス[換金](460G)'],
+             [ 84, '古代の種[換金](460G)'],
+             [ 85, '小さな宝石箱[換金](480G)'],
+             [ 86, 'ひびわれた姿見[換金](490G)'],
+             [ 87, '楓の書棚[換金](500G)'],
+             [ 88, '神秘的な仮面[換金](510G)'],
+             [ 89, 'トネリコの書見台[換金](520G)'],
+             [ 90, 'ビロードの紳士服[換金](530G)'],
+             [ 91, '磨きこまれた姿見[換金](540G)'],
+             [ 92, '大きな宝石箱[換金](550G)'],
+             [ 93, '白檀の文箱[換金](560G)'],
+             [ 94, '白樺の姿見[換金](570G)'],
+             [ 95, '焼き物の像[換金](580G)'],
+             [ 96, '彫刻付き姿見[換金](590G)'],
+             [ 97, '粘土の像[換金](610G)'],
+             [ 98, '銀の燭台[換金](620G)'],
+             [ 99, '鳳凰の剥製[換金](630G)'],
+             [100, '呪われた像[換金](640G)'],
+             [101, '魔王の椅子[換金](650G)'],
+             [102, '神域の椅子[換金](660G)'],
+             [103, '満杯の宝石箱[換金](680G)'],
+             [104, '大理石の貨幣[換金](690G)'],
+             [105, '絵の具のセット[換金](700G)'],
+             [106, '染料のセット[換金](710G)'],
+             [107, '木製の貨幣[換金](730G)'],
+             [108, '貝殻の貨幣[換金](740G)'],
+             [109, '覆い付き書見台[換金](750G)'],
+             [110, '手作りのゲーム盤[換金](760G)'],
+             [111, '祭祀用の像[換金](780G)'],
+             [112, '真鍮製のゲーム盤[換金](790G)'],
+             [113, '螺鈿の文箱[換金](800G)'],
+             [114, '青銅の書見台[換金](820G)'],
+             [115, '香木の小像[換金](830G)'],
+             [116, '魔法使いの像[換金](840G)'],
+             [117, '鍵のかかった宝石箱[換金](860G)'],
+             [118, 'いかさまゲーム盤[換金](870G)'],
+             [119, '魔法の紳士服[換金](890G)'],
+             [120, '化粧品のセット[換金](900G)'],
+             [121, '古代の姫の小像[換金](910G)'],
+             [122, '翡翠の小像[換金](930G)'],
+             [123, '真鍮の鉱石[換金](940G)'],
+             [124, '真鉄の塊[換金](960G)'],
+             [125, '銀の塊[換金](970G)'],
+             [126, '古代の貨幣[換金](990G)'],
+             [127, '瑪瑙の文箱[換金](1000G)'],
+             [128, '銀細工の書見台[換金](1020G)'],
+             [129, '粗末な馬具[換金](1030G)'],
+             [130, 'あふれた宝石箱[換金](1050G)'],
+             [131, '異文明の貨幣[換金](1060G)'],
+             [132, '透明な文箱[換金](1080G)'],
+             [133, '香水のセット[換金](1090G)'],
+             [134, '鳥が象られたティアラ[換金](1110G)'],
+             [135, '金細工の燭台[換金](1130G)'],
+             [136, '重厚な馬具[換金](1140G)'],
+             [137, '香辛料のセット[換金](1160G)'],
+             [138, '太陽のティアラ[換金](1170G)'],
+             [139, '豪華な姿見[換金](1190G)'],
+             [140, '月のティアラ[換金](1210G)'],
+             [141, '水晶のメガネ[換金](1220G)'],
+             [142, '魔法銀の塊[換金](1240G)'],
+             [143, '硝子のメガネ[換金](1260G)'],
+             [144, '金細工のゲーム盤[換金](1270G)'],
+             [145, '金剛石のメガネ[換金](1290G)'],
+             [146, '品のあるネックレス[換金](1310G)'],
+             [147, '植物の標本[換金](1330G)'],
+             [148, 'くすんだ首輪[換金](1340G)'],
+             [149, '空色鋼の塊[換金](1360G)'],
+             [150, '騎士の像[換金](1380G)'],
+             [151, 'ちぎれたネックレス[換金](1400G)'],
+             [152, '漆黒の呪具[換金](1410G)'],
+             [153, '精巧な馬具[換金](1430G)'],
+             [154, '飾り気のないティアラ[換金](1450G)'],
+             [155, '動物の標本[換金](1470G)'],
+             [156, '人造人間の標本[換金](1490G)'],
+             [157, 'しゃべる姿見[換金](1500G)'],
+             [158, '精巧なゲーム盤[換金](1520G)'],
+             [159, 'アルヴの貨幣[換金](1540G)'],
+             [160, '豪奢なネックレス[換金](1560G)'],
+             [161, '猫目石の腕輪[換金](1580G)'],
+             [162, '藍玉の腕輪[換金](1600G)'],
+            ]
+
+    return tableName, table
+  end
+
+  def getOtherTresureResultTable
+    tableName = "そのほか財宝表"
+    table = [
+             [  7, '百科事典(40G)'],
+             [  8, 'ライトメイス(40G)'],
+             [  9, '蘇生の宝珠(初級)(40G)'],
+             [ 10, '季節の野菜ポトフ(20G)x2'],
+             [ 11, '荷馬車の笛(45G)'],
+             [ 12, 'モンスター図鑑(初級)(40G)&お好みサンドイッチ(10G)'],
+             [ 13, 'ファインダガー(50G)'],
+             [ 14, 'パッデドアーマー[破損](55G)'],
+             [ 15, '鉄壁符(初級)(45G)'],
+             [ 16, 'ショートボウ[破損](55G)'],
+             [ 17, 'マーシャルアーツドレス[破損](60G)'],
+             [ 18, 'ケルピーの鱗(60G)'],
+             [ 19, 'クレセントバーガー(30G)x2'],
+             [ 20, 'シェフの気まぐれピザ(60G)'],
+             [ 21, 'スモールシールド(70G)'],
+             [ 22, 'ウィローバンド[破損](30G)&暗闇透視薬(初級)(40G)'],
+             [ 23, 'スモールシールド[破損](35G)&魔精の小瓶(初級)(40G)'],
+             [ 24, '鉢がね(80G)'],
+             [ 25, 'ベリーのタルト(40G)x2'],
+             [ 26, '防水バッグ[破損](35G)&水中活動の巻物(初級)(50G)'],
+             [ 27, 'スティレット(90G)'],
+             [ 28, '飛行の巻物(初級)(50G)&強酸毒(初級)(40G)'],
+             [ 29, 'スプリントメイル[破損](100G)'],
+             [ 30, '異次元バッグ(初級)[破損](45G)&おにぎり(50G)'],
+             [ 31, 'ウッデンラウンド(100G)'],
+             [ 32, '野太刀[破損](110G)'],
+             [ 33, '保護の巻物(初級)(70G)&消耗毒(初級)(50G)'],
+             [ 34, 'ライトクロスボウ[破損](115G)'],
+             [ 35, '良質な楽器(110G)'],
+             [ 36, 'ハンターフード[破損](60G)&猛毒薬(中級)(60G)'],
+             [ 37, 'とんすとんの焼き串盛り合わせ(120G)'],
+             [ 38, 'クィルブイリ[破損](135G)'],
+             [ 39, 'ケープ(140G)'],
+             [ 40, 'トートバッグ(140G)'],
+             [ 41, '三つ目鳥の卵炒飯(70G)&衰弱毒(中級)(80G)'],
+             [ 42, '範囲隠密の巻物(初級)(150G)'],
+             [ 43, 'アイアンシールド(150G)'],
+             [ 44, 'クイックカード(初級)(170G)'],
+             [ 45, 'イーグルカード(初級)(170G)'],
+             [ 46, '和弓[破損](170G)'],
+             [ 47, 'エストック[破損](170G)'],
+             [ 48, '食料バッグ(170G)'],
+             [ 49, 'トポト親方のコロッケ(90G)&拘束毒(中級)(100G)'],
+             [ 50, 'モーニングスター(200G)'],
+             [ 51, 'シーダーのスタッフ[破損](200G)'],
+             [ 52, 'カイトシールド(220G)'],
+             [ 53, '怪鳥肉の水炊き(210G)'],
+             [ 54, 'チェインシャツ[破損](245G)'],
+             [ 55, '大弓[破損](240G)'],
+             [ 56, '範囲飛行の巻物(初級)(210G)'],
+             [ 57, 'ポーションホルダー(中級)(240G)'],
+             [ 58, 'アウトレイジ[破損](240G)'],
+             [ 59, '定番カレー(130G)x2'],
+             [ 60, '蘇生の宝珠(中級)(240G)'],
+             [ 61, '招命の宝珠(中級)(120G)&焼きそばパン(150G)'],
+             [ 62, '異次元バッグ(中級)(280G)'],
+             [ 63, '名工の楽器(280G)'],
+             [ 64, '月狼の牙[破損](180G)&解毒薬(中級)(130G)'],
+             [ 65, '魚鱗の宝珠(中級)(280G)'],
+             [ 66, '防水ポーチ(320G)'],
+             [ 67, '収納シールド(320G)'],
+             [ 68, '火炎のイヤリング(330G)'],
+             [ 69, '絹の帽子(350G)'],
+             [ 70, '幻翼の宝珠(中級)(320G)'],
+             [ 71, 'バグナウ[破損](340G)'],
+             [ 72, 'ロングボウ[破損](370G)'],
+             [ 73, 'ショテル[破損](370G)'],
+             [ 74, 'デッキシールド[破損](185G)&チキンボックス(200G)'],
+             [ 75, 'コルセスカ[破損](365G)'],
+             [ 76, 'サイプレスのスタッフ[破損](425G)'],
+             [ 77, 'チェーンシックル[破損](230G)&苺のふんわりカップケーキ(230G)'],
+             [ 78, 'ブロードアックス(410G)'],
+             [ 79, 'クロスシールド(470G)'],
+             [ 80, 'モンスター図鑑(上級)(400G)'],
+             [ 81, '羽毛のケープ(400G)'],
+             [ 82, '鉄身のブーツ(400G)'],
+             [ 83, '看破の宝珠(中級)(420G)'],
+             [ 84, 'アドベンチャーザック(480G)'],
+             [ 85, '隠密の巻物(中級)(460G)'],
+             [ 86, 'ファルクス[破損](500G)'],
+             [ 87, 'ポーションホルダー(上級)(540G)'],
+             [ 88, 'カタール[破損](500G)'],
+             [ 89, '名匠の楽器(540G)'],
+             [ 90, '技能薬(上級)(480G)'],
+             [ 91, '火炎のタリスマン(620G)'],
+             [ 92, 'からすの手提げ袋(600G)'],
+             [ 93, '範囲隠密の巻物(中級)(570G)'],
+             [ 94, 'トラベラーハット(650G)'],
+             [ 95, '高位保護の巻物(中級)(520G)'],
+             [ 96, '長柄槍[破損](600G)'],
+             [ 97, '範囲飛行の巻物(中級)(570G)'],
+             [ 98, 'クイックカード(中級)(620G)'],
+             [ 99, 'トラベラーハット(650G)'],
+             [100, '剛手甲[破損](650G)'],
+             [101, '掃討の巻物(特級)(570G)'],
+             [102, 'エスパドン[破損](700G)'],
+             [103, '異次元バッグ(上級)(740G)'],
+             [104, 'まごころミートパイ(410G)x2'],
+             [105, '李満弓[破損](700G)'],
+             [106, '招命の宝珠(上級)(395G)x2'],
+             [107, '識別の巻物(上級)(690G)'],
+             [108, 'リネンクロス[破損](850G)'],
+             [109, '出前のおかもち(820G)'],
+             [110, '回復薬(上級)(720G)'],
+             [111, '蘇生の宝珠(上級)(790G)'],
+             [112, '再生薬(上級)(790G)'],
+             [113, 'エレメンタルカード(上級)(750G)'],
+             [114, '極上の楽器(900G)'],
+             [115, '踏破の巻物(上級)(750G)'],
+             [116, '垣間見の巻物(特級)(750G)'],
+             [117, '鉄壁符(上級)(820G)'],
+             [118, '魚鱗の宝珠(上級)(860G)'],
+             [119, '神奉刀[破損](850G)'],
+             [120, 'ヘヴィナックル[破損](950G)'],
+             [121, '巧指符(特級)(1000G)'],
+             [122, '幻翼の宝珠(上級)(930G)'],
+             [123, '火炎のベルト(1000G)'],
+             [124, 'シルキーベレー(1100G)'],
+             [125, 'ヒーリングワード(上級)(970G)'],
+             [126, '技能薬(特級)(930G)'],
+             [127, 'クレセントアックス[破損](1100G)'],
+             [128, 'ウォーハンマー(1100G)'],
+             [129, 'ミスルトゥのスタッフ[破損](1150G)'],
+             [130, '範囲保護の巻物(上級)(970G)'],
+             [131, '浄眼の巻物(上級)(970G)'],
+             [132, 'マジカルカード(上級)(1000G)'],
+             [133, '現身符(上級)(1000G)'],
+             [134, 'ジュエリーポーチ(1200G)'],
+             [135, '耐魔の軟膏(上級)(1000G)'],
+             [136, '鉄身の軟膏(上級)(1000G)'],
+             [137, '範囲飛行の巻物(上級)(1100G)'],
+             [138, '隠密の巻物(上級)(1100G)'],
+             [139, '魔性の楽器(1300G)'],
+             [140, '良業物[破損](1150G)'],
+             [141, '範囲付呪の巻物(上級)(1100G)'],
+             [142, 'ジャマダハル[破損](1300G)'],
+             [143, 'マスターパック(1400G)'],
+             [144, 'ドライハンダー[破損](1450G)'],
+             [145, 'グレイトヘルム[破損](1050G)'],
+             [146, '高位保護の巻物(上級)(1200G)'],
+             [147, '範囲隠密の巻物(上級)(1200G)'],
+             [148, 'ローズサークレット(1600G)'],
+             [149, '火炎のメダリオン(1600G)'],
+             [150, 'ファインジャベリン[破損](1200G)'],
+             [151, 'カッツバルゲル[破損](1250G)'],
+             [152, '紫の指輪[破損](1300G)'],
+             [153, '異次元バッグ(特急)(1600G)'],
+             [154, '五人張[破損](1750G)'],
+             [155, '皇国技芸寮打物[破損](1400G)'],
+             [156, 'ヘヴィランス[破損](1500G)'],
+             [157, '竜手甲[破損](1650G)'],
+             [158, 'ポロニアのワンド[破損](1700G)'],
+             [159, 'デスサイズ[破損](1700G)'],
+             [160, 'フランシスカ(1700G)'],
+             [161, '伝説の楽器(1700G)'],
+             [162, 'ウルミン(1900G)'],
             ]
 
     return tableName, table
@@ -687,14 +1335,14 @@ MESSAGETEXT
     return tableName, table
   end
 
-    def getGoblinTresureResultTable
+  def getGoblinTresureResultTable
     tableName = "ゴブリン財宝表"
     table = [
-             [ 7, 'ライトメイス'],
-             [ 8, '百科事典'],
+             [ 7, 'ライトメイス(40G)'],
+             [ 8, '百科事典(40G)'],
              [ 9, '42G'],
              [10, '古びた髑髏[コア素材](40G)'],
-             [11, '加速の巻物(初級)'],
+             [11, '加速の巻物(初級)(45G)'],
              [12, '46G'],
              [13, '48G'],
              [14, '50G'],
@@ -707,15 +1355,15 @@ MESSAGETEXT
              [21, '67G'],
              [22, '傷だらけの象牙像[換金](70G)'],
              [23, '72G'],
-             [24, 'ライトランス[破損]'],
+             [24, 'ライトランス[破損](75G)'],
              [25, '骨のネックレス[換金](80G)'],
              [26, '砂金混じりの石[換金](80G)'],
              [27, '85G'],
              [28, '香木の小片[換金](90G)'],
              [29, '奇妙なお面[換金](90G)'],
              [30, '96G'],
-             [31, 'シミター'],
-             [32, 'ウッデンラウンド'],
+             [31, 'シミター(100G)'],
+             [32, 'ウッデンラウンド(100G)'],
              [33, '風鳴りの鈴[コア素材](80G)&鉄の陣笠[コア素材](30G)'],
              [34, '112G'],
              [35, '山羊スライム(大)[換金](120G)'],
@@ -724,26 +1372,26 @@ MESSAGETEXT
              [38, '130G'],
              [39, '水晶のチェス駒[換金](135G)'],
              [40, '140G'],
-             [41, 'テント(キャンプ用)'],
-             [42, '怪しい丸薬[コア素材](30G)×2&100G'],
+             [41, 'テント(キャンプ用)(150G)'],
+             [42, '怪しい丸薬[コア素材](30G)x2&100G'],
              [43, '謎めいた毛皮[コア素材](80G)&怨念の鍔[コア素材](80G)'],
              [44, '儀式の骨剣[換金](160G)'],
              [45, '古ぼけたコイン[換金](165G)'],
-             [46, '狼牙棒'],
+             [46, '狼牙棒(170G)'],
              [47, '法理回路[コア素材](60G)&120G'],
              [48, '大きな魔力結晶[コア素材](120G)&60G'],
              [49, '189G'],
              [50, '上等な樽酒[換金](200G)'],
              [51, '201G'],
-             [52, '白の指輪'],
+             [52, '白の指輪(210G)'],
              [53, 'ヒスイの首飾り[換金](210G)'],
              [54, '金のゴブリン像[換金](220G)'],
-             [55, 'サーベル'],
+             [55, 'サーベル(230G)'],
              [56, '真鉄の刀身[コア素材](120G)&120G'],
              [57, '汚れた青磁の壺[換金](240G)'],
-             [58, 'とんすとんの焼き串盛り合わせ×2'],
+             [58, 'とんすとんの焼き串盛り合わせ(120G)x2'],
              [59, 'ゴブリン王の勲章[換金](255G)'],
-             [60, '垣間見の巻物(中級)×2'],
+             [60, '垣間見の巻物(中級)(130G)x2'],
              [61, '269G'],
              [62, '白狼の毛皮[換金](280G)'],
             ]
@@ -756,54 +1404,54 @@ MESSAGETEXT
 
     return nil unless("PTAG" === command)
 
-	tableName = "パーソナリティタグ表"
-	table = [
-	         '[おませさん]',
-	         '[好奇心旺盛]',
-	         '[寂しがりや]',
-	         '[生真面目]',
-	         '[食いしん坊]',
-	         '[やんちゃ]または[おてんば]',
+    tableName = "パーソナリティタグ表"
+    table = [
+             '[おませさん]',
+             '[好奇心旺盛]',
+             '[寂しがりや]',
+             '[生真面目]',
+             '[食いしん坊]',
+             '[やんちゃ]または[おてんば]',
 
-	         '[お人よし]',
-	         '[情熱家]',
-	         '[世話好き]',
-	         '[理知的]',
-	         '[頑固者]',
-	         '[兄貴肌]または[姉御肌]',
+             '[お人よし]',
+             '[情熱家]',
+             '[世話好き]',
+             '[理知的]',
+             '[頑固者]',
+             '[兄貴肌]または[姉御肌]',
 
-	         '[義理堅い]',
-	         '[気まぐれ]',
-	         '[職人気質]',
-	         '[熱血漢]',
-	         '[努力家]',
-	         '[男好き]または[女好き]',
+             '[義理堅い]',
+             '[気まぐれ]',
+             '[職人気質]',
+             '[熱血漢]',
+             '[努力家]',
+             '[男好き]または[女好き]',
 
-	         '[家庭的]',
-	         '[負けず嫌い]',
-	         '[純真]',
-	         '[朴念仁]',
-	         '[慈悲深い]',
-	         '[マイペース]',
+             '[家庭的]',
+             '[負けず嫌い]',
+             '[純真]',
+             '[朴念仁]',
+             '[慈悲深い]',
+             '[マイペース]',
 
-	         '[楽天家]',
-	         '[仲間思い]',
-	         '[誇り高い]',
-	         '[社交的]',
-	         '[冷静沈着]',
-	         '[ロマンチスト]',
+             '[楽天家]',
+             '[仲間思い]',
+             '[誇り高い]',
+             '[社交的]',
+             '[冷静沈着]',
+             '[ロマンチスト]',
 
-	         '[学者肌]',
-	         '[内向的]',
-	         '[苦労人]',
-	         '[派手好き]',
-	         '[勇猛果敢]',
-	         '[ミステリアス]',
-	        ]
+             '[学者肌]',
+             '[内向的]',
+             '[苦労人]',
+             '[派手好き]',
+             '[勇猛果敢]',
+             '[ミステリアス]',
+            ]
 
-	result, number = get_table_by_d66(table)
+    result, number = get_table_by_d66(table)
 
-	text = "#{tableName}(#{number})：#{result}"
+    text = "#{tableName}(#{number})：#{result}"
     return text
   end
 
@@ -814,7 +1462,7 @@ MESSAGETEXT
 
     tableName = "交友表"
     table = [
-	         "交友対象への庇護\nあなたは、交友対象を庇護したいと思っている。守ってあげられる強さがあなたにはある。",
+             "交友対象への庇護\nあなたは、交友対象を庇護したいと思っている。守ってあげられる強さがあなたにはある。",
              "交友対象への親愛\nあなたは、交友対象を好ましく思っている。その気持ちを彼に伝えているとは限らない。",
              "交友対象との義兄弟\nあなたは、交友対象と兄弟同然の間柄だ。もちろん、血は繋がっていないが、そんなことは些細なことだ。",
              "交友対象の英雄\nあなたは、交友対象から英雄視されている。それは例えあなたが否定しても変わらない。",
@@ -1032,48 +1680,48 @@ MESSAGETEXT
 
     tableName = "攻撃命中箇所"
     table = [
-	         '額',
-	         '頬',
-	         '鼻',
-	         '顎',
-	         '後頭部',
-	         '首',
+             '額',
+             '頬',
+             '鼻',
+             '顎',
+             '後頭部',
+             '首',
 
-	         '耳',
-	         '目',
-	         'こめかみ',
-	         '腕',
-	         '肘',
-	         '手',
+             '耳',
+             '目',
+             'こめかみ',
+             '腕',
+             '肘',
+             '手',
 
-	         '手の指',
-	         '心臓',
-	         '胃',
-	         '肺',
-	         '肋骨',
-	         '肩',
+             '手の指',
+             '心臓',
+             '胃',
+             '肺',
+             '肋骨',
+             '肩',
 
-	         '背',
-	         'わき腹',
-	         '腰',
-	         '下腹',
-	         '太もも',
-	         '喉',
+             '背',
+             'わき腹',
+             '腰',
+             '下腹',
+             '太もも',
+             '喉',
 
-	         'ふくらはぎ',
-	         'アキレス腱',
-	         'かかと',
-	         'すね',
-	         '足の小指',
-	         '膝',
+             'ふくらはぎ',
+             'アキレス腱',
+             'かかと',
+             'すね',
+             '足の小指',
+             '膝',
 
-	         '社会的信用',
-	         '人間関係',
-	         '初恋の思い出',
-	         '完璧なはずの予測',
-	         '心',
-	         '眼鏡',
-	        ]
+             '社会的信用',
+             '人間関係',
+             '初恋の思い出',
+             '完璧なはずの予測',
+             '心',
+             '眼鏡',
+            ]
     result, number = get_table_by_d66(table)
     "#{tableName}(#{number})：#{result}"
   end
@@ -1085,47 +1733,47 @@ MESSAGETEXT
 
     tableName = "PC名"
     table = [
-	         'Kuraudo',
-	         'Sefirosu',
-	         'Kirito',
-	         'Asuna',
-	         'Leeroy Jenkins',
-	         'Buront',
+             'Kuraudo',
+             'Sefirosu',
+             'Kirito',
+             'Asuna',
+             'Leeroy Jenkins',
+             'Buront',
 
-	         '水瀬陽夢',
-	         '三宝寺吾郎',
-	         '奈流麗夢',
-	         'フランチャイズ竜崎',
-	         '太宰治',
-	         'ねざ',
+             '水瀬陽夢',
+             '三宝寺吾郎',
+             '奈流麗夢',
+             'フランチャイズ竜崎',
+             '太宰治',
+             'ねざ',
 
-	         'クロウ・リー',
-	         'ダィテス',
-	         '達也',
-	         '深雪',
-	         'スレイ',
-	         'タカぼんさん',
+             'クロウ・リー',
+             'ダィテス',
+             '達也',
+             '深雪',
+             'スレイ',
+             'タカぼんさん',
 
-	         '黒の錬金術士',
-	         '†愛天使猫姫†',
-	         'デス★ガン',
-	         '卍漆黒の堕天使卍',
-	         '光速の異名を持ち重力を自在に操る高貴なる騎士',
-	         '新世界†英傑殺し',
+             '黒の錬金術士',
+             '†愛天使猫姫†',
+             'デス★ガン',
+             '卍漆黒の堕天使卍',
+             '光速の異名を持ち重力を自在に操る高貴なる騎士',
+             '新世界†英傑殺し',
 
-	         'ろぐほら',
-	         'ああああ',
-	         'そうこ',
-	         'ぎんこう',
-	         'あずかり',
-	         'もょもと',
+             'ろぐほら',
+             'ああああ',
+             'そうこ',
+             'ぎんこう',
+             'あずかり',
+             'もょもと',
 
-	         'サスケ',
-	         'ラファエロ',
-	         'ドナテロ',
-	         'ミケランジェロ',
-	         '川内',
-	         'フジキド',
+             'サスケ',
+             'ラファエロ',
+             'ドナテロ',
+             'ミケランジェロ',
+             '川内',
+             'フジキド',
             ]
 
     result, number = get_table_by_d66(table)
@@ -1337,78 +1985,1598 @@ MESSAGETEXT
     return "#{tableName}" + (is_roll ? "(#{type})" : '') + "：#{type_name}(#{dice})：#{result}"
   end
 
+  #イースタル探索表
   def getEastalDiceCommandResult(command)
+    return nil unless /ESTL(\d*)([\+\-\d]*)(\$(\d+))?/ =~ command
 
-    return nil unless command =~ /ESTL(\d+)/
+    #TKfix メソッドをまたぐと$xの中身がnilになっている
+    reg1 = $1
+    reg2 = $2
+    reg3 = $3
+    reg4 = $4
 
-    cr = $1.to_i
-    return nil unless (cr >= 1 && cr <= 10)
+    return command if (reg1.empty? && reg2.empty? && reg3.nil?)
 
+    rank = reg1.to_i
+    is_choice = (reg1.empty? || (not reg3.nil?))
+    modifyText = reg2
+    modify = getValue(modifyText, 0)
+    is_fix_roll = (not reg3.nil?)
+    dice_value = reg4
+    is_rank_enable = ( (not is_choice) || is_fix_roll)
+
+    tableName, table = getEastalExplorationResultTable
+
+    number, dice_str, =
+      if is_choice
+        [dice_value.to_i, dice_value]
+      else
+        roll(2, 6)
+      end
+
+    number += (rank * (is_rank_enable ? 5 : 0)) + modify
+
+    number = [getAdjustNumberMin(number, table), 162].min
+    result = get_table_by_number(number, table)
+    #result.chomp!
+    result = result.chomp # TKfix !
+
+    return "#{tableName}(#{number}#{dice_str ? '[' + dice_str + ']' : ''})\n#{result}"
+  end
+
+  def getEastalExplorationResultTable
     tableName = "イースタル探索表"
+    table = [
+             [  7, <<EOS
+香りに誘われて：
+　小さな街道沿いで、商隊が〈走り茸〉にまとわりつかれている。
+　奴らを追い払うと、商隊のリーダーがお礼にと人数分の「干した〈走り松茸〉[換金](40G)」をくれる。
+　追われていた理由はこれだったらしい。
+EOS
+             ],
+             [  8, <<EOS
+不幸な出会い：
+　街道を歩いていると、横合いの茂みから〈小牙竜鬼〉の一団が飛び出してきた。
+　突然のことでとても驚いたけど、向こうも〈冒険者〉がいるとは思っていなかったらしい。
+　一瞬の気まずい沈黙を経て、なし崩し的に乱戦が始まった。
+　PCは全員[疲労:10]を受ける。
+EOS
+             ],
+             [  9, <<EOS
+釣果はいかが？：
+　〈大地人〉の村近くの川で釣りが盛んなようだ。
+　挑戦するならば長靴の中に(2D*5)Gを見つけることができる。
+　この釣り場、ハズレなんじゃない？
+EOS
+             ],
+             [ 10, <<EOS
+近道だと思ったんだけど：
+　「この雑木林を抜けた方が早いよ」なんて誰が言い出したのやら。
+　近道に踏み込んだら、見事に道に迷ってしまった。
+　半日を余計に消費してしまう。
+　なんか疲れた……。
+　PCは全員【因果力】1点を失う。
+EOS
+             ],
+             [ 11, <<EOS
+おすすめの名産品：
+　キミたちは〈大地人〉の村に出くわした。
+　〈円卓会議〉と契約している農村で、〈冒険者〉にも好意的に接してくれる。
+　素朴だけど手厚いもてなしが心にしみた。
+　村の食堂でなにか食べるのなら10Gを支払い【因果力】に+1する。
+EOS
+             ],
+             [ 12, <<EOS
+思わぬ報酬：
+　畑を開墾していた〈大地人〉農夫が、切り株を掘り起こせずに途方に暮れていた。
+　手助けするならPCは全員[疲労:10]を受ける。
+　みんなで力を合わせて切り株を掘り起こすと、そこにはなんと「鉄の陣笠[コア素材](30G)」が埋まっていた！
+　キミたちは「鉄の陣笠」とお礼にもらった「お好みサンドイッチ(『LHZB1』P218)」2個を受け取る。
+EOS
+             ],
+             [ 13, <<EOS
+きりきり舞い：
+　野原の向こう側でつむじ風が舞っている、と思ったらどんどん速度を上げてこっちに向かってきた！
+　これはもう小さな竜巻だ。
+　PCは全員「難易度:9」の[耐久判定]を行なう。
+　失敗したPCは、風に耐えきれずにアイテムを吹き飛ばされる。
+　[消耗品]アイテムを1個失う。
+EOS
+             ],
+             [ 14, <<EOS
+小さな祝福：
+　雑木林の中を進んでいくとキラキラする光が集まってくる。
+　PCのなかに〈森呪遣い〉もしくは〈召喚術師〉がいるのならば小さな植物の妖精たちが手を振っているのがわかる。
+　自分たちに気づいたことに喜んだ妖精たちは〈森呪遣い〉と〈召喚術師〉にささやかな祝福をしてくれた。
+　【因果力】1点を得る。
+EOS
+             ],
+             [ 15, <<EOS
+だいなしピクニック：
+　林の中の広場でお弁当を広げて小休止。
+　モンスターの気配もなく、ちょっとしたピクニック気分……と思いきや、にわかに空がかき曇り、雷鳴とどろく大雨が降りだした！
+　キミたちはあわてて木陰に駆けこむ。
+　PCは全員1Dを振ること。
+　3以下のPCは一番高価な[消耗品]アイテム1つを失う。
+EOS
+             ],
+             [ 16, <<EOS
+キャラバンクルーズ：
+　林の木陰で休憩している〈第８商店街〉の商隊と出会った。
+　ギルドマスターのカラシン(『LHZB1』P351)みたいだ。
+　PCのうち希望者は気さくな彼とコネクションを取得することができる。
+　関係は「友情」または「雇用関係」となる(PCが選択)。
+　このコネクションはアフタープレイで消滅するがログチケット:アザーゲット1枚を消費すれば維持してもよい。
+EOS
+             ],
+             [ 17, <<EOS
+花畑にご用心：
+　花畑を渡る街道で〈暗殺蜜蜂〉の群れに出くわした。
+　追い払いながら縄張りから撤退する。
+　PCは全員[疲労:10]を受けるが、彼らが落とした「黒いローヤルゼリー[換金](70G)」1つを得る。
+EOS
+             ],
+             [ 18, <<EOS
+おとりはまかせろ！：
+　キミたちは〈緑小鬼〉に襲われていた〈大地人〉の荷馬車を逃がすため、おとりと護衛に分かれて行動を開始した。
+　おとりをかって出るPCを1人、相談して決めること。
+　そのPCは[疲労:20]を受け【因果力】1点を失うが、他のPCたちと荷馬車は安全に撤退することができる。
+EOS
+             ],
+             [ 19, <<EOS
+住処にお宝？：
+　〈フクロウ熊〉を撃退したキミたちは全員[疲労:10]を受ける。
+　もしかしたら近くに巣があるかもしれない。
+　代表者1名は「難易度:11」の[知覚判定]を行なう。
+　この判定は[偵察]タグがついているように扱う。
+　〔判定成功〕
+　　「壊れた武具[換金](90G)」を見つけた。
+　〔判定失敗〕
+　　何も得られない。
+　〔サブ職:狩人、辺境巡視〕
+　　判定に+2。
+EOS
+             ],
+             [ 20, <<EOS
+触らぬ狼に祟りなし：
+　遠くに狼の群れを発見した。
+　まだこちらには気づいてないようだが、このまま進めば鉢合わせすることになるだろう。
+　避けるのならPCの代表者1名は「難易度:10」の[解析判定]を行なうこと。
+　〔判定成功〕
+　　遠回り成功。
+　　PCは全員[疲労:10]を受ける。
+　〔判定失敗〕
+　　気づかれてしまった！
+　　撃退したものの、PCは全員[疲労:25]を受ける。
+　〔サブ職:斥候、遊牧民〕
+　　判定に+2。
+EOS
+             ],
+             [ 21, <<EOS
+ベリー摘み競争：
+　雑木林でベリーの茂みを見つけたキミたちは、誰が一番多く摘めるか競争を始めた。
+　PCは全員2Dを振ること。
+　一番低い出目のPCは、一番出目が高いPCにコネクションを取得する。
+　関係は「尊敬」か「ライバル」となる(PCが選択)。
+　このコネクションはアフタープレイで消滅するが、ログチケット:アザーゲット1枚を消費すれば維持してもよい。
+EOS
+             ],
+             [ 22, <<EOS
+村の形見：
+　滅びた〈大地人〉の村が眼前に広がる。
+　せめてもの弔いに何かを持ち帰って役立てよう。
+　PCは【因果力】を任意の数だけ消費してもよい。
+　各自消費した【因果力】と同数のダイスを振り、全員の出目を合計すること。
+　3以上なら「赤熱の小爪[コア素材](50G)」1つを得る。
+　8以上なら加えてさらに「小さな宝石箱[換金](70G)」1つを得る。
+　13以上なら加えてさらに「ミニポーチ(『LHZB1』P217)」1つを得る。
+EOS
+             ],
+             [ 23, <<EOS
+ぬかるみの強行軍：
+　折からの雨でひどく街道がぬかるんでいて、〈冒険者〉でも歩くには少々骨が折れる。
+　PCは全員「難易度:10」の[耐久判定]を行なう。
+　〔判定成功〕
+　　黙々と足を進め、無事目的地に到着した。
+　〔判定失敗〕
+　　疲れに加え少し足首をひねってしまったようだ。
+　　[疲労:20]を受ける。
+　〔サブ職:辺境巡視、配達屋〕
+　　判定に+2。
+EOS
+             ],
+             [ 24, <<EOS
+ハニーハント：
+　雑木林の中で、巨大蜂の巣を見つけた。
+　たっぷりのハチミツを手に入れるチャンスだ！
+　挑戦するならば、PCの代表者1名は「難易度:10」の[操作判定]を行なう。
+　〔判定成功〕
+　　PCたちは「黄金色のハチミツ[換金](120G)」を得る。
+　　-2(累積する、最大-10)を受ければもう一度チャレンジできる。
+　〔判定失敗〕
+　　怒ったハチの群れに追いかけられPC全員は[疲労:10]を受ける。
+　〔サブ職:採取人、農家〕
+　　判定に+2。
+EOS
+             ],
+             [ 25, <<EOS
+闇の群雲：
+　キミたちは〈闇精霊〉の「夜渡り」に遭遇する。
+　[暗視]状態、もしくは[魔法]プロップの影響を受けないPCは被害を受けない。
+　それ以外のPCは全員「難易度:10」の[抵抗判定]を行なう。
+　いまが夜なら(GM判断)「難易度:12」になる。
+　〔判定失敗〕
+　　【因果力】1点を失う。
+　〔サブ職:刻印呪師、魔法学者〕
+　　この判定に+2。
+EOS
+             ],
+             [ 26, <<EOS
+風のいたずら：
+　金色に波打つ麦畑で、一際強い突風が吹いた。
+　PCは全員2Dを振ること。
+　一番低い出目のPCは装備や性別に関係なく、シナリオ終了まで[ぱんちら]タグが与えられる。
+　このタグを持っているPCが何らかの判定でファンブルした時、特別にタグを[ぱんもろ]に変化させることでその判定を振り直してもよい。
+EOS
+             ],
+             [ 27, <<EOS
+古代の獣：
+　荒野を探索中、なかば地面にうずもれた巨大生物の骨格を見つけた。
+　キミたちはなにか素材が取れないか採取を試みる。
+　1Dを振ること(【因果力】1点で振り直してもよい)。
+　5か6が出たPCは「とがった爪[コア素材](60G)」もしくは「巨大な肋骨[換金](80G)」のどちらかを選んで得る。
+EOS
+             ],
+             [ 28, <<EOS
+死者の窪地：
+　丘陵地帯の窪地は不浄な死者がむらがっていた。
+　風に揺れるのはすべて〈蠢く墓標〉だったのだ！
+　おびただしい数の〈動く骸骨〉が迫る中での撤退戦の末、キミたち全員は[疲労:20]を受ける。
+EOS
+             ],
+             [ 29, <<EOS
+足跡は語る：
+　丘陵を行くキミたちはいくつもの足跡を発見した。
+　PCの代表者1名は[知識判定]を行なう。
+　〔判定成功〕
+　　足跡から〈醜豚鬼〉の小部隊の存在を看破したキミたちは奇襲を仕掛けることができる。
+　　仕掛ければ成功して120Gを入手する。
+　〔判定失敗もしくは奇襲しない〕
+　　足跡は意味不明で数が多すぎる。
+　　キミたちそっとその場を離れた。
+　〔サブ職:学者、追跡者〕
+　　判定に+2。
+EOS
+             ],
+             [ 30, <<EOS
+キリングゾーン：
+　キミたちは悪辣な〈小牙竜鬼〉たちが仕掛けたトラップ地帯に入り込んでしまった。
+　罠を解除しつつ抜け出すしかないようだ。
+　PCの代表者1名は「難易度:10」の[解除判定]を行なう。
+　〔判定成功〕
+　　最低限の被害で脱出できた。
+　　PC全員は[疲労:10]を受ける。
+　〔判定失敗〕
+　　いくつかのトラップにはまりPC全員は[疲労:30]を受ける。
+　〔達成値:14〕
+　　罠を突破して逆襲してやった！
+　　100Gを得る。
+　〔サブ職:機工師、罠師〕
+　　判定に+2。
+EOS
+             ],
+             [ 31, <<EOS
+果物天国：
+　通りかかった果樹園で〈大地人〉の農夫に果物の収穫を手伝ってほしいと頼まれた。
+　手伝うのならばPCは全員「難易度:10」の[解除判定]を行なう。
+　それぞれ難易度をいくつ上回ったかを記録し、その合計値が10以上ならば収穫は大成功だ。
+　喜んだ農夫は「かご一杯の果物(シェフの気まぐれピザ相当。『LHZB1』P218)」x2個をキミたちに持たせてくれる。
+EOS
+             ],
+             [ 32, <<EOS
+岩陰からこんにちわ：
+　岩陰から不意打ちを仕掛けてきた〈蜷局竜〉をなんとか撃退。
+　キミたちは戦利品として「風鳴りの鈴[コア素材](80G)」を手に入れたが、少なくない被害を受けた。
+　PC全員は1～6の数字2つを宣言し、代表者1名は1Dを振ること。
+　出目と同じ数字を宣言したPCは[疲労:25]を受ける。
+EOS
+             ],
+             [ 33, <<EOS
+野火から逃げろ！：
+　前方から焦げ臭い風が吹いてきた。
+　野火だ！
+　逃げようにも丘の合間を抜ける風が、炎を複雑に押し広げているようだ。
+　PCの代表者1名は「難易度:10」の[解析判定]を行なう。
+　〔判定成功〕
+　　風と地形を読み切り、PCは損害を免れた。
+　〔判定失敗〕
+　　何とか炎は免れたものの、思い切り煙を吸い込んでしまった。
+　　PCは全員[疲労:45]を受ける。
+　〔サブ職:生還者、星詠み〕
+　　判定に+2。
+EOS
+             ],
+             [ 34, <<EOS
+穴があったら探検だ：
+　丘の中腹に小規模なダンジョンを発見した。
+　探索する場合は1日を消費してPC全員は1Dを振る。
+　1人でも6が出た場合、戦利品として「小さな宝箱[換金](60G)」を3個手に入れた。
+　ホクホクだ。
+　そうでない場合、隅々まで探索したが全員が[疲労:20]を受けたのみだった。
+　誰かが探ったあとみたい。
+　〔サブ職:蒐集家、設計士〕
+　　2Dを振ってもよい。
+EOS
+             ],
+             [ 35, <<EOS
+一寸先は穴：
+　一面の田園地帯が広がっている。
+　景色に気をとられていたキミたちは、足元にポッカリ開いた堆肥の穴に気づかない！
+　PCは全員1Dを振ること。
+　出目が1～2ならば見事に穴にはまり込んでしまった！
+　しばらく臭いが取れそうにない。
+　被害者はコネクション1つ(被害者が選択)をして、シナリオ終了時まで関係を「くさい……」に変更する。
+EOS
+             ],
+             [ 36, <<EOS
+木漏れ日と水辺：
+　森の中できれいな泉を見つけた。
+　希望者はここで水浴びをすることも可能だ。
+　旅の汚れを落として気分もリフレッシュ、爽快な気分で冒険を続けられそうだ。
+　探索表のイベントで何らかの判定を行なう場合、その判定に+1Dする。
+　この効果はPCたち全員で共有し一度効果を受けるかシーン終了時まで持続する。
+EOS
+             ],
+             [ 37, <<EOS
+黒く光る宝物：
+　果樹園の木の上に巨大な鳥の巣を発見した。
+　巣の主が留守のうちに内部を調べてみると、散乱した羽毛や獣の骨の陰から「幻想の黒真珠[コア素材](100G)」を発見する。
+　いいのかな？
+　持っていくなら入手してもよい。
+EOS
+             ],
+             [ 38, <<EOS
+猪突猛進：
+　起伏の激しい丘陵を進んでいると、前方から〈猛猪〉の群れが激しい地響きを伴って迫ってくる！
+　PC全員は「難易度:11」の[運動判定]を行なう。
+　一人でも失敗すれば全員が判定失敗として扱おう。
+　〔判定成功〕
+　　キミたちは一目散に逃げきった。
+　〔判定失敗〕
+　　逃げた先には、今度ははぐれ〈巨石兵士〉だ！？
+　　延々逃げ回ってPC全員はそれぞれ[疲労:1D*10]を受ける。
+　〔サブ職:軽業師、傭兵〕
+　　全員の判定に+1。
+EOS
+             ],
+             [ 39, <<EOS
+へんてこ葡萄は蜜の味？：
+　日当たりのよい斜面に野性のブドウがたわわに実っている。
+　採取するPC全員は[知識判定]を行なう。
+　〔判定成功〕
+　　「貴腐ブドウ[換金](110G)」を手に入れる。
+　〔クリティカル〕
+　　さらに「常眠りの種子[コア素材](100G)」を得る。
+　〔サブ職:農家、採取人〕
+　　判定に+1D。
+EOS
+             ],
+             [ 40, <<EOS
+ここはどこ？：
+　灌木の生い茂る丘陵地帯、視界は悪いし歩きにくいことこの上ない。
+　どっちに向かっているかすらおぼつかない。
+　[飛行]状態のPCが1人もいない場合、回り道を強いられPC全員は【因果力】1点失うか、1日の時間をロスする。
+EOS
+             ],
+             [ 41, <<EOS
+狩りの邪魔：
+　澄んだ湧水を汲もうとしたキミたちに驚いて、立派な牡鹿が森の中に逃げていった。
+　すると近くの草むらから〈大地人〉の狩人がカンカンに怒って飛び出してきた。
+　どうやら彼の獲物を追い払ってしまったらしい。
+　PCは全員2Dを振ること。
+　一番低い出目のPCはこのNPCとコネクションを取得する。
+　関係は「恐縮」となる。
+　このコネクションはアフタープレイで消滅するがログチケット:アザーゲット1枚を消費すれば維持してもよい。
+EOS
+             ],
+             [ 42, <<EOS
+激闘の報酬：
+　キミたちは廃墟に住み着いた〈暗蛇竜〉を退治した。
+　遺跡が崩落するほど激しい戦いだったが、首尾は上々だ。
+　PCは【因果力】を任意の数だけ消費してもよい。
+　各自消費した【因果力】と同数のダイスを振り、全員の出目を合計すること。
+　ダイスの合計値が3以上ならPCたちは「ねじくれた角[コア素材](120G)」1つを得る。
+　8以上でさらに300Gを得る。
+　13以上でさらに「暗竜牙【魔触媒7】(60G)」5つを得る。
+EOS
+             ],
+             [ 43, <<EOS
+竜巻に追われて：
+　少しずつ風の強くなる中、キミたちが荒廃した高速道路を旅していると、背後から竜巻が追ってきていることに気付く。
+　勝てる相手じゃないから逃げださなければ！
+　PCは全員「難易度:12」の[運動判定]を行なう。
+　PCが[飛行]状態の場合は+2を得る。
+　〔判定成功〕
+　　PCは[疲労:20]を受ける。
+　　なんとか逃げ切れたようだ。
+　〔判定失敗〕
+　　PCは[疲労:50]を受け、【因果力】1点も失う。
+　〔サブ職:軽業師、生還者〕
+　　判定に+2。
+EOS
+             ],
+             [ 44, <<EOS
+毒を浄化せよ：
+　キミたちは〈緑小鬼〉の集団が泉に毒を投げ込んでいるのを目撃する。
+　村を滅ぼそうとする陰謀か！？
+　戦うのならば[疲労:20]を受ける。
+　BSを解除できるアイテムか特技があるのならば、泉を清浄化しておこう……。
+EOS
+             ],
+             [ 45, <<EOS
+させるか！：
+　道中、襲いかかってきた〈時計仕掛の蠍〉をキミたちは排除したはずだった。
+　とどめを刺したはずの一体が一瞬だけ息を吹き返し橙色の光線を放つ！
+　PCの代表者1名は[疲労:60]を受けなくてはならない。
+　その際に受ける[疲労]の強度を[軽減(光輝)]の強度ぶん減らしてもよい。
+EOS
+             ],
+             [ 46, <<EOS
+ない、ない、あれがない！：
+　森を歩いているとふと荷物整理を思い出した。
+　PC全員は2Dを振ること。
+　一番低い出目のPCは忘れ物に気づいた。
+　何を忘れたかはキミが決定してよい。
+　慌てたキミは仲間に相談する。
+　キミが望むなら、仲間にコネクションを得てもよい。
+　関係は「相棒」か「恩人」となる。
+　このコネクションはアフタープレイで消滅するがログチケット:アザーゲット1枚を消費すれば維持してもよい。
+EOS
+             ],
+             [ 47, <<EOS
+川底に眠る：
+　キミたちが川辺を歩いていると、川底で何かがキラリと光った。
+　なにかお宝があるかもしれない。
+　PCはここで半日過ごし[疲労:20]を受けるたびに1人1Dを振っても良い。
+　〔1D:1～2〕
+　　「水晶の砂[換金](20G)」を入手する。
+　〔1D:3～5〕
+　　「水晶の塊[換金](40G)」を入手する。
+　〔1D:6〕
+　　「蒼い鉱石[コア素材](140G)」を入手する。
+EOS
+             ],
+             [ 48, <<EOS
+沼より出ずるもの：
+　のどかな沼沢地帯を歩いてたキミたちは、沼からはい出てきた〈破裂の粘体〉に囲まれてしまっていた。
+　この包囲網から逃れるには協力が必要だ。
+　全員1Dを振る(【因果力】1点消費でさらに+1Dしてもよい)。
+　全員のダイスの合計が13以上なら、無事突破できる。
+　12以下の場合PCは全員[疲労:40]を受ける。
+EOS
+             ],
+             [ 49, <<EOS
+山の恵み：
+　豊かな沢の周りで山の幸を見つけよう！
+　PCは全員「難易度:12」の[知識判定]を行なう。
+　〔判定成功〕
+　　「大きな川魚(トポト親方のコロッケ相当。『LHZB1』P218)」を得る。
+　〔判定失敗〕
+　　「食べられるかもしれない水草[換金](1G)」2個を得る。
+　〔サブ職:料理人、食闘士〕
+　　判定に+2。
+EOS
+             ],
+             [ 50, <<EOS
+幻の先に：
+　のどかな沼沢地を進んでいると、突如毒の沼にはまり込んでしまった。
+　〈幻霊〉の幻覚による誘い込みの罠だったらしい！
+　強度10以上の[軽減(邪毒もしくは精神)]を持たないPCは[疲労:50]を受ける。
+EOS
+             ],
+             [ 51, <<EOS
+オークと一緒：
+　キミたちが崩れかけた高速道路を歩いていると、向こうから両手に骨付き肉を持った〈醜豚鬼〉の一団が踊りながら迫ってきた！
+　フレッシュミートゥ！
+　PCは全員2Dを振ること。
+　一番低い出目のPCの足元が崩れて〈醜豚鬼〉もろとも転落、そのPCには[尊い犠牲]タグが与えられる。
+　このタグはシナリオ終了まで解除されない。
+EOS
+             ],
+             [ 52, <<EOS
+物言わぬ番人：
+　森を横切る小川のほとりに、苔むした瑠璃色の機械人形が壊れてうずくまっている。
+　この機械人形の名前はなんだろう？
+　「難易度:14」の[知識判定]に成功し、1人でもかすれた文字が読めるなら「七色の透明捻子[コア素材](180G)」1つを得る。
+EOS
+             ],
+             [ 53, <<EOS
+鉄砲水！：
+　谷間で突然の大雨に見舞われる。
+　キミたちは鉄砲水の予兆に気付くだろうか？
+　PCの代表者1名は「難易度:14」の[知覚判定]を行なう。
+　〔判定成功〕
+　　予兆を察知して早々に避難したキミたちの足元を濁流が駆け抜けてゆく。
+　　危ないところだった。
+　〔判定失敗〕
+　　地響きが聞こえてきたときはもう手遅れだった。
+　　PCは全員[疲労:50]を受ける。
+　〔サブ職:探検家、生還者〕
+　　判定に+2。
+EOS
+             ],
+             [ 54, <<EOS
+廃橋の戦い：
+　高速道路の廃墟を進んでいると、〈鋼尾翼竜〉の群れに襲われた！
+　キミたちは勝利を納め、265Gを得るが、不安定な足場のため[疲労:30]を受ける。
+　[飛行]状態や[天然]もしくは[地形]プロップへの耐性があるPCは疲労は受けない。
+EOS
+             ],
+             [ 55, <<EOS
+転落注意：
+　谷あいの道を歩いていると、突然足元が崩れ始めた！
+　[飛行]状態でない、最も【行動力】が低いPCは崩落に巻き込まれて斜面を滑り落ちてしまう。
+　ダメージは大したことなかったけど、ついてない……。
+　滑落したPCは【因果力】1点を失うこと。
+EOS
+             ],
+             [ 56, <<EOS
+今夜はごちそう？：
+　葦の茂る沼沢地で〈大地人〉の兄妹が罠を仕掛けている。
+　鵞鳥を夕飯にしたいらしい。
+　1日かけて猟を手伝う場合、GMは兄弟にかわって2Dを振ること(サブ職が狩人、罠師のPCがいれば+1Dだ)。
+　8以上ならば、狩りは成功。
+　兄妹の家で料理を食べてPCたちは[疲労]をすべて回復する。
+　7以下ならば腹ペコだ。
+　兄妹の家で「タイミング:レストタイム」の行動が可能。
+EOS
+             ],
+             [ 57, <<EOS
+獅子の魔獣：
+　キミたちは手負いの〈人面獅子〉を廃墟の住処まで追いつめ、ついに討ち取った。
+　この偉業でPCは全員【因果力】1点を得る。
+　またPCの代表者1名は「難易度:15」の[知覚判定]を行なう。
+　判定に成功したなら、〈人面獅子〉の住処から「金色の毛玉[コア素材](220G)」を得る。
+　〔サブ職:狩人、追跡者〕
+　　判定に+2。
+EOS
+             ],
+             [ 58, <<EOS
+森の奇襲：
+　キミたちが木漏れ日にきらめくせせらぎで旅の疲れを癒していたところに〈醜豚鬼の遊撃兵〉が奇襲をしかけてきた！
+　とっさに武器を取りなんとか〈醜豚鬼の遊撃兵〉を蹴散らしたが、損害は決して小さくはなかった。
+　PCは全員[疲労:40]を受ける。
+EOS
+             ],
+             [ 59, <<EOS
+お肉入手のチャンス：
+　キミたちは水飲み場にやってきた。
+　足跡からすればここには動物が来そうな予感。
+　食料ゲットのチャンスだ！
+　半日のあいだ隠れて狙撃をするのならば狩りが可能だ。
+　[射撃攻撃]可能な武器を持つ代表者1名は「難易度:15」の[命中判定]を行なう。
+　〔判定成功〕
+　　PCたちは大きなイノシシを倒した。
+　　「魅惑のボタン肉(定番カレー相当。『LHZB1』P218)3個を得る。
+　〔判定失敗〕
+　　何も得られない。
+　〔サブ職:狩人、砲撃士〕
+　　判定に+2。
+EOS
+             ],
+             [ 60, <<EOS
+迷惑な怪魚：
+　大河を渡っているキミたちの後ろから巨大なモンスターが襲い掛かる。
+　ボートは転覆してしまうし、泳いで逃げるしかない。
+　[水棲]タグを持たないPCは慣れない水中戦闘に巻き込まれて[疲労:60]を受ける。
+EOS
+             ],
+             [ 61, <<EOS
+曙光：
+　山の稜線から差し込む朝日が、草の葉に溜まった朝露に反射する。
+　小鳥たちはさえずり、冷え込んだ夜の空気が追い払われていくのがわかる。
+　今朝のごはん当番は誰だっけ？
+　立候補したPC1名(メニューを発表すること！)には【因果力】1点が与えられる。
+　さあ、今日も一日頑張ろう！
+EOS
+             ],
+             [ 62, <<EOS
+吸血の森：
+　〈吸血ヒル〉の群生地である深い森の中を探索するキミたち。
+　PCは全員「難易度:13」の[耐久判定]を行なう。
+　成功した人数がパーティーの3名以上いれば、ヒルを振り払いながらキミたちは探索をなしとげる。
+　「錆びた聖印[コア素材](240G)」と2D*60Gを得る。
+　そうでなければキミたちは探索もそこそこに退却し、全員[疲労:20]を受ける。
+EOS
+             ],
+             [ 63, <<EOS
+落雷注意：
+　キミたちの頭上で雷雲が不気味に湧き上がり、雷鳴もゴロゴロと鳴り響いている。
+　そして不幸なことに、さえぎる物のない高原において雷はキミたちに狙いを定めたようだ！
+　[高位保護]を持たないPCは[疲労:80]を受ける。
+　〔サブ職:探検家、辺境巡視〕
+　　[疲労:80]ではなく[疲労:60]で済む。
+EOS
+             ],
+             [ 64, <<EOS
+ギブアンドテイク：
+　街道沿いに残る廃墟でキャンプする〈冒険者〉グループと遭遇した。
+　どうやら彼らには回復職のメンバーが欠けているようだ。
+　〈回復職〉PCが治療を申し出ると、彼らは治療のお礼に「定番カレー(『LHZB1』P218)」1個を分けてくれる。
+　もし特技喪失を受けるならさらに1個入手してよい。
+EOS
+             ],
+             [ 65, <<EOS
+悪魔の一族1：
+　大きな切り株に腰かけて一休み。
+　降り注ぐ木漏れ日が気持ちいい……と思ってたらカバンがない！？
+　少し離れたところでクロマミ族の連中がカバンをひっくり返してる！
+　PCは全員《異常探知》をすること。
+　最も達成値の低いPCは[消耗品]アイテムを2個失う。
+　〔ファンブル〕
+　　さらに[圧迫:2]を受ける。
+　〔サブ職:交易商人、配達屋〕
+　　判定に+2。
+EOS
+             ],
+             [ 66, <<EOS
+旅は道連れ：
+　キミたちは森の入り口で立ち往生している。
+　ヤーマという行商人の馬車を発見した。
+　護衛がいないため森を通り抜けるのが心配らしい。
+　もし一日の護衛を申し出るならヤーマからは大変感謝され、謝礼として300G渡される。
+　また、PCが望むならばヤーマとのコネクションを取得してもよい。
+　関係は「取引」となる。
+　このコネクションはアフタープレイで消滅するがログチケット:アザーゲット1枚を消費すれば維持してもよい。
+EOS
+             ],
+             [ 67, <<EOS
+勝負好きの〈妖精巨人〉：
+　高原の環状列石のそばを通りかかったキミたちに、〈妖精巨人〉が力比べを挑んできた。
+　勝負を受けるなら、PCの代表者が「難易度:14」の[耐久判定]、[知覚判定]、[知識判定]を1回ずつ行なう(それぞれの判定は別のPCが行なってもよい)。
+　2回以上成功すれば勝負に勝ち、「神鉄の組紐[コア素材](300G)」を入手する。
+　3回成功できたならさらにPC全員が【因果力】1点を入手する。
+　3回とも失敗した場合はPC全員が[疲労:35]を受ける。
+EOS
+             ],
+             [ 68, <<EOS
+不気味に響く音：
+　薄暗い森で〈魔狂蝙蝠〉の襲撃を受けた。
+　撃退はしたが強力な超音波攻撃で頭がガンガンする……。
+　PCは全員「難易度:16」の[耐久判定]を行なう。
+　〔判定成功〕
+　　【因果力】1点を失う。
+　〔判定失敗〕
+　　気分が悪くなってきた。
+　　【因果力】2点を失う。
+　〔自身:高位保護〕
+　　【因果力】を失わなくてよい。
+EOS
+             ],
+             [ 69, <<EOS
+手負いの獣にご用心：
+　森の小道を往くキミたちに、木々をへし折る重い音と地響きが近づいてくる。
+　なんと、手負いの〈魔熊〉が半狂乱となって飛び込んできた！
+　キミたちはなんとか〈魔熊〉の奇襲を撃退し、「傷ついた毛皮[換金](350G)」を手に入れたが被害も大きい。
+　「戦士職」もしくは「武器攻撃職」のPCは[疲労:45]を得る。
+EOS
+             ],
+             [ 70, <<EOS
+闇の連絡通路：
+　キミたちは遺跡と化したパーキングエリアを繋ぐトンネルを前にしている。
+　ここを通れば随分と近道ができるが、トンネルは闇に閉ざされ中をうかがい知ることができない。
+　キミたちはこのトンネルに向かってもよいし、迂回してもよい。
+　トンネルを通るならば[暗視]を持たないPC全員は特技喪失を受ける。
+　迂回するならばPC全員は[疲労:20]を受けて、もう一度この表を振らなければならない。
+EOS
+             ],
+             [ 71, <<EOS
+星月夜にて：
+　高原を旅するキミたちは満天の星の下で野営を行なう。
+　1Dを振って一番高い出目のPCが野営当番だ。
+　任意のPCひとりを見張りの相棒に指名し、今回の旅について話し合ってみよう(実時間で5分程度)。
+　あなたと相棒はお互いにコネクションを取るか、「関係変化表(『LHZB2』P59)」で1回ロールしてもよい。
+EOS
+             ],
+             [ 72, <<EOS
+最後の御奉公：
+　多数のメイド型ゴーレムが眠るかつてのパーキングエリアを思わせる遺跡で、キミたちはかろうじて機能を保っていた一体を発見した。
+　彼女はキミたちにきしむ腕で「刻印の銀板[コア素材](340G)」を託すと、静かにその活動を停止した。
+EOS
+             ],
+             [ 73, <<EOS
+冷たい霧に抱かれて：
+　灌木の生い茂る斜面を登っていると、急に濃霧がたちこめてきた。
+　霧はキミたちの体温を急速に奪っていく。
+　強度20以上の[軽減(冷気)]を持たないPCは[疲労:40]を受ける。
+EOS
+             ],
+             [ 74, <<EOS
+鹿肉パーティーだ！：
+　早朝の高原で、〈剣角鹿〉の大きな群れを発見した。
+　半日かけて狩りを行なうならば「新鮮な鹿肉[換金]45G」を2D+5個入手できる。
+　〔サブ職:料理人〕
+　　入手した「新鮮な鹿肉」を1個消費するごとにPC1人の[疲労]を-20してもよい。
+EOS
+             ],
+             [ 75, <<EOS
+手痛い報復：
+　清々しい風が吹き抜ける林を歩いていると、うっかり〈歩行毒樹〉の根を踏んで反撃を食らってしまった。
+　大したダメージは受けなかったけど、傷口がひどく腫れてとにかく痛い！
+　PCは全員2Dを振ること。
+　一番低い出目のPCは【因果力】1点を失う。
+EOS
+             ],
+             [ 76, <<EOS
+かぐわしい？におい：
+　キミたちは森でとある食虫植物の群生地に飛び込んでしまった。
+　大したダメージこそ受けなかったが、虫を引き寄せるべとべとが装備にこびりついて離れない。
+　これは困った！
+　PCは全員2Dを振ること。
+　一番低い出目のPCは[べとべと]タグが与えられる。
+　このタグはシナリオ終了まで解除されない。
+EOS
+             ],
+             [ 77, <<EOS
+誰も知らない花園：
+　険しい斜面を登り切った先には、一面の花畑が広がっていた。
+　その鮮やかさにキミたちは見ほれてしまう。
+　PCは全員1Dを振ること。
+　〔1D:1～4〕
+　　「きれいな花[換金](5G)」を好きなだけ摘んでもよい。
+　〔1D:5～6〕
+　　「うごめく花弁[コア素材](380G)」を手に入れる。
+EOS
+             ],
+             [ 78, <<EOS
+夜明け前の攻防：
+　夜明け前の薄暗がりから〈怨霊〉の大群が襲いかかってきた。
+　夜明けまで持ちこたえねば！
+　「回復職」と「魔法攻撃職」のPCは激戦のあまり【因果力】2点を失う。
+EOS
+             ],
+             [ 79, <<EOS
+何が出るかな……？：
+　瓦礫に半ばうずもれた遺跡でアルヴのものと思われる機械を発見した。
+　コインを入れるとランダムに品物が出てくるらしい。
+　挑戦したいPCは100Gを消費して、1Dを振ること。
+　なお、この機械は合計4回挑戦したところで完全に壊れてしまう。
+　〔1D:1〕
+　　ハズレ！
+　　何も出てこないぞ。
+　〔1D:2〕
+　　「鉄身の軟膏(初級)(『LHZB1』P219)」を入手。
+　〔1D:3～4〕
+　　「回復薬(中級)(『LHZB2』P146)」を入手。
+　〔1D:5〕
+　　「幻翼の宝珠(中級)(『LHZB2』P159)」を入手。
+　〔1D:6〕
+　　「保護の巻物(中級)(『LHZB2』P156)」を入手。
+EOS
+             ],
+             [ 80, <<EOS
+茨の森：
+　キミたちは毒の棘を持つ灌木が生い茂る森に迷い込んだ。
+　茎や葉の棘が露出した肌を容赦なく引っ掻いていく。
+　[中鎧]を装備したPCは[疲労:30]を受ける。
+　[軽鎧]もしくは鎧を装備していないPCは[疲労:30]と特技喪失を受ける。
+EOS
+             ],
+             [ 81, <<EOS
+朝まで歌合戦：
+　浜辺を行くキミたちの耳にきれいな歌声が聞こえてきた。
+　歌声の主を探すと、なんと岩場で〈人魚〉と〈翼持つ歌姫〉が歌の勝負をしている！？
+　ギャラリーに気づいた彼女たちはさらに熱心に歌い始めた。
+　キミたちはそそくさと立ち去ってもいいし、彼女らが満足するまで歌を聴いていてもいい。
+　そうした場合(一応美少女？の)祝福で【因果力】1点を得るが、[疲労:20]も受けてしまう。
+EOS
+             ],
+             [ 82, <<EOS
+ロードランナー：
+　狭い山道の向こうから、すごい勢いで〈時計仕掛の駝鳥〉が駆けてくる！
+　PC代表者は「難易度:17」を目標に[回避判定]をしてもよい(そのまま受け止めてもよい)。
+　失敗した(受け止めた)場合3D*20点の物理ダメージを受けてしまうが、キミが[戦闘不能]にならなければ〈時計仕掛の駝鳥〉も砕けて「真鋼のワイヤー[コア素材](440G)」を得る。
+　回避に成功するか[戦闘不能]になった場合、〈時計仕掛の駝鳥〉は走り去る。
+EOS
+             ],
+             [ 83, <<EOS
+森を舐める猛火：
+　山道を歩いていると、風に乗って焦げ臭いにおいが漂ってきた。
+　風上を見れば真っ赤な炎の舌が斜面を駆け上ってくるではないか！
+　[高位保護]を持たないPCは炎と煙に巻かれ[疲労:80]を受ける。
+EOS
+             ],
+             [ 84, <<EOS
+岩石林に咲くバラ：
+　キミたちは石の花が咲くという、石化ガスのたなびく石柱の林を訪れた。
+　探検するなら[高位保護]を持つPCにかぎり【因果力】を任意の数だけ消費してもよい。
+　各自消費した【因果力】と同数のダイスを振り、全員の出目を合計すること。
+　ダイスの合計値が3以上ならPCたちは「雲母のバラ[換金](480G)」1つを得る。
+　8以上でさらに「雲母のバラ[換金](480G)」3つを得る。
+　13以上でさらに「紅玉のバラ[換金](2000G)」1つを得る。
+EOS
+             ],
+             [ 85, <<EOS
+迫りくる海蛇竜：
+　浜辺で漁師たちが逃げまどっている。
+　〈海蛇竜〉が波打ち際で噛みつこうとしているのだ。
+　救助するならヘイトを集めてかばうしかない。
+　「戦士職」のPCは【因果力】2点を失う。
+EOS
+             ],
+             [ 86, <<EOS
+浜辺で小休止：
+　周囲の砂浜には椰子の木がたくさん生えている。
+　木陰に入って椰子の実ジュースを飲んでいると、なんだかちょっぴりリゾート気分だ。
+　十分に休憩したら、冒険を再開しよう！
+　この探索表の出来事で次に何らかの判定をする場合ロールに+1D。
+　この効果はひとりでも全員でも次の判定のみ、一度効果を受けるかシーン終了時まで持続する。
+EOS
+             ],
+             [ 87, <<EOS
+わんぱくイルカと〈冒険者〉：
+　目の前に広がる海で、何匹ものイルカが水面を蹴立てて遊んでいる。
+　キミたちに気づくと岸辺まで近づいてきた。
+　人懐っこい彼らはどうやらキミたちに遊んでほしいらしい。
+　もしも一緒に遊んでやるならPC全員は[疲労:40]を受けるが、遊んでくれたお礼にイルカたちが背中に乗せて送って行ってくれる。
+　次の探索表ロールを1回振りなおしてもよい。
+　〔自身:水棲〕
+　　受ける[疲労]の強度は20になる。
+EOS
+             ],
+             [ 88, <<EOS
+夢のリゾート計画：
+　海岸線を見渡せる崖の上に、打ち捨てられた白いコテージを発見した。
+　風雨にさらされ傷んでいるが、修理すればまだまだ使えそうだ。
+　機会があったらキミたちの別荘にしてみるのもよいかもしれない。
+　キミたちはこの建物の場所を忘れないよう、しっかりと地図に書き込んだ。
+EOS
+             ],
+             [ 89, <<EOS
+助けをよぶ声：
+　深い森の中を歩いていると、どこからともなく助けを呼ぶささやき声が聞こえてきた。
+　PC全員は[知覚判定]を行なうこと。
+　最も達成値の高いPCは、石の下から芽吹いたばかりの小さな双葉と、生まれたばかりの〈樹の精霊〉を発見する。
+　キミが石をどけてやると精霊は大いに感謝し、キミの[疲労]の強度を-60してくれる。
+　〔クリティカル〕
+　　精霊はさらに「芽吹きの甘露(苺のふんわりカップケーキ(『LHZB2』P145)相当)」をプレゼントしてくれた。
+　〔サブ職:農家、薔薇園の姫君〕
+　　判定を2回振って好きなほうを選べる。
+EOS
+             ],
+             [ 90, <<EOS
+草むらかき分けて：
+　キミたちが進む深い森の中は、背丈ほどもありそうな下生えが生い茂っている。
+　徒歩では視界も通らず、草を刈り、道を切り開きながらじりじりと進むしかない……。
+　PCは全員[疲労:60]を受ける。
+　〔自身:騎乗、飛行〕
+　　[疲労]を受けなくてもよい。
+EOS
+             ],
+             [ 91, <<EOS
+もうかりまっか？：
+　海沿いの道で〈大地人〉の行商人と出会った。
+　せっかくだから何か買って、顔を繋いでおくのもいいかもしれない。
+　PCのうち希望者はログチケット:アザーゲット1枚を消費してこの行商人とコネクションを取得してもよい。
+　関係は「ビジネス」か「敬意」となる(PCが選択)。
+　またコネクションとは無関係に一般的なアイテムを売買できる。
+EOS
+             ],
+             [ 92, <<EOS
+世捨て人の工房：
+　寂れた高原を旅するキミたちは古い庵を発見する。
+　中は無人だったが、なんと小さな魔法工房がしつらえられていた！
+　【因果力】を1点消費したPCは自身が取得している《ファーマシスト》、《スクライブスクロール》、《インビュー》のうち1つを即座に実行してもよい。
+EOS
+             ],
+             [ 93, <<EOS
+電光石火！？：
+　山の天気は不安定だ。
+　先ほどまでの晴天が嘘のように激しい雷雨が吹き荒れている。
+　うわっ、すぐ近くに落ちた！？
+　PCは全員「難易度:15」の[回避判定]を行なう。
+　強度20以上の[軽減(電撃)]を持つPCはこの判定に自動成功してよい。
+　〔判定成功〕
+　　回避成功！
+　　危うく直撃するところだった。
+　〔判定失敗〕
+　　避ける間もなく雷に打たれてしまい[疲労:100]を受ける。
+EOS
+             ],
+             [ 94, <<EOS
+砂の中の金貨：
+　キミたちは苦戦の末、砂浜で遭遇した〈巨大蟻地獄〉を退治した。
+　早速巣穴の中を探索だ！
+　PCの代表者1名は「難易度:18」の[知覚判定]を行なう。
+　この判定は[偵察]タグがついているかのように扱う。
+　〔判定成功〕
+　　キミたちは2D*120Gを手に入れた。
+　　代表者のサブ職が〈採掘師〉か〈賭博師〉ならば、2Dの出目に+1すること。
+　〔判定失敗〕
+　　何も得られない。
+EOS
+             ],
+             [ 95, <<EOS
+ローリングストーン：
+　針葉樹に覆われた高山の道を歩くキミたちの耳に「カラカラ」と不吉な音がする。
+　視線を上げれば落石に気が付く！
+　全員1Dを振ること(【因果力】1点で+1Dしてもよい)。
+　〔1D:1～3〕
+　　大きな石がキミに直撃！
+　　死ぬほど痛い上に【因果力】2点を失う。
+　〔1D:4～〕
+　　危うく難を逃れた。
+EOS
+             ],
+             [ 96, <<EOS
+イースタルの夕日：
+　真っ赤な夕日が金色に輝く波間に溶けていく。
+　潮騒だけが響く中、この世界でも変わらぬ姿を見せる夕日の中でキミたちは話し合う。
+　PCは全員2Dを振ること。
+　一番低い出目のPCは、一番高い出目のPCにコネクションを取得する。
+　関係は「親愛」か「友情」となる(PCが選択)。
+　このコネクションはアフタープレイで消滅するがログチケット:アザーゲット1枚を消費すれば維持してもよい。
+EOS
+             ],
+             [ 97, <<EOS
+大きな樹の迷宮：
+　大森林の中央にそびえる巨大樹の根元に薄暗い穴を発見。
+　探るのならば、代表者1名が「難易度:16」の[知覚判定]を行なう。
+　〔判定成功〕
+　　複雑に絡み合った通路の奥で「賢者の石[コア素材](620G)」と2D*100Gを得る。
+　〔判定失敗〕
+　　迷ってしまった！
+　　1日経過してしまう上に[疲労:50]を受ける。
+　〔サブ職:探検家、採掘師〕
+　　判定に+2。
+EOS
+             ],
+             [ 98, <<EOS
+前門の鬼、後門の竜：
+　深い山懐に抱かれた廃神殿には〈陽炎悪鬼〉の大群が住み着いていた。
+　多勢に無勢、キミたちは速やかに撤退を開始する。
+　PCの代表者1名は「難易度:18」の[運動判定]を行なう。
+　この判定は[偵察]タグがついているかのように扱う。
+　〔判定成功〕
+　　追跡を振り切る。
+　　損害はない。
+　〔判定失敗〕
+　　〈陽炎悪鬼〉を振り切ったと思ったら、前から〈翡翠竜〉が！
+　　PCは全員[疲労:60]を受け【因果力】1点を失う。
+　〔サブ職:生還者、追跡者〕
+　　判定に+2。
+EOS
+             ],
+             [ 99, <<EOS
+古代の工房：
+　山奥で古代の工房を発見した。
+　【因果力】1点を消費したPCは、自身が取得している《クロースメイク》、《ウッドクラフト》、《アーマーフォージ》、《ファーマシスト》を即座に1回実行してもよい。
+　アルヴの工房は性能抜群だ。
+EOS
+             ],
+             [100, <<EOS
+揺れるつり橋：
+　目の前には霧深い谷と、朽ちて落ちかけのつり橋が揺れている。
+　どうにか向こう岸に渡らねばならない。
+　PCは全員「難易度:18」の[運動判定]を行なう。
+　〔判定成功〕
+　　無事につり橋を渡りきることができた。
+　〔判定失敗〕
+　　足を滑らせてしまった！
+　　落下こそ免れたが、渡りきるのにひどく苦労してしまった。
+　　[疲労:100]を受ける。
+　〔自身:飛行〕
+　　判定は自動成功。
+　〔サブ職:軽業師、船乗り〕
+　　判定に+2。
+EOS
+             ],
+             [101, <<EOS
+真っ白なキミ：
+　キミたちは深い雪山を進む。
+　周囲の木々に降り積もった雪が、風にあおられ頭上から降り注いできた！
+　PCは全員2Dを振ること。
+　一番低い出目のPCは不運にも頭から雪まみれになってしまい、[雪だるマン]タグが与えられる。
+　このタグはシナリオ終了まで解除されない。
+EOS
+             ],
+             [102, <<EOS
+打ち捨てられた砦：
+　風化した砦の廃墟をキミたちは発見した。
+　内部を探索したい場合1日を必要とする。
+　探索した場合不思議な魔方陣を発見し、PC全員の[疲労]の強度が-50される。
+　また、「夜明け色の軟玉[コア素材](680G)」を得る。
+EOS
+             ],
+             [103, <<EOS
+白き魔物：
+　天候が急変し、ブリザードのような雪と風が吹き荒れる。
+　身体は凍りつきそうに冷え、視界は真っ白で何も見えない。
+　このままでは身動きが取れなくなってしまう。
+　強度20以上の[軽減(冷気)]と[暗視]の両方を持たないPCは[疲労:100]を受ける。
+EOS
+             ],
+             [104, <<EOS
+強襲の雪獣：
+　〈雪トロール〉の群れが襲撃してきた！
+　PCの代表者1名は「難易度:18」の[知識判定]を行なう。
+　〔判定成功〕
+　　キミの指示で敵の弱点を突き、危なげなく勝利を収めた！
+　　【因果力】3点を得る。
+　〔判定失敗〕
+　　敵の情報がつかめず、苦戦を強いられた。
+　　PCは全員[疲労:60]を受ける。
+　〔サブ職:学者、軍師〕
+　　判定に+2。
+EOS
+             ],
+             [105, <<EOS
+久しぶりのお客様：
+　野営場所を探すキミたちは、廃屋の前で〈家事妖精〉がぽつりと佇んでいるのを見つけた。
+　キミたちがここで一夜を過ごすことにするなら、彼女は久しぶりに訪れたお客様を精一杯もてなしてくれる。
+　この時キミたちは[食料]アイテムを提供してもよい。
+　翌朝には彼女の姿は消えているが、[食料]アイテムを提供したPCの手にはいつの間にかキラキラ光る小石が握られている。
+　【因果力】1点を得ること。
+EOS
+             ],
+             [106, <<EOS
+〈大地人〉の特訓：
+　夜になって一夜の宿と思い立ち寄った〈大地人〉の小さな砦でキミたちは手厚い歓待を受けた。
+　キミたちは、出発を1日遅らせて兵士たちの訓練を手伝ってもよい。
+　訓練を買って出るPCは、特技喪失する代わりに【因果力】3点を得る。
+　とても厳しい訓練を立派にやり遂げた兵士たちの歓声を背に、キミたちは再び旅路につく。
+EOS
+             ],
+             [107, <<EOS
+鳥妖の巣窟：
+　廃ビルに巣くう〈女面妖鳥〉がキミたちに対して集中攻撃を仕掛けてきたが、果敢に戦い見事に撃破。
+　戦利品として「人面鳥の羽根[コア素材](740G)」を手に入れた。
+　PC全員は1～6の好きな数字を申告すること。
+　代表者1名が1Dを振って、申告と同じ数字が出たPCは、戦闘のダメージで[疲労:80]を得る。
+　それ以外のPCは【因果力】1点を得る。
+EOS
+             ],
+             [108, <<EOS
+野営地の攻防：
+　野営地に押し寄せたアンデッド〈反魂悪党〉が思いのほか手ごわかった。
+　奴らは闇に紛れて不意打ちを仕掛けてきたのだ。
+　PCは全員1Dを振ること。
+　〔1D:1～4〕
+　　PCは[疲労:100]を受ける。
+　〔1D:5以上〕
+　　PCは[疲労:35]を受ける。
+　〔自身:暗視〕
+　　出目に+2。
+EOS
+             ],
+             [109, <<EOS
+今夜は天ぷらだ：
+　幅5mほどの名も知れぬ川をみつけた。
+　通り過ぎてもいいのだが、釣りをしてもよい。
+　釣りをする場合半日の時間を失うが「若葉アユ[換金](100G)」を1D匹釣り上げる。
+EOS
+             ],
+             [110, <<EOS
+吹雪の行軍：
+　山の天気は変わりやすい。
+　さっきまで青空が見えていたのに、今や地吹雪で足元すら見えないありさまだ。
+　[高位保護]を持たないPCは[疲労:100]を受ける。
+EOS
+             ],
+             [111, <<EOS
+ムーンライトフィーバー：
+　キミたちは道中で打ち棄てられた廃城で一晩を明かす。
+　ふと気が付くと月明りの下で輪になって妖精たちが踊っているようだ。
+　つられて踊りの輪に加わる場合、参加者は「難易度:16」の[運動判定]もしくは[操作判定]を行なう。
+　参加者に〈吟遊詩人〉がいるならば全員の判定に+2。
+　〔判定成功〕
+　　妖精たちはキミに大喝采と「回復薬(上級)(『LHZB2』P146)」1つを贈る。
+　〔判定失敗〕
+　　妖精たちは納得がいかない様子だ。
+　　キミは夜が明けるまで踊り、演奏を続けさせられた。
+　　[疲労:50]を受ける。
+EOS
+             ],
+             [112, <<EOS
+幽霊屋敷へようこそ！：
+　一夜の屋根にと求めた山中の古い屋敷はホラームード。
+　全員2Dを振る。
+　一番出目が小さかった人は夜中にトイレに行きたくなり、仕方なく暗闇の中へ……。
+　残ったPCの中で希望者は、トイレに行ったPCを驚かしてもよい。
+　その場合トイレに行ったPCから関係「仇敵」のコネクションを得る。
+　驚かさずに付き添うのならば「関係変化表(『LHZB2』P59)」で1回ロールしてもよい。
+EOS
+             ],
+             [113, <<EOS
+突然の雪崩！：
+　キミたちが穏やかな天候の下、雪深い山道を進んでいたその時、地鳴りと共に突然の大雪崩が襲いかかってきた！
+　PCの代表者1名は雪崩の方向を見極めるため「難易度:16」の[解析判定]を行なう。
+　〔判定成功〕
+　　[飛行]状態のPCは無事に逃げおおせたが、それ以外のPCは[疲労:30]を受ける。
+　〔判定失敗〕
+　　PC全員は[疲労:70]を受け特技喪失する。
+　〔サブ職:生還者、辺境巡視〕
+　　判定に+2。
+EOS
+             ],
+             [114, <<EOS
+ノーストリリアの遺産：
+　街道沿いに遺跡らしい地下への入り口を発見する。
+　内部を探索する場合は半日を消費する。
+　探索した場合、通路の奥にはアルヴとの戦争当時のものらしき回復の魔方陣が残っていた。
+　〈猫人族〉、〈狼牙族〉、〈狐尾族〉、〈法儀族〉のPCは、魔方陣の効果で[軽減(光輝、邪毒、精神):30]を得る。
+　この軽減はOSとして扱いシナリオ終了時まで持続する。
+EOS
+             ],
+             [115, <<EOS
+雪原の罠：
+　雪原を進むキミたちの足元が突如崩れ始める。
+　クレバスだ！？
+　最も【行動力】が低いPCが不運にも逃げ遅れ、クレバスの底に飲み込まれかける。
+　クレバスに落ちて[消耗品]タグを持つアイテムすべてを失うか、踏みとどまり【因果力】2点を失うか選ぶこと。
+EOS
+             ],
+             [116, <<EOS
+珍味、食べる？：
+　キミたちは旅の途中で立ち寄った城砦で、任期を終えた兵士たちの宴にでくわした。
+　そこでひときわ異彩を放っていたのが……魚の塩漬けだ！
+　とにかく臭い！
+　だが彼らはとても美味しそうに平らげている。
+　勇気を出して口にしたPCは、[疲労]の強度が-100される代わりに、セッション終了時まで[超臭い]タグを得る。
+EOS
+             ],
+             [117, <<EOS
+かつての星の里：
+　街道から少し離れた山腹に、キミたちは天文台の廃墟を発見した。
+　建物はかなり傷んではいるが、きちんと手を入れてやれば十分住むことはできそうだ。
+　ひょっとしたら、再び星を見ることができるかも知れない！
+　しかし残念ながら、今ここでできることは一夜の雨露をしのぐことぐらいだろう。
+EOS
+             ],
+             [118, <<EOS
+追いすがる吹雪：
+　深い山中で〈氷雪巨人〉に遭遇してしまい、一時間近くも追い回されることになった。
+　吹雪のブレスがキミたちを何度も襲う。
+　強度30以上の[軽減(冷気)]を持たないPCは[疲労:80]と特技喪失を受ける。
+EOS
+             ],
+             [119, <<EOS
+鎧袖一触！：
+　山道で〈大地人〉の隊商を襲う〈灰斑犬鬼〉の群れを発見した。
+　数こそ多かったが歴戦の〈冒険者〉であるキミたちの敵ではない！
+　残らず退治してやると、隊商の商人たちはたいそう喜んでいた。
+　900Gの謝礼をもらった。
+EOS
+             ],
+             [120, <<EOS
+止まったら死ぬ！？：
+　周囲が突然の地響きに包まれ、斜面の上に積もった雪が崩落を起こした。
+　雪崩だ！
+　難を逃れるため、PCは全員「難易度:17」の[運動判定]を行なう。
+　〔判定失敗〕
+　　逃げ遅れたキミはあっという間に雪の奔流に飲み込まれてしまった。
+　　抜け出すために[疲労:120]を受ける。
+　〔自身:騎乗〕
+　　判定に+1Dする。
+　〔サブ職:探検家、辺境巡視〕
+　　判定に+2。
+EOS
+             ],
+             [121, <<EOS
+〈冒険者〉の川流れ：
+　キミたちは目の前に広がる氷混じりの大河を渡る。
+　PCは全員2Dを振ること。
+　一番低い出目のPCは渡河の最中に足を滑らせ流されてしまうが、間一髪で他のPC(流されたPCが指定してよい)に救われる。
+　流されたPCは助けてくれた相手に対するコネクションを取得すること。
+　関係は「恩」か「友情」(助けたほうが選んでよい)となる。
+　このコネクションはアフタープレイで消滅するがログチケット:アザーゲット1枚を消費すれば維持してもよい。
+EOS
+             ],
+             [122, <<EOS
+サルベージ：
+　川底に何か光るものが見えた気がする。
+　PCの代表者1名は「難易度:17」の[知覚判定]を行なう。
+　〔判定成功〕
+　　川底の砂に半ば埋もれた小さな宝石箱を発見！
+　　「ブリザードダイヤ[コア素材](980G)」と「色とりどりの宝石[換金](300G)」x5を得る。
+　〔判定失敗〕
+　　何も見つけられなかった。
+　〔サブ職:海賊、漁師〕
+　　判定に+2。
+EOS
+             ],
+             [123, <<EOS
+ガスが濃くなってきたな：
+　渓谷を進んでいると周囲に毒々しい煙が立ち込めてきた。
+　これは……火山性の毒ガスだ！？
+　強度30以上の[軽減(邪毒)]を持たないPCは[疲労:100]を受ける。
+　〔サブ職:薬師、毒使い〕
+　　素早くガスの正体を見抜き、被害を受けなかった。
+EOS
+             ],
+             [124, <<EOS
+樹海の白狐：
+　樹海の奥深くを歩いていると突然狐に似た精霊が現れた。
+　面白い話をするか、[食料]アイテムを与えると、代わりに「エメラルドの首飾り[換金](500G)」1D個をプレゼントされる。
+　〔サブ職:ちんどん屋、料理人〕
+　　さらに「金のあぶらげ[換金](1000G)」をプレゼントされる。
+EOS
+             ],
+             [125, <<EOS
+吊るされた〈冒険者〉：
+　薄暗い密林でキミたちは足元に違和感を覚える。
+　しかし時すでに遅く、うごめくツタによって逆さ吊りにされ、所持品をあたりにぶちまけてしまった！
+　PC全員は[運動判定]を行なうこと。
+　この時[暗視]状態でないPCは、この判定に-1Dされる。
+　最も達成値の低いPCは[消耗品]アイテムを2個失い、[圧迫:1]を受ける。
+　〔ファンブル〕
+　　さらに[圧迫:1]を受ける。
+　〔サブ職:斥候、罠師〕
+　　この判定に+1D。
+EOS
+             ],
+             [126, <<EOS
+儀式の生贄：
+　樹海を進むキミたちは〈蜥蜴人〉の集落に出くわし、運悪く生贄にされかけた〈狐耳族〉の少女を救出する。
+　信心深い少女はPCの代表者(PCが相談して選択)を「イキガミさま」と崇拝する。
+　代表PCは[神さま]タグを得る。
+　このタグはシナリオ終了まで解除されない。
+EOS
+             ],
+             [127, <<EOS
+大自然のおこぼれ：
+　キミたちは密林で〈魔熊〉が〈地獄蜂〉の巣を襲い、ゆうゆうと食事をしているところに出くわした。
+　姿を隠すためにPCは全員「難易度:21」の[運動判定]を行なう。
+　〔全員が判定成功〕
+　　「地獄蜂の毒針[コア素材](1060G)」を得る。
+　〔1人でも判定失敗〕
+　　〈魔熊〉に気付かれて追いかけ回された。
+　　何も入手できない。
+EOS
+             ],
+             [128, <<EOS
+息詰まる戦い：
+　火山が近いのだろうか？
+　ガスがもくもくと立ち込める渓谷を進むキミたちは、襲ってくるモンスターを倒しながら先を急ぐ。
+　PCは全員1Dを振ること。
+　出目が1～2の場合、戦闘中に毒ガスを吸い込んでしまい[疲労:80]を受ける。
+EOS
+             ],
+             [129, <<EOS
+妖精のキッチン：
+　樹海を進むキミたちの目の前に、突如ファンシーな外見の建物が姿を現した。
+　中からはいい匂いが漂ってくる……
+　窓から覗き込んでみると、そこは〈厨房妖精〉が料理を作る妖精食堂だった！
+　もし、中に入ってみるなら「ニンゲンのお客様」と歓迎され、ひとりにつき「魔触媒」を1個消費することで不思議な妖精のお菓子を食べることができる。
+　それまでに受けていた[疲労]はすべて解除される。
+EOS
+             ],
+             [130, <<EOS
+天然の迷宮：
+　キミたちは樹海を進む。
+　深く複雑な緑の迷宮は方向感覚を狂わせる！
+　PCは全員「難易度:18」の[知覚判定]を行なう。
+　2人以上判定に失敗した場合、方角を見失ってしまう！
+　全員[疲労:80]を受ける。
+　〔ファンブル〕
+　　シナリオ終了まで[方向音痴]タグを得る。
+　〔サブ職:探検家、地図屋〕
+　　判定に+2。
+EOS
+             ],
+             [131, <<EOS
+謎の〈古来種〉：
+　樹海のはるか奥でキミたちは正体不明の〈古来種〉と出会った。
+　彼女はとてもとても退屈しており、PC(何人でもよい)は今までの冒険のエピソードを、実時間で5分以内にまとめて語ってもよい。
+　その場合、彼女はたいそう喜び褒美として「万年楓の小枝[魔触媒30][譲渡不可](820G)」を授ける。
+　望むPCは一時的なコネクションを取得してもよい。
+　関係は「友情」か「親愛」となる。
+　このコネクションはアフタープレイで消滅するがログチケット:アザーゲット1枚を消費すれば維持してもよい。
+EOS
+             ],
+             [132, <<EOS
+モンスター退治の後は：
+　キミたちは激しい戦闘の末、火口にほど近い谷間に巣食う〈一眼巨人〉の一団を退治した。
+　PCの代表者1名は「CR*5+2D」を振る。
+　このロールは財宝表であるかのように特技やアイテムの効果を受けてよい。
+　130以上ならばキミたちは立派な宝箱を発見した！
+　「極上の玉鋼[換金](500G)」が収められていた。
+　140以上ならばさらに「氷不死鳥の羽根[コア素材](1160G)」を発見する。
+EOS
+             ],
+             [133, <<EOS
+襲いくる水流：
+　上流で大雨が降ったのか、大河の水は氾濫し、濁流となってキミたちに襲い掛かってきた！
+　PCは全員「難易度:18」の[耐久判定]を行なう。
+　〔自身:[水棲]or[飛行]〕
+　　判定に自動成功してよい。
+　〔判定失敗〕
+　　水流に大きく流されてしまい[疲労:80]と[圧迫:1]を受ける。
+EOS
+             ],
+             [134, <<EOS
+たまには鉱夫もいいよね：
+　火山に連なる渓谷は、良質の鉱床が露出しており、採掘にはもってこいだ。
+　この場所で1日を費やすならば、つるはしやハンマーを振るい、それぞれ「血色鉱石[換金](80G)」*1D個を得る。
+　〔サブ職:採掘師、採取人〕
+　　入手個数に+2してよい。
+EOS
+             ],
+             [135, <<EOS
+バッドドリップ！？：
+　足元も見えない深い樹林を切り開いて進むキミたちは、うっかりカラフルな毒蛇の尻尾を踏みつけてしまった！
+　PCは全員2Dを振ること。
+　一番低い出目のPCは怒った蛇に噛み付かれてしまう。
+　毒が回って頭がクラクラしてきた。
+　【因果力】2点を失う。
+EOS
+             ],
+             [136, <<EOS
+大冒険の舞台：
+　高台から密林を見下ろすと、極彩色の鳥が群れを成して眼下を横切っていくのが見えた。
+　この雄大な自然は、キミたちの挑戦を待ち受けている。
+　さあ、冒険を続けようじゃないか！
+　この探索表の出来事で次に何らかの判定をする場合ロールに+1D。
+　この効果はひとりでも全員でも次の判定のみ、一度効果を受けるかシーン終了時まで持続する。
+EOS
+             ],
+             [137, <<EOS
+氷河の源：
+　キミたちは巨大な氷河を遡るうちに、氷の中に閉じ込められた魔法の輝きを見つける。
+　[火炎]タグのついた武器、もしくは魔法攻撃を持つPCがいるのならば運試しに1Dを振ること(1人1回まで)。
+　1～4なら「太古の氷[換金](100G)」を得るが、5以上なら「永久氷河の核[コア素材](1240G)」を得る。
+EOS
+             ],
+             [138, <<EOS
+捉えがたき強敵：
+　火山から吹き出す白煙に紛れて〈煙霧精霊〉の群れが忍び寄る。
+　戦いは乱戦から始まり、キミたちの勝利で終わったものの、実に厳しい戦いだった。
+　PC全員は【因果力】1点を失う。
+　さらにPCは全員1Dを振ること。
+　強度30以上の[軽減(邪毒)]を持つPCは出目に+2。
+　〔1D:1～4〕
+　　PCは特技喪失する。
+　〔1D:5以上〕
+　　特に損害はない。
+EOS
+             ],
+             [139, <<EOS
+小さな親切大きな宝石：
+　樹海の草木に埋もれた古い祭壇を見つけた。
+　祭壇を掃除して、お供えをしてもよい。
+　お供えにしてささげたアイテムの価格合計が300G以上なら、どこからともなく
+　「そ、そんな高価なものをささげないで！」
+　という言葉が聞こえてきて、頭の上から「大粒のルビー[換金](1200G)」が降ってくる。
+EOS
+             ],
+             [140, <<EOS
+密林の魔女：
+　キミたちは密林の奥で不思議な小屋を発見した。
+　するとドアがひとりでに開き、
+　「何をグズグズしてるんだい！水汲みに薪割りかまどの掃除、やることは山積みだよ！」
+　としゃがれた老婆の怒声が飛び出してくる。
+　なんて尊大な！
+　しかし素直に従うならば、[疲労:300]を作業したPCで自由に分配すること。
+　その後老婆はPCが所持するアイテム1つに、マジックグレード(1D/2)のプレフィックスド効果(任意)を付与してくれる。
+　〔サブ職:メイド、エルダーメイド〕
+　　各PCが受ける[疲労]の強度を-10する。
+EOS
+             ],
+             [141, <<EOS
+黒剣のお墨付き：
+　山上湖に巣くった〈火竜〉と戦う〈黒剣騎士団〉をキミたちは発見した。
+　彼らに加勢してモンスターとの戦闘に参加してもよい。
+　参加する場合半日を消費するが、希望者は「やるじゃねえか！」と感心した団長のアイザックとのコネクションを得る。
+　関係は「友情」となる。
+　このコネクションはアフタープレイで消滅するがログチケット:アザーゲット1枚を消費すれば維持してもよい。
+EOS
+             ],
+             [142, <<EOS
+アレだけは苦手なんだよね……。：
+　野営の明かりはいろんなことを思い起こさせる。
+　PC全員は苦手なエネミー(ゴキブリ、人面魚、幽霊、ムカデなどなど)や現象を話そう。
+　全員自分を再確認して[弱点(告白した対象からの攻撃):20]を得る。
+　これはOSとして扱いシナリオ終了時まで持続する。
+EOS
+             ],
+             [143, <<EOS
+悪意の嵐：
+　太古の戦場で突然の嵐に見舞われた。
+　暴風、落雷、大粒の雹と雨、そして怨嗟のうめき声。
+　PCの代表者1名は「難易度:22」の[解析判定]を行なう。
+　この判定は[偵察]タグがついているかのように扱う。
+　〔判定成功〕
+　　避難場所を見つけた。
+　　PCは全員[疲労:40]を受ける。
+　〔判定失敗〕
+　　避難場所が見つからない。
+　　怨嗟に蝕まれPCは全員[疲労:100]と特技喪失を受ける。
+EOS
+             ],
+             [144, <<EOS
+仕上げは魔触媒：
+　岩陰に作りかけの魔方陣を発見した。
+　魔力を足してやれば力を発揮しそうだ。
+　希望者は自分のCR以上のランクを持つ「魔触媒」1つをささげること。
+　ささげた希望者は、自身の「タイミング:メジャーアクション」の特技のSRを+1できる(最大SRは超えない)。
+　この効果はOSとして扱いシナリオ終了時まで持続する。
+EOS
+             ],
+             [145, <<EOS
+沈む！沈む！？：
+　じめじめとぬかるんだ、陰気な湿地帯を進むキミたちだが、突然ひざまで泥に埋まってしまう。
+　しまった、底なし沼だ！？
+　[飛行]状態ではないPCは【因果力】1点を失う。
+　[重鎧]もしくは[中鎧]を装備しているPCはさらに【因果力】1点を失う。
+EOS
+             ],
+             [146, <<EOS
+あばれる恐竜：
+　シダの生い茂る沼地で突然二足歩行の恐竜の群れが現れる。
+　逃げずに調教を試みるPCは「難易度:27」の[交渉判定]を行なう。
+　この時[疲労:20]を受けるごとに+1のボーナスを得られる(最大+4まで)。
+　〔判定成功〕
+　　「剣二足竜の角笛(『LHZB2』P161)」を得る。
+　〔判定失敗〕
+　　恐竜たちは去っていく。
+EOS
+             ],
+             [147, <<EOS
+一夜の勝負：
+　キミたちが野営をしていると、タキシード姿の青白い〈亡霊〉が姿を現した。
+　身構えるキミたちに〈亡霊〉はダイスで賭けをしようと持ちかけてくる。
+　もしも勝負を受けるなら代表者は【因果力】1点を支払い奇数か偶数を指定すること。
+　そののちGMは2Dを振ること。
+　的中したならば「財宝表:金銭」を1回ロールしてもよい。
+　4回勝負を行なうと〈亡霊〉は満足したのか朝日の中に消えていく。
+　勝負を受けず、〈亡霊〉を排除するならばPC全員は[疲労:100]を受ける。
+EOS
+             ],
+             [148, <<EOS
+不浄なる戦場：
+　キミたちはアルヴの怨念が満ちると伝えられる古戦場で、呪われた〈骸骨巨人〉と激しい遭遇戦を繰り広げた。
+　戦いにはなんとか勝利し「無骨な白骨[換金](400G)」を1D個手に入れたが、かなりの消耗を強いられることになってしまう。
+　PCは全員[疲労:120]と特技喪失を受ける。
+EOS
+             ],
+             [149, <<EOS
+飢えた〈魔狂鼠〉：
+　キミたちは黒い濁流のような、恐ろしい数の〈魔狂鼠〉の群れに遭遇した。
+　[食料]タグを持つアイテムすべてを投げ捨てて必死で逃げれば彼らを避けることができる。
+　食料を守るためには戦うしかない。
+　その場合は[疲労:120]を受け、さらに【因果力】1点を失う。
+EOS
+             ],
+             [150, <<EOS
+怨念の森：
+　陰鬱な森がキミたちに牙を剥いた。
+　足元の腐葉土は瘴気を放つ泥濘へと変じ、ぬめるような霧が魂魄を削り取ってゆく。
+　[軽減(邪毒)]もしくは[高位保護]を持たないPCは【因果力】3点を失う。
+EOS
+             ],
+             [151, <<EOS
+お憑かれ様：
+　アンデッドのうろつく古戦場で、キミたちは不注意から小さな祠を倒してしまった。
+　PCは全員2Dを振ること。
+　一番低い出目のPCは猛烈な寒気とともに[亡霊憑き]タグを得る。
+　このタグはシナリオ終了まで解除されず、タグを持つPCは判定でファンブルをするたびに[疲労:20]を受ける。
+EOS
+             ],
+             [152, <<EOS
+夜間飛行：
+　キミたちは湿地帯で野営することになった。
+　PCは全員2Dを振ること。
+　一番高い出目のPCが見張りをしていると、月明かりの下、小さな赤いクモの子供たちが何百匹も風に乗って飛んでいく、幻想的な光景を目にした。
+　[疲労]が-30される。
+　また「煉獄クモの燃糸[コア素材](1540G)」を入手する。
+EOS
+             ],
+             [153, <<EOS
+沸き立つ沼：
+　硫黄の臭いが立ち込める泥沼が突如沸騰し始めた！
+　毒ガスと灼熱の泥がキミたちに襲い掛かる。
+　[高位保護]もしくは[天然]プロップの影響を受けないPC以外は1Dを振る。
+　この時【因果力】1点を支払って1Dではなく2Dを振ってもよい。
+　出目の合計が2以下のPCは[戦闘不能]となる。
+　3～5ならば[疲労:100]を受ける。
+　6以上なら被害はない。
+EOS
+             ],
+             [154, <<EOS
+ハイパードロップタイム！：
+　ひょんなことからキミたちは大量のエネミーに追い回され、苦戦の末全滅させた。
+　PCは【因果力】を任意の数だけ消費してもよい。
+　各自消費した【因果力】と同数のダイスを振り、全員の出目を合計すること。
+　ダイスの合計値が3以上ならPCたちは1500Gを得る。
+　8以上でさらに「古代のメダル[コア素材](1540G)」と「魔凝結[魔触媒29](770G)」2つを得る。
+　13以上でさらに「ダイヤの原石[換金](2300G)」1つを得る。
+EOS
+             ],
+             [155, <<EOS
+心霊スポット？：
+　静かに水を湛えた湖のそばを通った時、急な悪寒に襲われた。
+　湖面に目をやると、水面下にひしめく亡者の群れと目が合ってしまった。
+　呪詛の視線がキミたちを捉える。
+　PCは全員1Dを振ること。
+　[高位保護]を持つPCは出目に+2。
+　〔1D:1～3〕
+　　ごっそり気力を削られる。
+　　【因果力】2点を失う。
+　〔1D:4以上〕
+　　視線に力を込めて睨み返す。
+　　やっぱ気合いだな。
+EOS
+             ],
+             [156, <<EOS
+武人の誉れ：
+　荒野の只中、数多の折れた武器や朽ちた鎧でできた玉座に、古めかしい甲冑をまとったアルヴの亡霊が座している。
+　彼はキミたちに「勝者には名誉を、敗者には死を」と告げ、一騎打ちを申し込んできた。
+　断るならば、彼の姿は玉座と共に掻き消える。
+　PCの代表者1名は「難易度25」の[耐久判定]を行なう。
+　判定に失敗した場合[疲労:300]を受ける。
+　判定後、判定したPCが[戦闘不能]状態でなければ、彼は「人斬りのギガントエストック[M2]([人間]に対しダメージ+40。『LHZB2』P136&P165)」を残し、光の粒子となって天に昇る。
+EOS
+             ],
+             [157, <<EOS
+見定める視線：
+　旅の道中で、キミたちは何度もモンスターの襲撃を受けた。
+　熟練の冒険者であるキミたちは危なげなくそれらを撃破し、合計で2D*200Gの戦利品を得た。
+　しかし、その様子を敵意に満ちた視線がずっと監視していたことには気づかなかったようだ。
+　このシナリオのクライマックスにおいて、GMは【因果力】(PC人数*2)点ぶんのGM用EXパワーを追加で使用できる。
+EOS
+             ],
+             [158, <<EOS
+突然の闇夜：
+　沼に潜む〈蛙竜〉に出会ったキミたちは、フォッグブレスで突然の暗黒に突き落とされる。
+　PCは全員1Dを振ること。
+　[水棲]、[飛行]、[暗視]を持っているのならば1つにつき出目に+1を得られる。
+　〔1D:1～4〕
+　　逃亡中に焼け付くような毒液を受けて[疲労:120]を受けてしまう。
+　〔1D:5～〕
+　　光を閉ざす霧の中を逃走して[疲労:60]を受ける。
+EOS
+             ],
+             [159, <<EOS
+幻の金鉱：
+　キミたちは瘴気を漂わせる金鉱を見つけた。
+　この金鉱に潜って金を探したいPCは2Dを振ること。
+　出た出目がいままでPCたち全員でこの金鉱を探った数と等しいか小さければ「金鉱石[換金]700G」を得る。
+　出目が大きい場合、PCたち全員は[疲労:80]を受けて気絶し丸一日が経過する。
+　目が覚めた時、金鉱は消えている。
+EOS
+             ],
+             [160, <<EOS
+幻惑草の群生地：
+　粘つくような湿気の原生林の中で、甘い香りを放つ花畑にでくわした。
+　PCは全員「難易度:20」の[耐久判定]を行なう。
+　強度30以上の[軽減(精神)]を持つPCは判定に+1D。
+　〔判定成功〕
+　　PCは意識をしっかり保って花畑を抜ける。
+　〔判定失敗〕
+　　PCは幻覚を見て仲間のひとり(失敗したPCが選択)に斬りかかり、その仲間に[疲労:100]を与えてしまう。
+　　この効果は失敗したPC以外であれば《かばう》で代わりに受けることができる。
+　　また失敗したPCは特技喪失を受ける。
+EOS
+             ],
+             [161, <<EOS
+硫黄の谷間：
+　黄褐色の霧が立ちこめる荒々しい谷間にたどり着く。
+　キミたちは呼吸を我慢するようにしてこの危険な谷をくぐり抜ける。
+　PCは全員1Dを振ること。
+　1が出たPCは濁った水たまりに動物の死体を見つけて気持ちが悪くなってしまい、シナリオ終了時まで[弱点(精神):20](OS扱い)を得る。
+EOS
+             ],
+             [162, <<EOS
+火竜との戦い：
+　火炎をまとい大空を支配する恐ろしい火竜との遭遇戦となった。
+　PCは全員1Dを振る。
+　自発的に特技喪失をしてもよく、その場合追加で+1Dしてもよい。
+　ダイスの合計値が(PC人数*4+2)以上ならば勝利して「赤色魔力ダイオード[コア素材](1640G)」を得る。
+　未満であった場合はPCは全員【因果力】1点を失い[疲労:50]を受ける。
+EOS
+             ],
+            ]
 
-    table = ["森のベリー摘み：あなたたちはベリーの茂みを見つける。探索をするＰＣは、難易度８の［知覚判定］を行うこと（ひとり１回だけ）。成功した人は「森のベリー（お好みサンドイッチ相当）」を１つ入手する。",
-             "木漏れ日の散歩道：木漏れ日のさす明るい雑木林が続く。適度に人の手が入っていて、散策気分で歩くことができる。",
-             "群生地にはご用心：あなたたちはトリフィドの群生地に踏み込んでしまった！　激しい戦いの末に辛くも勝利を掴んだあなたたち全員は［疲労：１０］をうける。",
-             "フェアトレード：あなたたちは〈大地人〉の農業家と出会った。２０Ｇを支払うなら「季節の野菜ポトフ」を購入してよい（全部でＰＣと同じ数まで）。",
-             "河の営み：河のほとりを土手に沿って歩く。小さな船着き場の方から威勢の良い声が聞こえる。魚を捕っているようだ。",
-             "越すに越されぬ：川が増水しているせいであなたたちは渡るところを探すのに苦労する。［水棲］タグを所持、得られる手段のあるＰＣがひとりでもいるなら問題はない。いないのであれば大きく迂回して全員［疲労：１０］をうける。",
-             "コボルド遭遇戦：あなたたちはコボルドの群れを蹴散らした！　全員１０Ｇを手に入れる。追いかけて殲滅をするなら全員［疲労：１０］をうけるが、さらに「小牙竜鬼の牙［換金］（１０Ｇ）」を１人１つ手に入れる。",
-             "豊かな田園風景：大きな田んぼが何枚も広がっている。秋ならば黄金色の稲穂が風に揺れている事だろう。",
-             "不運の予感：黒猫が目の前を横切り、カラスが頭上で鳴いて、靴ひもがぶつりと切れた。なんだか寒気まで。あなたたち全員は【因果力】を１点失う。",
-             "友達の友達は友達：ＰＣ全員は２Ｄをふる。最大の出目だった「あなた」は最小の出目だった「対象」と共通の知り合いの話題で盛り上がる。あなたと対象は「因果力チケット」１枚で、それぞれから相手へのコネクションをとってよい。関係は「友情、理解］のどれかになる。",
-             "時はうつろい：道沿いに、外壁だけが残った鉄筋のビルが建っている。今は鳥の住処になっているようだ。",
-             "突然の豪雨！：あなたたちは突然の豪雨に見舞われ、苦労の末なんとか廃屋にたどり着いた。あなたたち全員は［食料］アイテムをひとつ消費すること。ただし［疲労］の回復量は半分となる。［食料］アイテムを消費できない場合は【因果力】１点を失う。",
-             "野の花の咲き乱れる：あなたたちは一面の花畑にたどり着く。探索をするＰＣは、目標値９の［知識判定］を行うこと（ひとり１回だけ）。成功した人は「貴重な薬草［換金］（２０Ｇ）」をひとつ入手する。",
-             "緑のトンネル：立派な広葉樹が立ち並び、緑の天蓋を作り出している。足元には腐葉土が厚く積もり、所々で木の根が顔を見せている。",
-             "ゴブリン強襲部隊：あなたたちはゴブリン強襲部隊との遭遇戦に巻き込まれた。あなたたち全員は［疲労：２０］もしくは「シナリオｎ回」の特技の使用回数消費のどちらか（あなたたちが選択）をうける。",
-             "情報交換：あなたたちは冒険者の集団と出会った。代表者は目標１０の［交渉判定］を行なうこと（適切なサブ職であれば＋２のボーナスを得ても良いだろう）。成功したら付近の情報を教えてもらうことができる。あなたたちのうち１人が次に行う［偵察］行動に＋２をえる。",
-             "淵と漁師：滔々と水を湛えた淵に出た。漁師が投網を投げ入れる姿が見える。",
-             "寸断された小道：がけ崩れで道が崩れ落ちてしまっている。このルートは使えないようだ。［飛行］状態である、得られる手段のあるＰＣがひとりでもいるなら周辺を探索できるので問題はない。そうでないならあなたたちは随分な遠回りを強いられた。あなたたち全員は［疲労：２５］を受ける。",
-             "コボルドを蹴散らせ！：あなたたちはコボルドの巡回隊を蹴散らした。全員［疲労：（１Ｄ６＋５）］を受けてしまうが、戦利品として１人２０Ｇを手に入れる。",
-             "大地の恵み：畝に沿って野菜が青々とした葉を茂らせている。おいしい料理になりそうだ。",
-             "滑って転んで：街道を歩いていると突然巨大なワイルドボアが飛び出してくる。避けた拍子に転んで、あなたたち全員は［消耗品］アイテムを１つ失う。［消耗品］アイテムを持っていなければ、【因果力】１点を失う。",
-             "ぐるぐるずてん：ＰＣ全員は２Ｄをふる。最小の出目だった「あなた」は濡れてすべる草に足を取られて転げ落ちてしまう。助けてあげる仲間は「因果力チケット」１枚で「あなた」に対してコネクションをとってよい。関係は「庇護、親愛］のどれかになる。",
-             "つわものどもが夢の跡：半ば崩れた石垣が続く。亜人との激しい戦いの傷跡も今はかすかに残るばかりだ。",
-             "五里霧中：深い霧がたちこめ、伸ばした手の先すら見えないありさまになる。あなたたち全員は［疲労：２５］もしくは【因果力】１点の減少のどちらか（あなたたちが選択）をうける。",
-             "癒しの湧水：あなたたちはこんこんと湧きだす清水を見つける。目標値１３の［知覚判定］を行うこと。誰かひとりでも成功すれば、これが澄み切って魔力のある水だと判る。汲んでいくことはできないが飲めばその場で【因果力】１点を得る。",
-             "針葉樹の森：針葉樹がごつごつとした樹皮を見せている。下生えは少なく、少し寒々しい印象を受ける。",
-             "魚人の襲撃：あなたたちはサファギンの回遊小隊と激しい戦いを繰り広げた！　あなたたち全員は［疲労：２５］もしくは［消耗品］アイテム２個の喪失のどちらか（あなたたちが選択）をうける。",
-             "助け合い：あなたたちは立ち往生している旅人の馬車と出会った。彼らを手助けする場合、助けたいメンバー全員で［操作判定］を行うこと。達成値の合計が（ＰＣ人数ｘ１０）以上なら、馬車は街道へと戻る。その場合１人２０Ｇの謝礼を受け取る。",
-             "せせらぎとそよかぜ：ゆるやかに流れる小川のほとりに出た。川面を渡る風が心地よい",
-             "ファイト！　一発！：あなたたちは断崖沿いの細い通路を進み、体力だけでなく気力の消耗も強いられた。あなたたち全員は［疲労：１５］を受ける。加えて［疲労：１５］もしくは【因果力】１点の減少のどちらか（あなたたちが選択）をうける。",
-             "巨人と出会う：森の上にひょっこりと頭を出したのは禿頭の巨人だった。戦闘に勝利したあなたたちは全員２５Ｇを手に入れる。巨人のねぐらを見つけるには代表者が［解析判定］で目標値１２に成功しなければならない。成功した場合ねぐらから［魔触媒６］を１Ｄ６＋１個手に入れる。",
-             "休憩する場所：休憩するのにぴったりの道路脇の空き地を見つける。かつて誰かもそう考えたのかたき火の跡もある。",
-             "そんなはずじゃ：道に迷ってしまった。あなたたち全員は【因果力】を２点失う。【因果力】を失えない場合、足りない【因果力】１点につき［疲労：１５］をうける。あなたたちの仲間に［地図屋］か［斥候］がいれば［疲労：１５］は［疲労：１０］で済む。",
-             "好きな子はいる？：休憩、どういったわけかあなたたちは好きな異性の話題へと突入してしまう。［愛情］もしくは［恋］［ときめき］に準じるコネクションを持っているＰＣはそれを告白してもよい。その場合ＧＭはそのＰＣの勇気に感じ入って【因果力】１を与えよう。もし告白の相手がいまパーティーにいる仲間であるならロールプレイ次第でその相手にも与える事。",
-             "あの橋を渡って：赤く朽ちた鉄橋が川面に影を落としている。かつてはこの橋を電車が走っていたのかレールがコケの中に顔を見せている。",
-             "氷のつぶて：大粒の雹が降り注ぎ、容赦なくあなたたちを打ち据える。あたりはひらけた土地で回避のしようもない。あなたたち全員は［疲労：３０］を受ける。",
-             "ドキドキ洞窟：あなたたちは不気味な洞窟を見つける。探索をするＰＣは、目標１２の［知覚判定］を行うこと。失敗した人は［疲労：１０］をうける。１人でも成功した場合、洞窟の奥から１００Ｇはいった金貨の袋を発見する。山分けにしよう！",
-             "茂みを抜けて：背の低い灌木の茂みが点在し、まっすぐに進むのが難しい。しかし踏み分け道があるおかげで迷わずに済みそうだ。",
-             "ここで食い止めろ！：あなたたちは〈大地人〉集落の周辺でワイバーンとの戦闘に突入した！　あなたたち全員は［疲労：３０］もしくは【因果力】２点の減少のどちらか（あなたたちが選択）をうける。",
-             "いつか来た道：あなたたちはまだ駆け出しの〈冒険者〉たちと出会った。何か手助けをしてあげたいなら、［消耗品］をいくつか渡してあげても良いし、何かアドバイスをしてあげても良い。彼らは何度も礼を述べつつ旅立つ。ＧＭはＰＣたちのロールプレイ次第では【因果力】１点を与えても良い。",
-             "渓谷に響く水音：急峻な谷底を勢いよく水が流れる。そう遠くないところから滝の轟く音も聞こえてくる。",
-             "毒の沼：あなたたちは腐臭漂う沼地を横切った。あなたたち全員は目標値１４の［運動判定］をおこなうこと。失敗したＰＣは［消耗品］アイテムを２個失う。",
-             "ゴルゴンスレイヤー：あなたたちは鉄の肌を持つ牛を蹴散らした！　全員【因果力】１点をうしなう。戦利品として「鉄牛の鱗［換金］（３５Ｇ）」を人数分手に入れる。",
-             "開拓者たち：何人かの大地人たちが森の開墾を進めている。種を撒けるようになるにはもう少しかかりそうだ。",
-             "まさかの遭遇：１０メートルを超える緑のドラゴンとでくわす。なんでこんなところに！あなたたちは全力で逃げ出した。あなたたち全員は【因果力】を２点失うか、［補助装備］ひとつに［破損］タグをつける。どちらも出来ない場合、［疲労：３５］をうける。",
-             "しりとりは鬼門：あなたたちは旅のいろどりのために全員でしりとりをしながら歩いている。ＰＣ全員は「ぬ」（ＧＭは変えても良い）で始まる単語をひとつを発表すること。一番恥ずかしい単語を発表したＰＣをみんなで決めよう。そのＰＣに対するコネクションをもっている他のＰＣは、コネクションの関係を［あわれみ］［がっかり］［面白い］に変更しても良い。",
-             "古い神域：大きな石の鳥居が、長年の風雪に耐えてそびえ立つ。厳かな空気が辺りに満ちていた。",
-             "タービュランス：突風は竜巻となった！　あなたたちはなんとかやり過ごすことができたが、無傷とはいかなかった。あなたたちは［疲労：３５］もしくは［消耗品］２個の喪失のどちらか（あなたたちが選択）をうける。",
-             "狙うは大物：あなたたちは素晴らしい釣り場を見つける。探索をするＰＣは［食料］アイテムをひとつ消費して、目標値１３の［操作判定］を行うこと（漁師屋それに準じたサブ職には＋２を与えても良い）。成功した人は「脂の乗ったマス（定番カレー相当）を［１Ｄ／２］匹（最低１匹）入手する。",
-             "古樹の迷宮：節くれだった古木が絡み合い、圧迫感を感じさせる。森がすごしてきた年月の重みが、肩にのしかかってくるようだ。",
-             "多勢に無勢：あなたたちは大蛇の大群に囲まれた！　あなたたち全員は［疲労：４０］もしくは補助装備ひとつに［破損］タグ（あなたたちが選択）をうける。",
-             "祝福の輪：あなたたちは大地人の村でささやかな結婚の宴席に出会う。祝福をしてみたい場合、代表するＰＣは難易度１４の［交渉判定］を行うこと。差し入れの［食料］タグを持つアイテムひとつごとに＋２のボーナスを得る。成功した場合、どこからか爽やかな風が吹きあなたたち全員は【因果力】１点を得る。そうでない場合にも宴会には招かれ、素朴なご馳走を食べて全員の［疲労］は－２０される。",
-             "大河の一滴：清らかな水がこんこんと湧きだしている。この水はきっとあの大河に繋がっているのだろう。",
-             "冷え切った山道：風が吹き付けて凍える山道をあなたたちは進んでいく。［軽減（冷気）］をもつ、あるいは得られるＰＣはこの影響を受けない。そうでないＰＣは【因果力】２点もしくは補助装備ひとつに［破損］タグ（あなたたちが選択）をうける。どちらも選択できない場合、［疲労：５０］をうける。",
-             "ゴブリンの野営地：あなたたちはゴブリンの野営地を蹴散らした！　全員５０Ｇを手に入れる。目標値１４の［知覚判定］を行うこと（１人１回）。成功した人は「ゴブの英雄像［換金］（６０Ｇ）」を１つ入手する。",
-             "連綿と続く営み：山の斜面一面に段々畑が広がる。これだけの田畑を拓くのに、いったいどれほどの時間がかかったのだろう。きっと近くには〈大地人〉の村があるはずだ。"]
-
-    d1, dummy = roll(1, 6)
-    d2, dummy = roll(1, 6)
-    value = (d1.to_i + d2.to_i - 2 + (cr - 1) * 5)
-
-    result = table[value]
-
-    return "#{tableName}CR#{cr}(#{value + 7}[#{d1},#{d2}])\n#{result}"
+    return tableName, table
   end
 end
