@@ -35,9 +35,7 @@ testDataPath = nil
 # テストデータ番号
 dataIndex = nil
 
-$stderr.puts('Tkfix') # Tkfix
-
-# HELP_MESSAGE = "Usage: #{File.basename($0)} [TEST_DATA_PATH] [DATA_INDEX]" # Tkfix
+HELP_MESSAGE = "Usage: #{File.basename($0)} [TEST_DATA_PATH] [DATA_INDEX]"
 
 if ARGV.include?('-h') || ARGV.include?('--help')
   $stdout.puts(HELP_MESSAGE)
